### 2023-12-09 5197fdb
fix(data): `reduce_int_span` with nullable dtypes
