### 2023-12-09 fb52a6f
test: Fixup pandas dtypes
