+++ .github/workflows/pre-commit.yml
@@ -15,7 +15,7 @@ jobs:
         with:
           submodules: recursive
       - name: Setup Python 3.10
-        uses: actions/setup-python@v4
+        uses: actions/setup-python@v5
         with:
           python-version: "3.10"
       - name: Install pre-commit
