+++ docs/reference/scheduling/queue_monitor.md
@@ -1,0 +1,5 @@
+## Queue Monitor
+
+::: amltk.scheduling.queue_monitor
+    options:
+        members: False
