### 2023-12-12 166c725
fix(data): `reduce_int_span` with nullable dtypes (#200)
