+++ tests/data/test_dtype_reduction.py
@@ -7,12 +7,26 @@ from pytest_cases import parametrize
 from amltk.data.dtype_reduction import reduce_dtypes
 
 
+@parametrize(
+    "series, expected",
+    (
+        (pd.Series([1, 2, None], dtype=pd.Int64Dtype()), "UInt8"),
+        (pd.Series([1, 2342, None], dtype=pd.Int64Dtype()), "UInt16"),
+        (pd.Series([-1, 1, None], dtype=pd.Int64Dtype()), "Int8"),
+        (pd.Series([-1, 2342, None], dtype=pd.Int64Dtype()), "Int16"),
+    ),
+)
+def test_reduce_dtypes_with_pandas_nan_dtypes(series: pd.Series, expected: str) -> None:
+    reduced = reduce_dtypes(series)
+    assert reduced.dtype.name == expected
+
+
 def test_reduce_dtypes_mixed_df() -> None:
     # Default 8 bytes per number
     mixed_df = pd.DataFrame({"a": np.arange(100), "b": np.linspace(0, 1, 100)})
     reduced_df = reduce_dtypes(mixed_df)
 
-    assert reduced_df["a"].dtype == np.uint8
+    assert reduced_df["a"].dtype == pd.UInt8Dtype()
     assert reduced_df["b"].dtype == pd.Float32Dtype()
 
 
@@ -20,14 +34,14 @@ def test_reduce_dtypes_mixed_df() -> None:
     "dtype, expected",
     [
         # For int's we squeeze to smallest possible that holds, max/min
-        (np.uint8, np.uint8),
-        (np.uint16, np.uint8),
-        (np.uint32, np.uint8),
-        (np.uint64, np.uint8),
-        (np.int8, np.uint8),
-        (np.int16, np.uint8),
-        (np.int32, np.uint8),
-        (np.int64, np.uint8),
+        (np.uint8, pd.UInt8Dtype()),
+        (np.uint16, pd.UInt8Dtype()),
+        (np.uint32, pd.UInt8Dtype()),
+        (np.uint64, pd.UInt8Dtype()),
+        (np.int8, pd.UInt8Dtype()),
+        (np.int16, pd.UInt8Dtype()),
+        (np.int32, pd.UInt8Dtype()),
+        (np.int64, pd.UInt8Dtype()),
         # For floats, we only do single step in precision reduction and
         # we default to pandas nullable float
         (np.float16, pd.Float32Dtype()),
