### 2023-12-13 8c127db
fix(sklearn): Better default categorical dtypes
