### 2023-12-13 aafa615
refactor (EmissionsTrackerPlugin): use _EmissionsTrackerWrapper
