### 2023-12-15 4fa295e
doc: Fix typos
