+++ docs/guides/scheduling.md
@@ -1,6 +1,6 @@
 AutoML-toolkit was designed to make offloading computation
 away from the main process __easy__, to foster increased ability for
-interact-ability deployment and control. At the same time,
+interact-ability, deployment and control. At the same time,
 we wanted to have an event based system to manage the complexity
 that comes with AutoML systems, all while making the API intuitive
 and extensible.
@@ -58,11 +58,11 @@ perform with the `Scheduler` and start the system's gears turning.
     or in a Notebook, simply leave it as the last object of a cell.
 
     You'll have to install with `amltk[jupyter]` or
-    `pip install rich[jupyter]` manually.k
+    `pip install rich[jupyter]` manually.
 
 ## Scheduler
 The core engine of the AutoML-Toolkit is the [`Scheduler`][amltk.scheduling.Scheduler].
-It purpose it to allow you to create workflows in an event driven manner. It does
+Its purpose is to allow you to create workflows in an event driven manner. It does
 this by allowing you to [`submit()`][amltk.scheduling.Scheduler.submit] functions
 with arguments to be computed in the background, while the main process can continue
 to do other work. Once this computation has completed, you can react with various
@@ -72,13 +72,13 @@ callbacks, most likely to submit more computations.
 
     If you're familiar with pythons `await/async` syntax, then this description
     might sound similar. The `Scheduler` is powered by an asynchronous event loop
-    but hides this complexity in it's API. We do have an asynchronous API which
+    but hides this complexity in its API. We do have an asynchronous API which
     we will discuss later.
 
 ### Backend
 
 The first thing to do is define where this computation should happen.
-A [`Scheduler`][amltk.scheduling.Scheduler] builds upon, an
+A [`Scheduler`][amltk.scheduling.Scheduler] builds upon an
 [`Executor`][concurrent.futures.Executor],
 an interface provided by python's [`concurrent.futures`][concurrent.futures]
 module. This interface is used to abstract away the details of how the
@@ -91,11 +91,10 @@ or even custom backends.
     You can find a list of these in our
     [executor reference](../reference/scheduling/executors.md).
 
-The simplest one is a [`ProcessPoolExecutor`][concurrent.futures.ProcessPoolExecutor]
+The simplest one is a [`ProcessPoolExecutor`][concurrent.futures.ProcessPoolExecutor],
 which will create a pool of processes to run the compute in parallel. We provide
 a convenience function for this as
-[`Scheduler.with_processes()`][amltk.scheduling.Scheduler.with_processes]
-well as some other builder
+[`Scheduler.with_processes()`][amltk.scheduling.Scheduler.with_processes].
 
 ```python exec="true" source="material-block" html="True"
 from concurrent.futures import ProcessPoolExecutor
@@ -107,10 +106,10 @@ from amltk._doc import doc_print; doc_print(print, scheduler)  # markdown-exec: 
 
 ### Running the Scheduler
 
-You may have noticed from the above example that there are many events the shceduler will emit,
+You may have noticed from the above example that there are many events the scheduler will emit,
 such as `@on_start` or `@on_future_done`. One particularly important one is
 [`@on_start`][amltk.scheduling.Scheduler.on_start], an event to signal
-the scheduler has started and ready to accept tasks.
+the scheduler has started and is ready to accept tasks.
 
 ```python exec="true" source="material-block" html="True"
 from amltk.scheduling import Scheduler
@@ -124,7 +123,7 @@ from amltk._doc import doc_print; doc_print(print, scheduler, output="html", fon
 ```
 
 From the output, we can see that the `print_hello()` function was registered
-to the event `@on_start`, but it was never called and no `#!python "hello"` printed.
[…]
