+++ docs/guides/pipelines.md
@@ -11,8 +11,8 @@ if you just want to quickly look something up.
 
 ## Introduction
 The kinds of pipelines that exist in an AutoML system come in many different
-forms. For example, one might be an [sklearn.pipeline.Pipeline][], other's
-might be some deep-learning pipeline while some might even stand for some
+forms. For example, one might be an [sklearn.pipeline.Pipeline][], others
+might be some deep-learning pipeline, while some might even stand for some
 real life machinery process and the settings of these machines.
 
 To accommodate this, what AutoML-Toolkit provides is an **abstract** representation
@@ -23,7 +23,7 @@ We categorize this into 4 steps:
 
 1. Parametrize your pipeline using the various [components](../reference/pipelines/pipeline.md),
     including the kinds of items in the pipeline, the search spaces and any additional configuration.
-    Each of the various types of components give a syntactic meaning when performing the next steps.
+    Each of the various types of components gives a syntactic meaning when performing the next steps.
 
 2. [`pipeline.search_space(parser=...)`][amltk.pipeline.Node.search_space],
     Get a useable search space out of the pipeline. This can then be passed to an
@@ -34,7 +34,7 @@ We categorize this into 4 steps:
     an optimizer.
 
 4. [`pipeline.build(builder=...)`][amltk.pipeline.Node.build],
-    Build your configured pipeline definition into something usable, i.e.
+    Build your configured pipeline definition into something useable, i.e.
     an [`sklearn.pipeline.Pipeline`][sklearn.pipeline.Pipeline] or a
     [`torch.nn.Module`](https://pytorch.org/docs/stable/generated/torch.nn.Module.html).
 
@@ -145,7 +145,8 @@ from amltk._doc import doc_print; doc_print(print, component_with_function, outp
 ### Search Space
 If interacting with an [`Optimizer`](../reference/optimization/optimizers.md), you'll often require some
 search space object to pass to it.
-To extract a search space from a `Component`, we can call [`search_space(parser=)`][amltk.pipeline.Node.search_space],
+To extract a search space from a `Component`, we can call 
+[`search_space(parser=...)`][amltk.pipeline.Node.search_space],
 passing in the kind of search space you'd like to get out of it.
 
 ```python exec="true" source="material-block" result="python" session="Pipeline-Component"
@@ -165,7 +166,7 @@ You may also define your own `parser=` and use that if desired.
 
 
 ### Configure
-Pretty straight forward but what do we do with this `config`? Well we can
+Pretty straight forward, but what do we do with this `config`? Well we can
 [`configure(config=...)`][amltk.pipeline.Node.configure] the component with it.
 
 ```python exec="true" source="material-block" html="true" session="Pipeline-Component"
@@ -309,8 +310,8 @@ from amltk._doc import doc_print; doc_print(print, configured_component, fontsiz
 
 !!! tip inline end "Transform Context"
 
-    There may be times where you may have some additional context which you may only
-    know at configuration time, you may pass this to `configure(..., transform_context=...)`
+    There may be times where you have some additional context, which you may only know at configuration time. 
+    In this case, it is possible to pass this additional context to `configure(..., transform_context=...)`, 
     which will be forwarded as the second argument to your `.config_transform`.
 
 ## Sequential
@@ -337,8 +338,8 @@ doc_print(print, rf)  # markdown-exec: hide
 !!! info inline end "Infix `>>`"
 
     To join these two components together, we can either use the infix notation using `>>`,
-    or passing them directly to a [`Sequential`][amltk.pipeline.Sequential]. However
-    a **random name** will be given.
+    or passing them directly to a [`Sequential`][amltk.pipeline.Sequential]. However,
+    a **random name** will be given when using the infix notation.
 
[…]
