### 2023-12-18 ea3276c
chore: update dask requirement from <=2023.4 to <=2023.12.1
Updates the requirements on [dask](https://github.com/dask/dask) to permit the latest version.
- [Changelog](https://github.com/dask/dask/blob/main/docs/release-procedure.md)
- [Commits](https://github.com/dask/dask/compare/0.2.0...2023.12.1)

---
updated-dependencies:
- dependency-name: dask
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
