### 2023-12-19 eafd06c
docs: Fix typos in readme and guides (#204)
