### 2023-12-19 f8bcebd
chore: bump actions/upload-artifact from 3 to 4 (#205)
Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
