+++ .github/workflows/release.yml
@@ -51,7 +51,7 @@ jobs:
           mv changelog-increment.md changelog-output/changelog-increment.md
           cat changelog-output/changelog-increment.md
       - name: Upload the changelog
-        uses: actions/upload-artifact@v3
+        uses: actions/upload-artifact@v4
         with:
           name: changelog
           path: changelog-output
@@ -72,7 +72,7 @@ jobs:
       - run: python -m pip install build
       - run: python -m build --sdist
       - name: Store the distribution packages
-        uses: actions/upload-artifact@v3
+        uses: actions/upload-artifact@v4
         with:
           name: build-output
           path: dist
