### 2023-12-19 2960fc4
chore: bump actions/download-artifact from 3 to 4 (#206)
Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
