+++ .github/workflows/release.yml
@@ -122,13 +122,13 @@ jobs:
           fetch-tags: 1
           fetch-depth: 0
       - name: Download the build artifiact
-        uses: actions/download-artifact@v3
+        uses: actions/download-artifact@v4
         with:
             name: build-output
             path: dist
       - run: ls -R dist
       - name: Download the changelog
-        uses: actions/download-artifact@v3
+        uses: actions/download-artifact@v4
         with:
           name: changelog
           path: changelog-output
@@ -156,7 +156,7 @@ jobs:
       - uses: actions/checkout@v4
         with:
           ref: "main"  # Necessary to download the latest of main as this will have been updated on the step before
-      - uses: actions/download-artifact@v3
+      - uses: actions/download-artifact@v4
         with:
             name: build-output
             path: dist
