### 2023-12-22 961d481
Update emissions_tracker_plugin.py
