### 2023-12-22 07ea690
refactor (EmissionsTrackerPlugin): Update docstring
