### 2023-12-22 74eed54
refactor (_EmissionsTrackerWrapper): Update docstring
