### 2024-01-07 a6d21e2
refactor (_EmissionsTrackerWrapper): Update docstring and use Generic
