+++ src/amltk/scheduling/plugins/emissions_tracker_plugin.py
@@ -1,29 +1,27 @@
 from codecarbon import EmissionsTracker
-from typing import Callable, Any, ClassVar, TypeVar
-from typing_extensions import Self
+from typing import Callable, Any, ClassVar, Generic, TypeVar
+from typing_extensions import Self, ParamSpec
 
 from amltk.scheduling.plugins.plugin import Plugin
 from amltk.scheduling.task import Task
 
-P = TypeVar("P")
+P = ParamSpec("P")
 R = TypeVar("R")
 
 
-class _EmissionsTrackerWrapper:
+class _EmissionsTrackerWrapper(Generic[P, R]):
     """A wrapper around codecarbon package to measure emissions."""
 
     def __init__(
-            self,
-            fn: Callable[[P], R],
-            task: Task,
-            *codecarbon_args: Any,
-            **codecarbon_kwargs: Any,
+        self,
+        fn: Callable[[P], R],
+        *codecarbon_args: Any,
+        **codecarbon_kwargs: Any,
     ):
         """Initialize the wrapper.
 
         Args:
             fn: The function to wrap.
-            task: The task the function is being wrapped for.
             *codecarbon_args: arguments to pass to
                 [`codecarbon.EmissionsTracker`][codecarbon.EmissionsTracker].
             **codecarbon_kwargs: keyword arguments to pass to
@@ -31,23 +29,19 @@ class _EmissionsTrackerWrapper:
         """
         super().__init__()
         self.fn = fn
-        self.task = task
         self.codecarbon_args = codecarbon_args
         self.codecarbon_kwargs = codecarbon_kwargs
 
-    def __call__(self, *args: any, **kwargs: any) -> R:
+    def __call__(self, *args: P.args, **kwargs: P.kwargs) -> R:
         with EmissionsTracker(*self.codecarbon_args, **self.codecarbon_kwargs) as tracker:
             result = self.fn(*args, **kwargs)
             return result
 
 
 class EmissionsTrackerPlugin(Plugin):
-    """A plugin that tracks carbon emissions using codecarbon library."""
-
-    name: ClassVar = "emissions-tracker"
-    """The name of the plugin."""
-
     """
+    A plugin that tracks carbon emissions using codecarbon library.
+
     Usage Example:
 
     ```python
@@ -65,7 +59,10 @@ class EmissionsTrackerPlugin(Plugin):
 
     # Create a task with the emissions tracker plugin
     task = scheduler.task(some_function, plugins=[
-        EmissionsTrackerPlugin(log_level="info", save_to_file=False) # pass any codecarbon args here
+        # Please,  check out the official documentation for more details on the codecarbon arguments:
+        # https://mlco2.github.io/codecarbon/parameters.html
+        # Pass any codecarbon parameters as args here
+        EmissionsTrackerPlugin(log_level="info", save_to_file=False)
     ])
 
     @scheduler.on_start
@@ -74,15 +71,18 @@ class EmissionsTrackerPlugin(Plugin):
 
     @task.on_submitted
     def on_submitted(future, *args, **kwargs):
-        print(f"Task was submitted", future, args, kwargs) 
+        print(f"Task was submitted", future, args, kwargs)
 
     @task.on_done
     def on_done(future):
         print("Task done: ", future.result()) # result is the return value of the function
 
-    scheduler.run()    
-    ```  
+    scheduler.run()
+    ```
     """
+
+    name: ClassVar = "emissions-tracker"
+    """The name of the plugin."""
 
     def __init__(self, *args: Any, **kwargs: Any):
         super().__init__()
@@ -95,11 +95,11 @@ class EmissionsTrackerPlugin(Plugin):
[…]
