### 2024-01-08 6f5e535
refactor (_EmissionsTrackerWrapper): Fix linter warning.
