+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "amltk"
-version = "1.4.0"
+version = "1.5.0"
 dependencies = [
   "typing_extensions", # Better typing
   "more_itertools",    # Better iteration
@@ -97,7 +97,7 @@ exclude_lines = [
 
 [tool.commitizen]
 name = "cz_conventional_commits"
-version = "1.4.0"
+version = "1.5.0"
 update_changelog_on_bump = true
 version_files = ["pyproject.toml:version", "src/amltk/__version__.py"]
 changelog_start_rev = "1.0.0"
