### 2024-01-10 68198b9
feat(history): Implement `best()`
