### 2024-01-10 0e578d3
feat(history): Get `best()` from History (#209)
