### 2024-01-10 b752efd
bump: version 1.5.0 → 1.6.0
