+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.6.0 (2024-01-10)
+
+### Feat
+
+- **history**: Get `best()` from History (#209)
+
 ## 1.5.0 (2024-01-09)
 
 ### Feat
