+++ pyproject.toml
@@ -54,7 +54,7 @@ smac = ["smac>=2.0", "amltk[configspace]"]
 optuna = ["optuna"]
 configspace = ["configspace>=0.6"]
 loky = ["loky"]
-dask = ["dask<=2023.4", "distributed"]
+dask = ["dask<=2024.1.0", "distributed"]
 pynisher = ["pynisher>=1.0.10"]
 wandb = ["wandb"]
 threadpoolctl = ["threadpoolctl"]
