### 2024-01-15 4e981d6
feat(scheduler): Handled errors with specific method
