### 2024-01-15 5893ad5
feat(scheduler): Handled errors with specific method
