+++ src/amltk/scheduling/scheduler.py
@@ -254,7 +254,7 @@ import asyncio
 import logging
 import warnings
 from asyncio import Future
-from collections.abc import Callable, Iterable
+from collections.abc import Callable, Iterable, Mapping
 from concurrent.futures import Executor, ProcessPoolExecutor
 from dataclasses import dataclass
 from enum import Enum, auto
@@ -271,6 +271,8 @@ from typing import (
 )
 from typing_extensions import Self, override
 from uuid import uuid4
+
+from more_itertools import first
 
 from amltk._asyncm import ContextEvent
 from amltk._functional import Flag
@@ -499,8 +501,12 @@ class Scheduler(RichRenderable):
         # This is triggered when run is called
         self._running_event = asyncio.Event()
 
-        # This is set once `run` is called
-        self._end_on_exception_flag = Flag(initial=False)
+        # This is set once `run` is called.
+        # Either contains the mapping from exception to what to do,
+        # otherwise a boolean whether to end or not.
+        self._end_on_exception_flag: Flag[
+            bool | dict[type[Exception], Literal["raise", "end", "ignore"]]
+        ] = Flag(initial=False)
 
         # This is used to manage suequential queues, where we need a Thread
         # timer to ensure that we don't get caught in an endless loop waiting
@@ -1065,8 +1071,40 @@ class Scheduler(RichRenderable):
         exception = future.exception()
         if exception:
             self.on_future_exception.emit(future, exception)
-            if self._end_on_exception_flag and future.done():
-                self.stop(stop_msg="Ending on first exception", exception=exception)
+            flag_value = self._end_on_exception_flag.value
+            match flag_value:
+                case False:
+                    pass
+                case True:
+                    self.stop(stop_msg="Ending on first exception", exception=exception)
+                case Mapping():
+                    method: Literal["raise", "end", "ignore"]
+                    match first(
+                        (
+                            (err_type, method)
+                            for err_type, method in flag_value.items()
+                            if isinstance(exception, err_type)
+                        ),
+                        default=None,
+                    ):
+                        # Not explicitly handled
+                        case None:
+                            self.stop(
+                                msg=f"Ending on exception type {type(exception)} as"
+                                f"it's not contained in {flag_value}",
+                                exception=exception,
+                            )
+                        # Handled and should end
+                        case (err_type, method) if method in ("end", "raise"):
+                            self.stop(
+                                msg=f"{method}ing on exception type {type(exception)}"
+                                f" as {err_type} is {method} from"
+                                f" `on_exception={flag_value}`",
+                                exception=exception,
+                            )
+                        # Handled and should ignore
+                        case _:
+                            pass
         else:
             result = future.result()
             self.on_future_result.emit(future, result)
@@ -1226,7 +1264,10 @@ class Scheduler(RichRenderable):
         timeout: float | None = None,
         end_on_empty: bool = True,
         wait: bool = True,
-        on_exception: Literal["raise", "end", "ignore"] = "raise",
+        on_exception: (
+            Literal["raise", "end", "ignore"]
+            | Mapping[type[Exception], Literal["raise", "end", "ignore"]]
+        ) = "raise",
         asyncio_debug_mode: bool = False,
         display: bool | list[RenderableType] = False,
     ) -> ExitState:
@@ -1256,6 +1297,38 @@ class Scheduler(RichRenderable):
                 * If `#!python "end"`, similar to `#!python "raise"`, the scheduler
                     will stop but no exception will occur and the control flow
                     will return gracefully to the point where you called `run()`.
+
+                * If a `dict`, then it's a mapping from error types to what should be
+                    done about them. This can be used with something like the
+                    [`amltk.scheduling.plugins.plugins.PynisherPlugin`] to prevent
+                    it from killing the scheduler when an individual trial fails.
+
+                    ```python
[…]
