### 2024-01-15 518f1c9
refactor(Scheduler): Use function for subclass mapping
