### 2024-01-15 b7a5baa
doc(_functional): `subclass_map` documentation
