### 2024-01-15 c043e83
doc(_functional): Add example for `subclass_map`
