### 2024-01-15 76e589f
fix(Scheduler): Make sure to raise BaseExceptions
