### 2024-01-15 38124c2
refactor(Scheduling): Cleanup async handler
