### 2024-01-15 5c2c4be
refactor(Scheduling): Simply `_end_on_exception_flag` handling
