### 2024-01-15 69d5e0d
refactor(Scheduler): Simplify return `async_run`
