### 2024-01-16 9c68d29
test(Scheduler): Test ordering for `on_exception`
