### 2024-01-16 6696402
feat(scheduler): Handled errors with specific method (#213)
