+++ src/amltk/scheduling/scheduler.py
@@ -254,7 +254,7 @@ import asyncio
 import logging
 import warnings
 from asyncio import Future
-from collections.abc import Callable, Iterable
+from collections.abc import Callable, Iterable, Mapping
 from concurrent.futures import Executor, ProcessPoolExecutor
 from dataclasses import dataclass
 from enum import Enum, auto
@@ -273,7 +273,7 @@ from typing_extensions import Self, override
 from uuid import uuid4
 
 from amltk._asyncm import ContextEvent
-from amltk._functional import Flag
+from amltk._functional import Flag, subclass_map
 from amltk._richutil.renderable import RichRenderable
 from amltk.exceptions import SchedulerNotRunningError
 from amltk.scheduling.events import Emitter, Event, Subscriber
@@ -499,8 +499,12 @@ class Scheduler(RichRenderable):
         # This is triggered when run is called
         self._running_event = asyncio.Event()
 
-        # This is set once `run` is called
-        self._end_on_exception_flag = Flag(initial=False)
+        # This is set once `run` is called.
+        # Either contains the mapping from exception to what to do,
+        # otherwise a boolean whether to end or not.
+        self._end_on_exception_flag: Flag[
+            bool | dict[type[Exception], Literal["raise", "end", "ignore"]]
+        ] = Flag(initial=False)
 
         # This is used to manage suequential queues, where we need a Thread
         # timer to ensure that we don't get caught in an endless loop waiting
@@ -1063,13 +1067,57 @@ class Scheduler(RichRenderable):
         self.on_future_done.emit(future)
 
         exception = future.exception()
-        if exception:
-            self.on_future_exception.emit(future, exception)
-            if self._end_on_exception_flag and future.done():
-                self.stop(stop_msg="Ending on first exception", exception=exception)
-        else:
+        if exception is None:
             result = future.result()
             self.on_future_result.emit(future, result)
+            return
+
+        self.on_future_exception.emit(future, exception)
+        if self._end_on_exception_flag.value is False:
+            return
+
+        if self._end_on_exception_flag.value is True:
+            self.stop(stop_msg="Ending on first exception", exception=exception)
+            return
+
+        exception_handle_mapping = self._end_on_exception_flag.value
+        match exception:
+            case BaseException() if not isinstance(exception, Exception):
+                self.stop(
+                    msg=f"Ending as BaseException {type(exception)} occured.",
+                    exception=exception,
+                )
+            case _:
+                assert isinstance(exception, Exception)
+                method: Literal["raise", "end", "ignore"]
+                match subclass_map(exception, exception_handle_mapping):
+                    # Not explicitly handled
+                    case None:
+                        self.stop(
+                            msg=(
+                                f"Ending on exception type {type(exception)} as"
+                                "it's not contained in"
+                                f"`on_exception={exception_handle_mapping}"
+                            ),
+                            exception=exception,
+                        )
+                    # Handled and should end
+                    case (err_type, method) if method in ("end", "raise"):  # type: ignore
+                        self.stop(
+                            msg=(
+                                f"{method}ing on exception '{type(exception)}'"
+                                f" as {err_type} is {method} as specified from"
+                                f" `on_exception={exception_handle_mapping}`"
+                            ),
+                            exception=exception,
+                        )
+                    case (err_type, "ignore"):
+                        pass
+                    case thing:
+                        raise RuntimeError(
+                            f"Unexpected case {thing=} for {exception=}."
+                            "\nPlease report this as a bug!",
+                        ) from exception
 
     async def _monitor_queue_empty(self) -> None:
         """Monitor for the queue being empty and trigger an event when it is."""
@@ -1226,7 +1274,10 @@ class Scheduler(RichRenderable):
         timeout: float | None = None,
         end_on_empty: bool = True,
[…]

[…]
