### 2024-01-16 61d4932
feat(Scheduler): Respond to cancelled futures
