### 2024-01-16 fbfa781
refactor(Scheduler): Change `"ignore"` to `"continue"`
