+++ src/amltk/scheduling/scheduler.py
@@ -503,12 +503,12 @@ class Scheduler(RichRenderable):
         # Either contains the mapping from exception to what to do,
         # otherwise a boolean whether to end or not.
         self._on_exc_method_map: Flag[
-            dict[type[Exception], Literal["raise", "end", "ignore"]],
+            dict[type[Exception], Literal["raise", "end", "continue"]],
         ] = Flag(initial={Exception: "raise"})
 
         # This is set once `run` is called.
         # Indicates what to do when a cancellation occurs
-        self._on_cancel_method: Flag[Literal["raise", "end", "ignore"]] = Flag(
+        self._on_cancel_method: Flag[Literal["raise", "end", "continue"]] = Flag(
             initial="raise",
         )
 
@@ -1071,10 +1071,10 @@ class Scheduler(RichRenderable):
             match self._on_cancel_method.value:
                 case "raise" | "end":
                     self.stop(
-                        msg="Ending on cancellation of a future",
+                        stop_msg="Ending on cancellation of a future",
                         exception=CancelledError(f"Future {future} was cancelled"),
                     )
-                case "ignore":
+                case "continue":
                     pass
             return
 
@@ -1098,14 +1098,25 @@ class Scheduler(RichRenderable):
                     exception=exception,
                 )
             case Exception():
+                self.on_future_exception.emit(future, exception)
                 match subclass_map(exception, self._on_exc_method_map.value):
-                    case (err_type, "ignore"):
+                    case (err_type, "continue"):
                         return
-                    case (err_type, method) if method in ("raise", "end"):
+                    case (err_type, "end"):
                         self.stop(
-                            msg=(
-                                f"{method}ing on exception '{type(exception)}'"
-                                f" as {err_type} is {method} as specified from"
+                            stop_msg=(
+                                f"Ending on exception '{type(exception)}'"
+                                f" as {err_type} is 'end' as specified from"
+                                f"`on_exception={self._on_exc_method_map.value}"
+                            ),
+                            exception=exception,
+                        )
+                        return
+                    case (err_type, "raise"):
+                        self.stop(
+                            stop_msg=(
+                                f"raising on exception '{type(exception)}'"
+                                f" as {err_type} is 'raise' as specified from"
                                 f"`on_exception={self._on_exc_method_map.value}"
                             ),
                             exception=exception,
@@ -1113,7 +1124,7 @@ class Scheduler(RichRenderable):
                         return
                     case None:
                         self.stop(
-                            msg=(
+                            stop_msg=(
                                 f"Ending on exception type {type(exception)} as"
                                 "it's not contained in"
                                 f"`on_exception={self._on_exc_method_map.value}"
@@ -1286,10 +1297,10 @@ class Scheduler(RichRenderable):
         end_on_empty: bool = True,
         wait: bool = True,
         on_exception: (
-            Literal["raise", "end", "ignore"]
-            | Mapping[type[Exception], Literal["raise", "end", "ignore"]]
+            Literal["raise", "end", "continue"]
+            | Mapping[type[Exception], Literal["raise", "end", "continue"]]
         ) = "raise",
-        on_cancelled: Literal["raise", "end", "ignore"] = "raise",
+        on_cancelled: Literal["raise", "end", "continue"] = "raise",
         asyncio_debug_mode: bool = False,
         display: bool | Iterable[RenderableType] = False,
     ) -> ExitState:
@@ -1313,9 +1324,9 @@ class Scheduler(RichRenderable):
 
                 * If `#!python "raise"`, the scheduler will stop and raise the
                     exception at the point where you called `run()`.
-                * If `#!python "ignore"`, the scheduler will continue running,
[…]
