### 2024-01-16 d8e6366
feat(Scheduler): Respond to cancelled futures (#214)
