+++ src/amltk/scheduling/scheduler.py
@@ -255,7 +255,7 @@ import logging
 import warnings
 from asyncio import Future
 from collections.abc import Callable, Iterable, Mapping
-from concurrent.futures import Executor, ProcessPoolExecutor
+from concurrent.futures import CancelledError, Executor, ProcessPoolExecutor
 from dataclasses import dataclass
 from enum import Enum, auto
 from functools import partial
@@ -502,9 +502,15 @@ class Scheduler(RichRenderable):
         # This is set once `run` is called.
         # Either contains the mapping from exception to what to do,
         # otherwise a boolean whether to end or not.
-        self._end_on_exception_flag: Flag[
-            bool | dict[type[Exception], Literal["raise", "end", "ignore"]]
-        ] = Flag(initial=False)
+        self._on_exc_method_map: Flag[
+            dict[type[Exception], Literal["raise", "end", "continue"]],
+        ] = Flag(initial={Exception: "raise"})
+
+        # This is set once `run` is called.
+        # Indicates what to do when a cancellation occurs
+        self._on_cancel_method: Flag[Literal["raise", "end", "continue"]] = Flag(
+            initial="raise",
+        )
 
         # This is used to manage suequential queues, where we need a Thread
         # timer to ensure that we don't get caught in an endless loop waiting
@@ -1062,62 +1068,78 @@ class Scheduler(RichRenderable):
 
         if future.cancelled():
             self.on_future_cancelled.emit(future)
+            match self._on_cancel_method.value:
+                case "raise" | "end":
+                    self.stop(
+                        stop_msg="Ending on cancellation of a future",
+                        exception=CancelledError(f"Future {future} was cancelled"),
+                    )
+                case "continue":
+                    pass
             return
 
         self.on_future_done.emit(future)
 
         exception = future.exception()
-        if exception is None:
-            result = future.result()
-            self.on_future_result.emit(future, result)
-            return
-
-        self.on_future_exception.emit(future, exception)
-        if self._end_on_exception_flag.value is False:
-            return
-
-        if self._end_on_exception_flag.value is True:
-            self.stop(stop_msg="Ending on first exception", exception=exception)
-            return
-
-        exception_handle_mapping = self._end_on_exception_flag.value
         match exception:
+            case None:
+                result = future.result()
+                self.on_future_result.emit(future, result)
+                return
+            # We do not allow users to handle `BaseException`s as they are
+            # not meant to be caught.
             case BaseException() if not isinstance(exception, Exception):
+                self.on_future_exception.emit(future, exception)
                 self.stop(
-                    msg=f"Ending as BaseException {type(exception)} occured.",
+                    stop_msg=(
+                        f"BaseException occured - {type(exception)}:{exception}."
+                        "\nEnding as this is not meant to be caught."
+                    ),
                     exception=exception,
                 )
-            case _:
-                assert isinstance(exception, Exception)
-                method: Literal["raise", "end", "ignore"]
-                match subclass_map(exception, exception_handle_mapping):
-                    # Not explicitly handled
+            case Exception():
+                self.on_future_exception.emit(future, exception)
+                match subclass_map(exception, self._on_exc_method_map.value):
+                    case (err_type, "continue"):
+                        return
+                    case (err_type, "end"):
+                        self.stop(
+                            stop_msg=(
+                                f"Ending on exception '{type(exception)}'"
+                                f" as {err_type} is 'end' as specified from"
+                                f"`on_exception={self._on_exc_method_map.value}"
+                            ),
+                            exception=exception,
+                        )
+                        return
+                    case (err_type, "raise"):
+                        self.stop(
+                            stop_msg=(
+                                f"raising on exception '{type(exception)}'"
[…]
