### 2024-01-16 66883cd
test: Update fixtures to be function scope
