### 2024-01-16 251e1b5
test: Update fixtures to be explicitly function scope (#215)
