+++ tests/scheduling/plugins/test_call_limiter_plugin.py
@@ -51,7 +51,7 @@ def case_loky_executor() -> ProcessPoolExecutor:
     return get_reusable_executor(max_workers=2)  # type: ignore
 
 
-@fixture()
+@fixture(scope="function")
 @parametrize_with_cases("executor", cases=".", has_tag="executor")
 def scheduler(executor: Executor) -> Iterator[Scheduler]:
     yield Scheduler(executor)
