### 2024-01-16 672889b
bump: version 1.6.0 → 1.7.0
