### 2024-01-17 f0a9db9
refactor(Trial): Add clear error message for using the Metric class
