+++ src/amltk/optimization/trial.py
@@ -72,7 +72,7 @@ logger = logging.getLogger(__name__)
 @dataclass(kw_only=True)
 class Trial(RichRenderable, Generic[I]):
     """A [`Trial`][amltk.optimization.Trial] encapsulates some configuration
-    that needs to be evaluated. Typically this is what is generated by an
+    that needs to be evaluated. Typically, this is what is generated by an
     [`Optimizer.ask()`][amltk.optimization.Optimizer.ask] call.
 
     ??? tip "Usage"
@@ -84,7 +84,7 @@ class Trial(RichRenderable, Generic[I]):
         If all went smooth, your trial was successful and you can use
         [`trial.success()`][amltk.optimization.Trial.success] to generate
         a success [`Report`][amltk.optimization.Trial.Report], typically
-        passing what your chosen optimizer expects, e.g. `"loss"` or `"cost"`.
+        passing what your chosen optimizer expects, e.g., `"loss"` or `"cost"`.
 
         If your trial failed, you can instead use the
         [`trial.fail()`][amltk.optimization.Trial.fail] to generate a
@@ -121,7 +121,7 @@ class Trial(RichRenderable, Generic[I]):
 
     What you can return with [`trial.success()`][amltk.optimization.Trial.success]
     or [`trial.fail()`][amltk.optimization.Trial.fail] depends on the
-    [`metrics`][amltk.optimization.Trial.metrics] of the trial. Typically
+    [`metrics`][amltk.optimization.Trial.metrics] of the trial. Typically,
     an optimizer will provide the trial with the list of metrics.
 
     ??? tip "Metrics"
@@ -130,12 +130,52 @@ class Trial(RichRenderable, Generic[I]):
             options:
                 members: False
 
-    Some important properties is that they have a unique
+    Some important properties are that they have a unique
     [`.name`][amltk.optimization.Trial.name] given the optimization run,
-    a candidate [`.config`][amltk.optimization.Trial.config]' to evaluate,
+    a candidate [`.config`][amltk.optimization.Trial.config] to evaluate,
     a possible [`.seed`][amltk.optimization.Trial.seed] to use,
-    and an [`.info`][amltk.optimization.Trial.info] object which is the optimizer
+    and an [`.info`][amltk.optimization.Trial.info] object, which is the optimizer
     specific information, if required by you.
+
+    !!! tip
+    When using the `success` method, make sure to provide values for all metrics
+    specified in the `metrics` attribute. Each metric has a unique name, and it's
+    crucial to use the correct names when reporting success. Failure to provide
+    a value for any metric will result in a `ValueError`. To ensure correctness,
+    refer to the `self.metrics` attribute for the list of metrics and use their
+    exact names in the `trial.success` method. For example:
+
+    ```python
+    trial = Trial(
+        name="example_trial",
+        config={"param": 42},
+        metrics=[Metric(name="accuracy", minimize=False)]
+    )
+
+    # Incorrect usage (will raise a ValueError)
+    report = trial.success(invalid_metric=0.95)
+
+    # Correct usage
+    report = trial.success(accuracy=0.95)
+    ```
+
+    Additionally, if you encounter a metric name mismatch,
+    consider renaming the metric using the provided key:
+
+    ```python
+    trial = Trial(
+        name="example_trial",
+        config={"param": 42},
+        metrics=[Metric(name="poly_metric", minimize=False)]
+    )
+
+    # Incorrect usage (will raise a ValueError)
+    report = trial.success(poly_metric=0.95)
+
+    # Correct usage
+    report = trial.success(provided_key=0.95)
+    trial.metrics = [Metric(name="provided_key", minimize=False, bounds=None)]
+    ```
 
     If using [`Plugins`][amltk.scheduling.plugins.Plugin], they may insert
[…]
