### 2024-01-17 ccb075d
doc(Trial): Fixup docstring
