### 2024-01-17 66f7b32
doc(Trial): Better indication of what to do with the Metric name for report (#216)
Co-authored-by: eddiebergman <eddiebergmanhs@gmail.com>
