### 2024-01-18 b5a11d7
feat(Pipeline): `factorize()` a pipeline into its possibilities
