### 2024-01-18 e26d4de
fix(Pipeline): `factorize(..., assign_child=...)` recieves copy
