### 2024-01-18 02b5897
other: Remove mention of changelog as it's done automatically
