### 2024-01-18 bf610f7
other: Add issue templates for doc and ux
