+++ .github/ISSUE_TEMPLATE/doc_request.md
@@ -1,0 +1,17 @@
+---
+name: Doc request
+about: Suggest something that should be documented
+title: ''
+labels: ['doc']
+assignees: ''
+
+---
+
+**Was there something unclear in the documentation?**
+If something in the documentation was unclear, please describe what you interpreted
+it to mean and what it actually meant.
+
+If something wasn't there, please let us know what you were looking for.
+
+**Experience with AMLTK?**
+What's your experience with AMLTK, are you a newcomer or been using it for a while?
