+++ .github/ISSUE_TEMPLATE/ux_request.md
@@ -1,0 +1,15 @@
+---
+name: User Experience
+about: What would make using/developing AMLTK easier?
+title: ''
+labels: ['ux']
+assignees: 'eddiebergman'
+
+---
+
+**What would have made your life easier using AMLTK?**
+* Perhaps it was an error message which wasn't clear. If so, maybe
+you could suggest what would have been clearer and the actual solution once fixed.
+
+* If developing, maybe there's some part of the development experience that could
+be improved?
