### 2024-01-20 3712b43
feat(pipeline): `factorize` based on depth
