### 2024-01-20 dd0eb6e
fix(pipeline): Factorize exit early yield
