### 2024-01-20 0196072
doc: Add note on how to use `factor_by=`
