### 2024-01-20 e781284
fix(pipeline): Forward new factorize args to `Node`
