### 2024-01-22 abc1f75
feat(Pipeline): `factorize()` a pipeline into its possibilities (#217)
