### 2024-01-22 372fca3
feat(Pynisher): Detect tasks with `Trial` to report `FAIL`
If a task which was used for a `Trial`, such as those when
optimizing trials from an Optimizer, exceeded it's constraints,
it would previously just be a generic `Task` exception. However
it's often the case that we would rather get back a `Trial.fail()`
such that we can handle it in a uniform manner with other trials
that failed.
