+++ src/amltk/scheduling/plugins/pynisher.py
@@ -82,14 +82,17 @@ performed in processes.
 """  # noqa: E501
 from __future__ import annotations
 
+import traceback
 from collections.abc import Callable
+from dataclasses import dataclass
 from multiprocessing.context import BaseContext
-from typing import TYPE_CHECKING, ClassVar, Literal, TypeAlias, TypeVar
+from typing import TYPE_CHECKING, ClassVar, Generic, Literal, TypeAlias, TypeVar
 from typing_extensions import ParamSpec, Self, override
 
 import pynisher
 import pynisher.exceptions
 
+from amltk.optimization import Trial
 from amltk.scheduling.events import Event
 from amltk.scheduling.plugins.plugin import Plugin
 
@@ -102,6 +105,62 @@ if TYPE_CHECKING:
 
 P = ParamSpec("P")
 R = TypeVar("R")
+
+
+@dataclass
+class _PynisherWrap(Generic[P, R]):
+    fn: Callable[P, R]
+    memory_limit: int | tuple[int, str] | None = None
+    cputime_limit: int | tuple[float, str] | None = None
+    walltime_limit: int | tuple[float, str] | None = None
+    terminate_child_processes: bool = True
+    context: BaseContext | None = None
+    disable_trial_handling: bool = False
+
+    def __call__(self, *args: P.args, **kwargs: P.kwargs) -> R:
+        if any(
+            limit is not None
+            for limit in (self.memory_limit, self.cputime_limit, self.walltime_limit)
+        ):
+            fn = pynisher.Pynisher(
+                self.fn,
+                memory=self.memory_limit,
+                cpu_time=self.cputime_limit,
+                wall_time=self.walltime_limit,
+                terminate_child_processes=True,
+                context=self.context,
+            )
+        else:
+            fn = self.fn
+
+        trial: Trial | None = None
+        if not self.disable_trial_handling:
+            if len(args) > 0 and isinstance(args[0], Trial):
+                trial = args[0]
+            elif (_trial := kwargs.get("trial")) is not None:
+                if not isinstance(_trial, Trial):
+                    raise ValueError(
+                        f"Expected 'trial' to be a Trial instance, got {type(trial)}"
+                        f"\n{trial=}",
+                    )
+                trial = _trial
+            else:
+                raise ValueError(
+                    "Could not find a Trial instance in the args for PynisherPlugin."
+                    "\nPlease pass it as the first argument or as a keyword argument"
+                    " or disable trial handling with `disable_trial_handling=True`.",
+                )
+
+        if trial is not None:
+            try:
+                return fn(*args, **kwargs)
+            except pynisher.PynisherException as e:
+                tb = traceback.format_exc()
+                trial.exception = e
+                trial.traceback = tb
+                return trial.fail()  # type: ignore
+        else:
+            return fn(*args, **kwargs)
 
 
 class PynisherPlugin(Plugin):
@@ -259,6 +318,7 @@ class PynisherPlugin(Plugin):
         cputime_limit: int | tuple[float, str] | None = None,
         walltime_limit: int | tuple[float, str] | None = None,
         context: BaseContext | None = None,
+        disable_trial_handling: bool = False,
     ):
         """Initialize a `PynisherPlugin` instance.
 
@@ -274,12 +334,61 @@ class PynisherPlugin(Plugin):
                 `("s", "m", "h")`. Defaults to `None`.
             context: The context to use for multiprocessing. Defaults to `None`.
                 See [`multiprocessing.get_context()`][multiprocessing.get_context]
+            disable_trial_handling: By default, the `PynisherPlugin` will auto-detect
+                if the task is one for a `Trial`. If so, it will catch any pynisher
+                specific exceptions and return a `Trial.Report` with `trial.fail()`,
+                instead of raising the expcetion. This has the effect that the report
[…]
