### 2024-01-22 7422982
fix(Pynisher): Don't crash on non-trial task
