### 2024-01-22 f844a5e
ci: Update pre-commit (#221)
