+++ .github/workflows/test.yml
@@ -1,4 +1,7 @@
 name: tests
+concurrency:
+  group: ${{ github.workflow }}-${{ github.event.pull_request.number || github.ref }}
+  cancel-in-progress: true
 
 on:
   workflow_dispatch:
