+++ .github/workflows/docs.yml
@@ -1,5 +1,8 @@
 # This workflow is just to test that the docs build successfully.
 name: docs
+concurrency:
+  group: ${{ github.workflow }}-${{ github.event.pull_request.number || github.ref }}
+  cancel-in-progress: true
 on:
   workflow_dispatch:
   push:
