+++ .github/workflows/pre-commit.yml
@@ -1,4 +1,7 @@
 name: pre-commit
+concurrency:
+  group: ${{ github.workflow }}-${{ github.event.pull_request.number || github.ref }}
+  cancel-in-progress: true
 on:
   workflow_dispatch:
   push:
