### 2024-01-22 ec6f235
feat(Pynisher): Detect tasks with `Trial` to report `FAIL` (#220)
