+++ src/amltk/__version__.py
@@ -1,3 +1,3 @@
 from __future__ import annotations
 
-version = "1.7.0"
+version = "1.8.0"
