### 2024-01-22 f9f0c78
fix(Pynisher): Ensure system supports limit
