### 2024-01-22 d118960
feat(Optimizer): Allow for batch ask requests
