### 2024-01-22 90bcf52
test(Optimizer): Test for uniqueness of batch ask
