### 2024-01-22 #6133 @JBorrow
Hi @Jammy2211, would you be willing to review this submission to JOSS?
