### 2024-01-22 8a5fb03
test: `ask(n)` should return `n` configs
