### 2024-01-22 eaa5082
feat(Pipeline): `optimize()` directly on a pipeline
