### 2024-01-23 545fcef
feat(Database): Add sqlite integration to store report in a db.
