### 2024-01-23 6ae4f98
feat(Database): Add sqlite integration to store report in a db.
