### 2024-01-23 3836b26
doc: Use env variables to disable running code blocks locally
This was done to speed up the time taken to render docs
