+++ docs/disable_markdown_exec.py
@@ -1,0 +1,45 @@
+"""This disable markdown_exec based on an environment variable.
+
+This is done by overwriting the module responsible for compiling and executing the code
+via overriding it `exec` global varialbe and hijacking it to print a message instead.
+
+https://github.com/pawamoy/markdown-exec/blob/adff40b2928dbb2d22f27684e085f02d39a07291/src/markdown_exec/formatters/python.py#L42-L70
+"""
+from __future__ import annotations
+
+import logging
+import os
+from typing import Any
+
+import mkdocs
+import mkdocs.plugins
+import mkdocs.structure.pages
+
+RUN_CODE_BLOCKS_ENV_VAR = "AMLTK_EXEC_DOCS"
+
+logger = logging.getLogger("mkdocs")
+
+
+def _print_msg(compiled_code: Any, exec_globals: dict) -> None:
+    _print = exec_globals["print"]
+    _print(
+        f"Env variable {RUN_CODE_BLOCKS_ENV_VAR}=0 - No code to display."
+        "\nUse `just docs-code` (or `just docs-full` for examples) to run"
+        " the code block and display output."
+    )
+
+truthy_values = {"yes", "on", "true", "1"}
+
+@mkdocs.plugins.event_priority(100)
+def on_startup(**kwargs: Any):
+    disable = os.environ.get(RUN_CODE_BLOCKS_ENV_VAR, "no")
+    if disable.lower() in falsy_values:
+        logger.warning(
+            f"Disabling markdown-exec due to {RUN_CODE_BLOCKS_ENV_VAR}={disable}"
+            "\n.Use `just docs-full` to run and render examples.",
+        )
+        from markdown_exec.formatters import python
+
+        setattr(python, "exec", _print_msg)
+
+falsy_values = {"", "no", "off", "false", "0"}
