+++ mkdocs.yml
@@ -90,6 +90,10 @@ markdown_extensions:
       emoji_index: !!python/name:material.extensions.emoji.twemoji
       emoji_generator: !!python/name:material.extensions.emoji.to_svg
 
+hooks:
+  - docs/prevent_info_messages.py
+  - docs/disable_markdown_exec.py
+
 plugins:
   - search
   - autorefs
@@ -111,7 +115,7 @@ plugins:
   - mkdocstrings:
       default_handler: python
       enable_inventory: true
-      custom_templates: docs/_templates
+      #custom_templates: docs/_templates
       handlers:
         python:
           paths: [src]
@@ -130,18 +134,20 @@ plugins:
               trim_doctest_flags: true
             show_docstring_attributes: true
             show_docstring_description: true
-            show_root_heading: false
-            show_root_toc_entry: false
+            show_root_heading: true
+            show_root_toc_entry: true
             show_object_full_path: false
+            show_root_members_full_path: true
+            signature_crossrefs: true
             merge_init_into_class: true
+            show_symbol_type_heading: true
+            show_symbol_type_toc: true
             docstring_style: google
             show_if_no_docstring: false
             show_bases: true
             show_source: true
             members_order: "source"
-            # Would like to set `group_by_category` to false
-            # https://github.com/mkdocstrings/mkdocstrings/issues/579
-            group_by_category: true
+            group_by_category: false
             show_signature: true
             separate_signature: false
             show_signature_annotations: false
