### 2024-01-23 2d7da37
doc: Improve clarity between classes
