### 2024-01-23 ecbdbfb
doc: Styling cleanup
