### 2024-01-24 6232114
doc: remove old unused code
