### 2024-01-25 76e03bf
doc: Add description to mkdocs.yml
Also adds a note about being allowed to remove the css
