+++ mkdocs.yml
@@ -1,3 +1,17 @@
+# This project uses mkdocs to generate the documentation.
+# Specifically it uses the mkdocs-material theme, which provides a whole
+# host of nice features and customization
+#
+# mkdocs: https://www.mkdocs.org/getting-started/#getting-started-with-mkdocs
+# mkdocs-material: https://squidfunk.github.io/mkdocs-material/
+#
+# Please refer to these links for more information on how to use mkdocs
+#
+# For serving the docs locally, you can take a look at the `justfile` at
+# the root of this repository, it contains a few commands for generating the docs
+# with different levels of execution.
+#
+# Please refer to individual sections for any additional notes
 site_name: "AutoML-Toolkit"
 repo_url: https://github.com/automl/amltk/
 repo_name: automl/amltk
@@ -43,6 +57,10 @@ theme:
         name: Switch to dark mode
 
 
+# The `mike` versioning provider
+# https://github.com/jimporter/mike
+#
+# This is what allows us to create versioned docs in the github cli
 extra:
   version:
     provider: mike
@@ -52,6 +70,9 @@ extra:
     - icon: fontawesome/brands/twitter
       link: https://twitter.com/automl_org
 
+# We do have some extra custom css
+# If for whatever reason you think this is breaking something,
+# please feel free to remove it.
 extra_css:
   - stylesheets/custom.css
 
@@ -91,8 +112,13 @@ markdown_extensions:
       emoji_index: !!python/name:material.extensions.emoji.twemoji
       emoji_generator: !!python/name:material.extensions.emoji.to_svg
 
+# These are files that are run when serving the docs.
 hooks:
+  # This prevents logging messages from poluting the doc build
   - docs/prevent_info_messages.py
+  # This prevents markdown_exec (plugin) from executing code blocks
+  # dependant on environemnt variables. These env variables are
+  # automatically set with the `justfile` commands to build docs
   - docs/disable_markdown_exec.py
 
 plugins:
@@ -116,10 +142,10 @@ plugins:
   - mkdocstrings:
       default_handler: python
       enable_inventory: true
-      #custom_templates: docs/_templates
       handlers:
         python:
           paths: [src]
+          # Extra objects which allow for linking to external docs
           import:
             - 'https://docs.python.org/3/objects.inv'
             - 'https://numpy.org/doc/stable/objects.inv'
@@ -128,6 +154,9 @@ plugins:
             - 'https://scikit-learn.org/stable/objects.inv'
             - 'https://pytorch.org/docs/stable/objects.inv'
             - 'https://jobqueue.dask.org/en/latest/objects.inv'
+          # Please do not try to change these without having
+          # looked at all of the documentation and seeing if it
+          # causes the API docs to look weird anywhere.
           options:  # https://mkdocstrings.github.io/python/usage/
             docstring_section_style: spacy
             docstring_options:
