### 2024-01-25 0ae52f5
doc: Fixup logging silencing for building docs
