### 2024-01-25 9c60611
doc: Add missing command to hide print command
