### 2024-01-25 bd13a19
doc: Add `black` for mkdocstrings signature formatting
