### 2024-01-25 77109de
doc: Add hook to log which page is being rendered
This only triggers if the environment variable to actually
execute and render code is enabled.
