### 2024-01-25 d3e04b6
doc: Add module description to hooks
