+++ docs/disable_markdown_exec.py
@@ -1,4 +1,5 @@
 """This disable markdown_exec based on an environment variable.
+This speeds up the build of the docs for faster iteration.
 
 This is done by overwriting the module responsible for compiling and executing the code
 via overriding it `exec` global varialbe and hijacking it to print a message instead.
