+++ docs/prevent_info_messages.py
@@ -1,3 +1,10 @@
+"""The module is a hook which disables warnings and log messages which pollute the
+doc build output.
+
+One possible downside is if one of these modules ends up giving an actual
+error, such as OpenML failing to retrieve a dataset. I tried to make sure ERROR
+log message are still allowed through.
+"""
 import logging
 import warnings
 from typing import Any
@@ -29,6 +36,10 @@ def on_pre_page(
 ) -> mkdocs.structure.pages.Page | None:
     # NOTE: mkdocs says they're always normalized to be '/' seperated
     # which means this should work on windows as well.
+
+    # This error is actually demonstrated to the user which causes amltk
+    # to log the error. I don't know how to disable it for that one code cell
+    # put I can at least limit it to the file in which it's in.
     if page.file.src_uri == "guides/scheduling.md":
         scheduling_logger = logging.getLogger("amltk.scheduling.task")
         scheduling_logger.setLevel(logging.CRITICAL)
