### 2024-01-25 da0ca40
doc: Better module doc for disabling markdown_exec
