### 2024-01-25 dc0e0b3
doc: Move hooks to own folder
