### 2024-01-25 5efd647
doc: Fix env variables so default runs everything
