+++ docs/example_runner.py
@@ -143,8 +143,8 @@ class Example:
         if not runnable:
             return False
 
-        env_var = os.environ.get(RUN_EXAMPLES_ENV_VAR, None)
-        if env_var is None:
+        env_var = os.environ.get(RUN_EXAMPLES_ENV_VAR, "false")
+        if env_var in ("false", "", "0", "no", "off"):
             return False
 
         if env_var == "all":
