+++ docs/hooks/debug_which_page_is_being_rendered.py
@@ -22,7 +22,7 @@ def on_pre_page(
     config: Any,
     files: Any,
 ) -> mkdocs.structure.pages.Page | None:
-    render_examples = os.environ.get(RENDER_EXAMPLES_ENV_VAR, "no")
-    render_code = os.environ.get(EXEC_DOCS_ENV_VAR, "no")
+    render_examples = os.environ.get(RENDER_EXAMPLES_ENV_VAR, "true")
+    render_code = os.environ.get(EXEC_DOCS_ENV_VAR, "true")
     if render_examples.lower() in truthy_values or render_code.lower() in truthy_values:
         log.info(f"{page.file.src_path}")
