### 2024-01-25 e8bdbd0
doc: Fix example env var default
