### 2024-01-25 8a878c6
doc: Fix examples env var warning
