### 2024-01-26 #6133 @eddiebergman
Hi @JBorrow, thank you very much for the reminder!

* Martin Hirzel [hirzel](https://github.com/hirzel) - One of the main contributors to IMB Lale AutoML Tool and second author of the paper.
* Felix Mohr [fmohr](https://github.com/fmohr) - Author of ML-Plan.
* Anna Kozak [kozaka93](https://github.com/kozaka93) - Automated machine learning and responsible machine learning
* Katarzyna Woźnica [woznicak](https://github.com/woznicak) - PhD student in Machine Learning and AutoML

Please let me know if you need more names!
