### 2024-01-26 6c8edcb
doc: Cleaner API docs, cleanup pipeline reference, faster rendering options, less noise in doc build output, changelog (#234)
PR #234 Introduced a much cleaner doc layout with the latest version of the mkdocs stack we are using! Along with this, the reference for pipelines was made more streamlined and clear, while the API docs for pipelines was also cleaned up. Also, a viewable change log!

For developers, we defaulted `just docs` to not run any code cells for faster iteration, with the `just docs-code` to now run the code and `just docs-full` to run code and examples.
