+++ src/amltk/pipeline/components.py
@@ -1,73 +1,11 @@
-"""You can use the various different node types to build a pipeline.
-
-You can connect these nodes together using either the constructors explicitly,
-as shown in the examples. We also provide some index operators:
-
-* `>>` - Connect nodes together to form a [`Sequential`][amltk.pipeline.components.Sequential]
-* `&` - Connect nodes together to form a [`Join`][amltk.pipeline.components.Join]
-* `|` - Connect nodes together to form a [`Choice`][amltk.pipeline.components.Choice]
-
-There is also another short-hand that you may find useful to know:
-
-* `{comp1, comp2, comp3}` - This will automatically be converted into a
-    [`Choice`][amltk.pipeline.Choice] between the given components.
-* `(comp1, comp2, comp3)` - This will automatically be converted into a
-    [`Join`][amltk.pipeline.Join] between the given components.
-* `[comp1, comp2, comp3]` - This will automatically be converted into a
-    [`Sequential`][amltk.pipeline.Sequential] between the given components.
-
-For each of these components we will show examples using
-the [`#! "sklearn"` builder][amltk.pipeline.builders.sklearn.build]
-
-The components are:
-
-### Component
-
-::: amltk.pipeline.components.Component
-    options:
-        members: false
-
-### Sequential
-
-::: amltk.pipeline.components.Sequential
-    options:
-        members: false
-
-### Choice
-
-::: amltk.pipeline.components.Choice
-    options:
-        members: false
-
-### Split
-
-::: amltk.pipeline.components.Split
-    options:
-        members: false
-
-### Join
-
-::: amltk.pipeline.components.Join
-    options:
-        members: false
-
-### Fixed
-
-::: amltk.pipeline.components.Fixed
-    options:
-        members: false
-
-### Searchable
-
-::: amltk.pipeline.components.Searchable
-    options:
-        members: false
-"""  # noqa: E501
+"""The provided subclasses of a [`Node`][amltk.pipeline.node.Node]
+that can be used can be assembled into a pipeline.
+"""
 from __future__ import annotations
 
 import inspect
 from collections.abc import Callable, Iterator, Mapping, Sequence
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from typing import TYPE_CHECKING, Any, ClassVar, TypeVar, overload
 from typing_extensions import Self, override
 
@@ -159,93 +97,306 @@ def as_node(  # noqa: PLR0911
 
 
 @dataclass(init=False, frozen=True, eq=True)
-class Join(Node[Item, Space]):
-    """[`Join`][amltk.pipeline.Join] together different parts of the pipeline.
+class Component(Node[Item, Space]):
+    """A [`Component`][amltk.pipeline.Component] of the pipeline with
+    a possible item and **no children**.
 
-    This indicates the different children in
-    [`.nodes`][amltk.pipeline.Node.nodes] should act in tandem with one
-    another, for example, concatenating the outputs of the various members of the
-    `Join`.
+    This is the basic building block of most pipelines, it accepts
+    as it's [`item=`][amltk.pipeline.node.Node.item] some function that will be
+    called with [`build_item()`][amltk.pipeline.components.Component.build_item] to
+    build that one part of the pipeline.
+
+    When [`build_item()`][amltk.pipeline.Component.build_item] is called, whatever
+    the config of the component is at that time, will be used to construct the item.
+
+    A common pattern is to use a [`Component`][amltk.pipeline.Component] to
+    wrap a constructor, specifying the [`space=`][amltk.pipeline.node.Node.space]
+    and [`config=`][amltk.pipeline.node.Node.config] to be used when building the
+    item.
 
     ```python exec="true" source="material-block" html="true"
[…]
