### 2024-01-26 66313fb
refactor(pipeline): Factor out setting up the optimization loop
