+++ src/amltk/pipeline/node.py
@@ -44,6 +44,7 @@ import warnings
 from collections.abc import Callable, Iterable, Iterator, Mapping, Sequence
 from dataclasses import dataclass, field
 from datetime import datetime
+from functools import partial
 from pathlib import Path
 from typing import (
     TYPE_CHECKING,
@@ -878,75 +879,7 @@ class Node(RichRenderable, Generic[Item, Space]):
 
         return new_config
 
-    @overload
-    def optimize(
-        self,
-        target: (
-            Callable[[Trial, Node], Trial.Report]
-            | Task[[Trial, Node], Trial.Report]
-            | EvaluationProtocol
-        ),
-        metric: Metric | Sequence[Metric],
-        *,
-        # With setup_only, we return a Scheduler and the `run()` arguments
-        # will be ignored
-        setup_only: Literal[True],
-        # `scheduler.run()`
-        # * display: bool = True,
-        # * wait: bool = True,
-        # * on_scheduler_exception: Literal["raise", "end", "continue"] = "raise",
-        optimizer: type[Optimizer] | Optimizer.CreateSignature | None = ...,
-        seed: Seed | None = ...,
-        working_dir: str | Path | PathBucket | None = ...,
-        plugins: Plugin | Iterable[Plugin] | None = ...,
-        n_workers: int = ...,
-        scheduler: Scheduler | None = ...,
-        history: History | None = ...,
-        max_trials: int | None = ...,
-        on_trial_exception: Literal["raise", "end", "continue"] = "raise",
-        process_memory_limit: int | tuple[int, str] | None = ...,
-        process_walltime_limit: int | tuple[float, str] | None = ...,
-        process_cputime_limit: int | tuple[float, str] | None = ...,
-        threadpool_limit_ctl: bool | None = ...,
-        on_begin: OnBeginCallbackSignature | None = ...,
-    ) -> Scheduler:
-        ...
-
-    @overload
-    def optimize(
-        self,
-        target: (
-            Callable[[Trial, Node], Trial.Report]
-            | Task[[Trial, Node], Trial.Report]
-            | EvaluationProtocol
-        ),
-        metric: Metric | Sequence[Metric],
-        *,
-        # If setup only is false (default), we use `run()` arguments and will
-        # just return the history
-        setup_only: Literal[False] = False,
-        optimizer: type[Optimizer] | Optimizer.CreateSignature | None = ...,
-        seed: Seed | None = ...,
-        timeout: float = ...,
-        working_dir: str | Path | PathBucket | None = ...,
-        plugins: Plugin | Iterable[Plugin] | None = ...,
-        n_workers: int = ...,
-        scheduler: Scheduler | None = ...,
-        history: History | None = ...,
-        max_trials: int | None = ...,
-        on_trial_exception: Literal["raise", "end", "continue"] = "raise",
-        process_memory_limit: int | tuple[int, str] | None = ...,
-        process_walltime_limit: int | tuple[float, str] | None = ...,
-        process_cputime_limit: int | tuple[float, str] | None = ...,
-        on_begin: OnBeginCallbackSignature | None = ...,
-        display: bool = ...,
-        wait: bool = ...,
-        threadpool_limit_ctl: bool | None = ...,
-        on_scheduler_exception: Literal["raise", "end", "continue"] = ...,
-    ) -> History:
-        ...
-
-    def optimize(  # noqa: C901, PLR0915, PLR0912
+    def register_optimization_loop(  # noqa: C901, PLR0915, PLR0912
         self,
         target: (
             EvaluationProtocol
@@ -959,7 +892,6 @@ class Node(RichRenderable, Generic[Item, Space]):
         seed: Seed | None = None,
         max_trials: int | None = 3,
         n_workers: int = 1,
[…]
