### 2024-01-26 ef9f33b
feat(rich): `is_jupyter()` function to detect if running in display context
