### 2024-01-26 7fa02bc
feat(scheduler): "auto" to display `run()` in notebook by default
