+++ src/amltk/_richutil/__init__.py
@@ -1,6 +1,6 @@
 from amltk._richutil.renderable import RichRenderable
 from amltk._richutil.renderers import Function, rich_make_column_selector
-from amltk._richutil.util import df_to_table, richify
+from amltk._richutil.util import df_to_table, is_jupyter, richify
 
 __all__ = [
     "df_to_table",
@@ -8,4 +8,5 @@ __all__ = [
     "RichRenderable",
     "Function",
     "rich_make_column_selector",
+    "is_jupyter",
 ]
