### 2024-01-26 f27db12
feat(pipeline): propogate `"auto"` for display to `optimize()`
