### 2024-01-26 6874f42
fix(pipeline): Correctly add threadpool ctl to plugins
