### 2024-01-26 edd24b9
doc(optimizer): Correct counting typo (grrr)
