### 2024-01-26 377abab
doc: More verbosity in description of threadpool heuristic
