### 2024-01-26 8b69e40
fix(pipeline): Another `plugin` varname bug -_-
