### 2024-01-26 bded378
feat(Pipeline): Optimize pipelines directly with `optimize()` (#230)
