+++ src/amltk/pipeline/node.py
@@ -36,11 +36,16 @@ directly, but instead use the various components to create the pipeline.
     means two nodes are considered equal if they look the same and they are
     connected in to nodes that also look the same.
 """  # noqa: E501
+# ruff: noqa: PLR0913
 from __future__ import annotations
 
 import inspect
-from collections.abc import Callable, Iterator, Mapping, Sequence
+import warnings
+from collections.abc import Callable, Iterable, Iterator, Mapping, Sequence
 from dataclasses import dataclass, field
+from datetime import datetime
+from functools import partial
+from pathlib import Path
 from typing import (
     TYPE_CHECKING,
     Any,
@@ -50,6 +55,7 @@ from typing import (
     Literal,
     NamedTuple,
     ParamSpec,
+    Protocol,
     TypeAlias,
     TypeVar,
     overload,
@@ -59,10 +65,15 @@ from typing_extensions import override
 from more_itertools import first_true
 from sklearn.pipeline import Pipeline as SklearnPipeline
 
-from amltk._functional import classname, mapping_select, prefix_keys
+from amltk._functional import classname, funcname, mapping_select, prefix_keys
 from amltk._richutil import RichRenderable
 from amltk.exceptions import RequestNotMetError
-from amltk.types import Config, Item, Space
+from amltk.optimization.history import History
+from amltk.optimization.optimizer import Optimizer
+from amltk.scheduling import Task
+from amltk.scheduling.plugins import Plugin
+from amltk.store import PathBucket
+from amltk.types import Config, Item, Seed, Space
 
 if TYPE_CHECKING:
     from typing_extensions import Self
@@ -71,8 +82,12 @@ if TYPE_CHECKING:
     from rich.console import RenderableType
     from rich.panel import Panel
 
+    from amltk.evalutors.evaluation_protocol import EvaluationProtocol
+    from amltk.optimization.metric import Metric
+    from amltk.optimization.trial import Trial
     from amltk.pipeline.components import Choice, Join, Sequential
     from amltk.pipeline.parsers.optuna import OptunaSearchSpace
+    from amltk.scheduling import Scheduler
 
     NodeLike: TypeAlias = (
         set["Node" | "NodeLike"]
@@ -83,6 +98,81 @@ if TYPE_CHECKING:
     )
 
     SklearnPipelineT = TypeVar("SklearnPipelineT", bound=SklearnPipeline)
+
+
+class OnBeginCallbackSignature(Protocol):
+    """A  calllback to further define control flow from
+    [`pipeline.optimize()`][amltk.pipeline.node.Node.optimize].
+
+    In one of these callbacks, you can register to specific `@events` of the
+    [`Scheduler`][amltk.scheduling.Scheduler] or [`Task`][amltk.scheduling.Task].
+
+    ```python
+    pipeline = ...
+
+    # The callback will get the task, scheduler and the history in which results
+    # will be stored
+    def my_callback(task: Task[..., Trial.Report], scheduler: Scheduler, history: History) -> None:
+
+        # You can do early stopping based on a target metric
+        @task.on_result
+        def stop_if_target_reached(_: Future, report: Trial.Report) -> None:
+            score = report.metrics["accuracy"]
+            if score >= 0.95:
+                scheduler.stop(stop_msg="Target reached!"))
+
+        # You could also perform early stopping based on iterations
+        n = 0
+        last_score = 0.0
+
+        @task.on_result
+        def stop_if_no_improvement_for_n_runs(_: Future, report: Trial.Report) -> None:
+            score = report.metrics["accuracy"]
+            if score > last_score:
+                n = 0
+                last_score = score
+            elif n >= 5:
+                scheduler.stop()
+            else:
+                n += 1
+
+        # Really whatever you'd like
+        @task.on_result
[…]
