+++ 1.10.0/example_runner.py
@@ -1,0 +1,295 @@
+"""Generates the examples pages."""
+from __future__ import annotations
+
+import logging
+import os
+import textwrap
+from dataclasses import dataclass
+from itertools import takewhile
+from pathlib import Path
+from typing import Any
+from typing_extensions import override
+
+import mkdocs_gen_files
+from more_itertools import first_true, peekable
+
+logger = logging.getLogger("mkdocs")
+
+RUN_EXAMPLES_ENV_VAR = "AMLTK_DOC_RENDER_EXAMPLES"
+
+
+@dataclass
+class CodeSegment:
+    lines: list[str]
+    session: str
+    exec: bool
+
+    def code(self, code: list[str]) -> str:
+        points_start = first_true(code, pred=lambda _l: _l.startswith("# 1."))
+        if points_start is not None:
+            points_start_index = code.index(points_start)
+
+            points = [""]
+            points.extend([_l.lstrip("#")[1:] for _l in code[points_start_index:]])
+            points.append("")
+
+            body = code[:points_start_index]
+        else:
+            points = []
+            body = code
+
+        # Trim off any excess leading lines which only have whitespace
+        while body and body[0].strip() == "":
+            body.pop(0)
+
+        hl_lines = []
+        for i, line in enumerate(body):
+            _l = line.rstrip()
+            if "<!>" in _l:
+                hl_lines.append(str(i + 1))
+                _l = _l.replace("<!>", "").rstrip()
+
+            for strip in ["# !", "#!", "#"]:
+                if _l.endswith(strip):
+                    _l = _l[: -len(strip)]
+
+            body[i] = _l
+
+        if any(hl_lines):
+            hl_lines = " ".join(hl_lines)
+            hl_string = f'hl_lines="{hl_lines}"'
+        else:
+            hl_string = ""
+
+        # We generate two tabs if executing
+        if self.exec:
+            indented_body = "\n".join(f"    {_l}" for _l in body)
+
+            code_annotations = " ".join(
+                [
+                    "{",
+                    ".python",
+                    ".annotate",
+                    hl_string,
+                    "}",
+                ],
+            )
+            tab1 = "\n".join(
+                [
+                    '=== "Code"',
+                    "",
+                    f"    ``` {code_annotations}",
+                    indented_body,
+                    "    ```",
+                    *[f"    {point}" for point in points],
+                    "",
+                ],
+            )
+
+            run_annotations = " ".join(
+                [
+                    "{",
+                    ".python",
+                    f"session='{self.session}'",
+                    'exec="True"',
+                    'result="python"',
+                    "}",
+                ],
+            )
+
+            tab2 = "\n".join(
+                [
+                    '=== "Run"',
+                    "",
+                    f"    ``` {run_annotations}",
+                    indented_body,
+                    "    ```",
+                ],
+            )
+
+            return "\n".join([tab1, "", tab2])
+
+        annotations = " ".join(["{", ".python", ".annotate", hl_string, "}"])
+        top = f"```{annotations}"
+        bottom = "```"
+
+        s = [top, *body, bottom, *points]
+        body = "\n".join(s)
+        return body
+
+    @override
+    def __str__(self) -> str:
+        return self.code(self.lines)
+
+
+@dataclass
+class CommentSegment:
+    lines: list[str]
+
+    @override
+    def __str__(self) -> str:
+        return "\n".join(self.lines)
+
+
+@dataclass
+class Example:
+    name: str
+    filepath: Path
+    description: str
+    segments: list[CodeSegment | CommentSegment]
+
+    @classmethod
+    def should_execute(cls, *, name: str, runnable: bool) -> bool:
+        if not runnable:
+            return False
+
+        env_var = os.environ.get(RUN_EXAMPLES_ENV_VAR, "all")
[…]
