### 2024-01-26 fef5795
feat(metric): Create a metric definition from an sklearn scorer
