### 2024-01-26 b76deca
doc(metric): Add hint why you might want to use `from_sklearn()`
