### 2024-01-26 c129092
fix: Sklearn `as_metric` uses _name and gives better warning
