### 2024-01-27 33c2263
refactor(trial): Deprecate default behaviour of `begin()`
Currently `begin()` will silently eat errors and relies on an
explicit idiom of checking the exception after the block.
While this is convenient, it's silent behaviour and it's better
to have users explicitly enable it.
