### 2024-01-27 7a1a24c
doc: Fixup doc string of change
