### 2024-01-27 da43b1f
doc: Update new usage on error handling where appropriate
