### 2024-01-27 1930d62
refactor(trail): Deprecated `begin()`
