+++ src/amltk/optimization/trial.py
@@ -28,7 +28,8 @@ from __future__ import annotations
 
 import copy
 import logging
-import traceback
+import traceback as traceback_module
+import warnings
 from collections.abc import Callable, Iterable, Iterator, Mapping, Sequence
 from contextlib import contextmanager
 from dataclasses import dataclass, field
@@ -36,7 +37,7 @@ from datetime import datetime
 from enum import Enum
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, Generic, Literal, TypeVar, overload
-from typing_extensions import ParamSpec, Self, override
+from typing_extensions import ParamSpec, Self, deprecated, override
 
 import numpy as np
 import pandas as pd
@@ -264,9 +265,11 @@ class Trial(RichRenderable, Generic[I]):
     """The summary of the trial. These are for summary statistics of a trial and
     are single values."""
 
-    exception: BaseException | None = field(repr=True, default=None)
+    # @deprecated(1.11.0)
+    exception: Exception | None = field(repr=True, default=None)
     """The exception raised by the trial, if any."""
 
+    # @deprecated(1.11.0)
     traceback: str | None = field(repr=False, default=None)
     """The traceback of the exception, if any."""
 
@@ -283,6 +286,14 @@ class Trial(RichRenderable, Generic[I]):
         """The profiles of the trial."""
         return self.profiler.profiles
 
+    @deprecated(
+        "`trial.begin() silently catches exceptions and attaches them to"
+        " To be more explicit and have the same functionality, please use"
+        " `with trial.profile('some tag'):` to profile and manually use"
+        " `try: ... except: ...` to catch exceptions and report them"
+        " with `trial.fail(exception)`."
+        "\nDeprecated in 1.11.0",
+    )
     @contextmanager
     def begin(
         self,
@@ -290,6 +301,18 @@ class Trial(RichRenderable, Generic[I]):
         memory_unit: Memory.Unit | Literal["B", "KB", "MB", "GB"] | None = None,
     ) -> Iterator[None]:
         """Begin the trial with a `contextmanager`.
+
+        !!! warning "Deprecated"
+
+            This method is deprecated, please use the following idiom instead:
+
+            ```python
+            with trial.profile("trial"):
+                try:
+                    # Do some work
+                except Exception as error:
+                    return trial.fail(error)
+            ```
 
         Will begin timing the trial in the `with` block, attaching the profiled time and memory
         to the trial once completed, under `.profile.time` and `.profile.memory` attributes.
@@ -336,7 +359,7 @@ class Trial(RichRenderable, Generic[I]):
                 yield
             except Exception as error:  # noqa: BLE001
                 self.exception = error
-                self.traceback = traceback.format_exc()
+                self.traceback = traceback_module.format_exc()
             finally:
                 self.time = self.profiler["trial"].time
                 self.memory = self.profiler["trial"].memory
@@ -438,7 +461,13 @@ class Trial(RichRenderable, Generic[I]):
             metric_values=tuple(_recorded_values),
         )
 
-    def fail(self, **metrics: float | int) -> Trial.Report[I]:
+    def fail(
+        self,
+        exception: Exception | None = None,
+        traceback: str | None = None,
+        /,
+        **metrics: float | int,
+    ) -> Trial.Report[I]:
         """Generate a failure report.
 
         !!! note "Non specifed metrics"
@@ -467,22 +496,48 @@ class Trial(RichRenderable, Generic[I]):
         Returns:
             The result of the trial.
         """
-        _recorded_values: list[Metric.Value] = []
-        for _metric in self.metrics:
-            if (raw_value := metrics.get(_metric.name)) is not None:
-                _recorded_values.append(_metric.as_value(raw_value))
-            else:
[…]
