### 2024-01-27 54aa2b5
doc: Scrub away `begin()`
