### 2024-01-27 b2d06ab
fix: Remove reference to trial.exception
