### 2024-01-28 96c6a82
fix: Add timestamp `generated_at` to report for sorting purposes
