### 2024-01-28 da0b6a9
refactor(trial): Remove `begin()` (#238)
