+++ src/amltk/optimization/optimizers/smac.py
@@ -46,11 +46,14 @@ def target_function(trial: Trial, pipeline: Node) -> Trial.Report:
     X_train, X_test, y_train, y_test = train_test_split(X, y)
     clf = pipeline.configure(trial.config).build("sklearn")
 
-    with trial.begin():
-        clf.fit(X_train, y_train)
-        y_pred = clf.predict(X_test)
-        accuracy = accuracy_score(y_test, y_pred)
-        return trial.success(accuracy=accuracy)
+    with trial.profile("trial"):
+        try:
+            clf.fit(X_train, y_train)
+            y_pred = clf.predict(X_test)
+            accuracy = accuracy_score(y_test, y_pred)
+            return trial.success(accuracy=accuracy)
+        except Exception as e:
+            return trial.fail(e)
 
     return trial.fail()
 from amltk._doc import make_picklable; make_picklable(target_function)  # markdown-exec: hide
@@ -109,6 +112,7 @@ from smac.runhistory import (
 
 from amltk.optimization import Metric, Optimizer, Trial
 from amltk.pipeline import Node
+from amltk.profiling import Profile
 from amltk.randomness import as_int
 from amltk.store import PathBucket
 
@@ -140,6 +144,7 @@ class SMACOptimizer(Optimizer[SMACTrialInfo]):
         bucket: PathBucket | None = None,
         metrics: Metric | Sequence[Metric],
         fidelities: Mapping[str, FidT] | None = None,
+        time_profile: str | None = None,
     ) -> None:
         """Initialize the optimizer.
 
@@ -148,6 +153,9 @@ class SMACOptimizer(Optimizer[SMACTrialInfo]):
             bucket: The bucket given to trials generated by this optimizer.
             metrics: The metrics to optimize.
             fidelities: The fidelities to use, if any.
+            time_profile: The profile to use to get time information to the
+                optimizer. Must use `trial.profile(time_profile)` in your
+                target function then.
         """
         # We need to very that the scenario is correct incase user pass in
         # their own facade construction
@@ -158,7 +166,9 @@ class SMACOptimizer(Optimizer[SMACTrialInfo]):
         self.facade = facade
         self.metrics = metrics
         self.fidelities = fidelities
+        self.time_profile = time_profile
 
+    @override
     @classmethod
     def create(
         cls,
@@ -166,6 +176,7 @@ class SMACOptimizer(Optimizer[SMACTrialInfo]):
         space: ConfigurationSpace | Node,
         metrics: Metric | Sequence[Metric],
         bucket: PathBucket | str | Path | None = None,
+        time_profile: str | None = None,
         deterministic: bool = True,
         seed: Seed | None = None,
         fidelities: Mapping[str, FidT] | None = None,
@@ -179,6 +190,9 @@ class SMACOptimizer(Optimizer[SMACTrialInfo]):
             space: The config space to optimize.
             metrics: The metrics to optimize.
             bucket: The bucket given to trials generated by this optimizer.
+            time_profile: The profile to use to get time information to the
+                optimizer. Must use `trial.profile(time_profile)` in your
+                target function then.
             deterministic: Whether the function your optimizing is deterministic, given
                 a seed and config.
             seed: The seed to use for the optimizer.
@@ -247,7 +261,13 @@ class SMACOptimizer(Optimizer[SMACTrialInfo]):
                 scenario=scenario,
             ),
         )
-        return cls(facade=facade, fidelities=fidelities, bucket=bucket, metrics=metrics)
+        return cls(
+            facade=facade,
+            fidelities=fidelities,
+            bucket=bucket,
+            metrics=metrics,
+            time_profile=time_profile,
+        )
 
     @overload
[…]
