### 2024-01-28 a3a803b
refactor: Make things more context manager
