+++ src/amltk/scheduling/scheduler.py
@@ -256,6 +256,7 @@ import warnings
 from asyncio import Future
 from collections.abc import Callable, Iterable, Mapping
 from concurrent.futures import CancelledError, Executor, ProcessPoolExecutor
+from contextlib import nullcontext
 from dataclasses import dataclass
 from enum import Enum, auto
 from functools import partial
@@ -275,6 +276,7 @@ from uuid import uuid4
 from amltk._asyncm import ContextEvent
 from amltk._functional import Flag, subclass_map
 from amltk._richutil.renderable import RichRenderable
+from amltk._util import ignore_warnings, mutli_context
 from amltk.exceptions import SchedulerNotRunningError
 from amltk.scheduling.events import Emitter, Event, Subscriber
 from amltk.scheduling.executors import SequentialExecutor
@@ -497,7 +499,7 @@ class Scheduler(RichRenderable):
         self._queue_has_items_event = asyncio.Event()
 
         # This is triggered when run is called
-        self._running_event = asyncio.Event()
+        self._running_event = ContextEvent()
 
         # This is set once `run` is called.
         # Either contains the mapping from exception to what to do,
@@ -1170,123 +1172,120 @@ class Scheduler(RichRenderable):
         logger.debug("Stop event triggered, stopping scheduler")
         return True
 
-    async def _run_scheduler(  # noqa: C901, PLR0912, PLR0915
+    async def _run_scheduler(  # noqa: C901, PLR0915
         self,
         *,
         timeout: float | None = None,
         end_on_empty: bool = True,
         wait: bool = True,
     ) -> ExitState.Code | BaseException:
-        self.executor.__enter__()
-        self._stop_event = ContextEvent()
+        with mutli_context(
+            self.executor,
+            self._live_output if self._live_output is not None else nullcontext(),
+            ignore_warnings() if self._live_output is not None else nullcontext(),
+        ):
+            self._stop_event = ContextEvent()
 
-        if self._live_output is not None:
-            self._live_output.__enter__()
+            # Start a Thread Timer as our timing mechanism.
+            # HACK: This is required because the SequentialExecutor mode
+            # will not allow the async loop to run, meaning we can't update
+            # any internal state. For this reason, we use a Thread
+            if timeout is not None:
+                self._timeout_timer = Timer(timeout, lambda: None)
+                self._timeout_timer.start()
 
-            # If we are doing a live display, we have to disable
-            # warnings as they will screw up the display rendering
-            # However, we re-enable it after the scheduler has finished running
-            warning_catcher = warnings.catch_warnings()
-            warning_catcher.__enter__()
-            warnings.filterwarnings("ignore")
-        else:
-            warning_catcher = None
+            with self._running_event:
+                self.on_start.emit()
 
-        # Declare we are running
-        self._running_event.set()
+                # Monitor for `stop` being triggered
+                stop_triggered = asyncio.create_task(
+                    self._stop_when_triggered(self._stop_event),
+                )
 
-        # Start a Thread Timer as our timing mechanism.
-        # HACK: This is required because the SequentialExecutor mode
-        # will not allow the async loop to run, meaning we can't update
-        # any internal state.
-        if timeout is not None:
-            self._timeout_timer = Timer(timeout, lambda: None)
-            self._timeout_timer.start()
+                # Monitor for the queue being empty
+                monitor_empty = asyncio.create_task(self._monitor_queue_empty())
+                if end_on_empty:
+                    self.on_empty(lambda: monitor_empty.cancel(), hidden=True)
 
-        self.on_start.emit()
+                # The timeout criterion is satisfied by the `timeout` arg
+                await asyncio.wait(
+                    [stop_triggered, monitor_empty],
+                    timeout=timeout,
[…]
