### 2024-01-28 6615af6
fix(dask-jobqueue): Make sure to close client
