### 2024-01-28 844ce6a
test: Fix filesockets open leakage from dask
