+++ tests/scheduling/plugins/test_comm_plugin.py
@@ -90,6 +90,8 @@ def case_dask_executor() -> ClientExecutor:
 @parametrize_with_cases("executor", cases=".", has_tag="executor")
 def scheduler(executor: Executor) -> Iterator[Scheduler]:
     yield Scheduler(executor)
+    if isinstance(executor, ClientExecutor):
+        executor._client.close()
 
 
 def test_sending_worker(scheduler: Scheduler) -> None:
