+++ pyproject.toml
@@ -55,7 +55,7 @@ smac = ["smac>=2.0", "amltk[configspace]"]
 optuna = ["optuna"]
 configspace = ["configspace>=0.6"]
 loky = ["loky"]
-dask = ["dask<=2023.4", "distributed"]
+dask = ["dask", "distributed"]
 pynisher = ["pynisher>=1.0.10"]
 wandb = ["wandb"]
 threadpoolctl = ["threadpoolctl"]
