### 2024-01-28 a2a97d4
refactor(trial): Use a mapping for metrics
This is to allow a more native `trial.metrics["name"]` usage
