### 2024-01-28 6888bb6
feat(Metric): Allow for attached fn
