### 2024-01-28 23f42ad
typing(Metric): more permissive Mapping to from_colection
