### 2024-01-28 951a27a
refactor(Metric): remove Metric.Value
