+++ src/amltk/optimization/metric.py
@@ -29,7 +29,7 @@ from __future__ import annotations
 
 from collections.abc import Callable, Iterable, Iterator, Mapping
 from dataclasses import dataclass, field
-from typing import TYPE_CHECKING, Generic, ParamSpec
+from typing import TYPE_CHECKING, Any, Generic, Literal, ParamSpec
 from typing_extensions import Self, override
 
 if TYPE_CHECKING:
@@ -37,6 +37,8 @@ if TYPE_CHECKING:
 
 
 P = ParamSpec("P")
+
+SklearnResponseMethods = Literal["predict", "predict_proba", "decision_function"]
 
 
 @dataclass(frozen=True)
@@ -86,8 +88,25 @@ class Metric(Generic[P]):
             )
         return self.fn(*args, **kwargs)
 
-    def as_sklearn_scorer(self) -> _Scorer:
-        """Convert a metric to a sklearn scorer."""
+    def as_scorer(
+        self,
+        *,
+        response_method: (
+            SklearnResponseMethods | Iterable[SklearnResponseMethods] | None
+        ) = None,
+        **scorer_kwargs: Any,
+    ) -> _Scorer:
+        """Convert a metric to a sklearn scorer.
+
+        Args:
+            response_method: The response method to use for the scorer.
+                This can be a single method or an iterable of methods.
+            scorer_kwargs: Additional keyword arguments to pass to the
+                scorer during the call. Forwards to [`sklearn.metrics.make_scorer`][].
+
+        Returns:
+            The sklearn scorer.
+        """
         from sklearn.metrics import get_scorer, make_scorer
 
         match self.fn:
@@ -100,7 +119,12 @@ class Metric(Generic[P]):
                         " Please provide a function to `Metric(fn=...)`.",
                     ) from e
             case fn:
-                return make_scorer(fn, greater_is_better=not self.minimize)
+                return make_scorer(
+                    fn,
+                    greater_is_better=not self.minimize,
+                    response_method=response_method,
+                    **scorer_kwargs,
+                )
 
     @override
     def __str__(self) -> str:
@@ -146,94 +170,42 @@ class Metric(Generic[P]):
         return cls(name=name, minimize=minimize, bounds=bounds)
 
     @property
-    def worst(self) -> Metric.Value:
+    def worst(self) -> float:
         """The worst possible value of the metric."""
-        if self.bounds:
-            v = self.bounds[1] if self.minimize else self.bounds[0]
-            return self.as_value(v)
+        if self.bounds is not None:
+            return self.bounds[1] if self.minimize else self.bounds[0]
 
-        v = float("inf") if self.minimize else float("-inf")
-        return self.as_value(v)
+        return float("inf") if self.minimize else float("-inf")
 
     @property
-    def optimal(self) -> Metric.Value:
+    def optimal(self) -> float:
         """The optimal value of the metric."""
         if self.bounds:
-            v = self.bounds[0] if self.minimize else self.bounds[1]
-            return self.as_value(v)
-        v = float("-inf") if self.minimize else float("inf")
-        return self.as_value(v)
+            return self.bounds[0] if self.minimize else self.bounds[1]
 
-    def as_value(self, value: float | int) -> Metric.Value:
-        """Convert a value to an metric value."""
-        return Metric.Value(metric=self, value=float(value))
+        return float("-inf") if self.minimize else float("inf")
 
-    @dataclass(frozen=True, order=True)
-    class Value:
-        """A recorded value of an metric."""
+    def distance_to_optimal(self, v: float, /) -> float | None:
+        """The distance to the optimal value, using the bounds if possible."""
+        match self.bounds:
+            case None:
+                return None
+            case (lower, upper) if lower <= v <= upper:
[…]
