### 2024-01-28 69ca8eb
refactor(Trial): simply metric access
