+++ src/amltk/optimization/trial.py
@@ -29,7 +29,14 @@ from __future__ import annotations
 import copy
 import logging
 import traceback as traceback_module
-from collections.abc import Callable, Iterable, Iterator, Mapping
+from collections.abc import (
+    Callable,
+    Hashable,
+    Iterable,
+    Iterator,
+    Mapping,
+    MutableMapping,
+)
 from contextlib import contextmanager
 from dataclasses import dataclass, field
 from datetime import datetime
@@ -44,7 +51,7 @@ import pandas as pd
 from amltk._functional import dict_get_not_none, mapping_select, prefix_keys
 from amltk._richutil.renderable import RichRenderable
 from amltk._util import parse_timestamp_object
-from amltk.optimization.metric import Metric
+from amltk.optimization.metric import Metric, MetricCollection
 from amltk.profiling import Memory, Profile, Profiler, Timer
 from amltk.store import Bucket, PathBucket
 
@@ -225,40 +232,94 @@ class Trial(RichRenderable, Generic[I]):
     config: Mapping[str, Any]
     """The config of the trial provided by the optimizer."""
 
-    bucket: PathBucket = field(
-        default_factory=lambda: PathBucket("unknown-trial-bucket"),
-    )
+    bucket: PathBucket
     """The bucket to store trial related output to."""
 
-    info: I | None = field(default=None, repr=False)
+    info: I | None = field(repr=False)
     """The info of the trial provided by the optimizer."""
 
-    metrics: Mapping[str, Metric] = field(default_factory=dict)
-    """The metrics associated with the trial."""
+    metrics: MetricCollection
+    """The metrics associated with the trial.
+
+    You can access the metrics by name, e.g. `#!python trial.metrics["loss"]`.
+    """
 
     seed: int | None = None
     """The seed to use if suggested by the optimizer."""
 
-    fidelities: dict[str, Any] | None = None
+    fidelities: Mapping[str, Any]
     """The fidelities at which to evaluate the trial, if any."""
 
-    profiler: Profiler = field(
-        repr=False,
-        default_factory=lambda: Profiler(memory_unit="B", time_kind="wall"),
-    )
+    profiler: Profiler = field(repr=False)
     """A profiler for this trial."""
 
-    summary: dict[str, Any] = field(default_factory=dict)
+    summary: MutableMapping[str, Any]
     """The summary of the trial. These are for summary statistics of a trial and
     are single values."""
 
-    storage: set[Any] = field(default_factory=set)
+    storage: set[Any]
     """Anything stored in the trial, the elements of the list are keys that can be
     used to retrieve them later, such as a Path.
     """
 
-    extras: dict[str, Any] = field(default_factory=dict)
+    extras: MutableMapping[str, Any]
     """Any extras attached to the trial."""
+
+    @classmethod
+    def create(  # noqa: PLR0913
+        cls,
+        name: str,
+        config: Mapping[str, Any] | None = None,
+        *,
+        metrics: Metric | Iterable[Metric] | Mapping[str, Metric] | None = None,
+        info: I | None = None,
+        seed: int | None = None,
+        fidelities: Mapping[str, Any] | None = None,
+        profiler: Profiler | None = None,
+        bucket: PathBucket | None = None,
+        summary: MutableMapping[str, Any] | None = None,
+        storage: set[Hashable] | None = None,
+        extras: MutableMapping[str, Any] | None = None,
+    ) -> Trial[I]:
+        """Create a trial.
+
+        Args:
+            name: The name of the trial.
+            metrics: The metrics of the trial.
+            config: The config of the trial.
+            info: The info of the trial.
+            seed: The seed of the trial.
+            fidelities: The fidelities of the trial.
+            bucket: The bucket of the trial.
[…]
