### 2024-01-28 37b8d41
refactor(trial): Cleanup the  part of trial storage
