+++ src/amltk/optimization/trial.py
@@ -30,7 +30,6 @@ import copy
 import logging
 import traceback as traceback_module
 from collections.abc import (
-    Callable,
     Hashable,
     Iterable,
     Iterator,
@@ -510,34 +509,15 @@ class Trial(RichRenderable, Generic[I]):
             traceback=traceback,
         )
 
-    def store(
-        self,
-        items: Mapping[str, T],
-        *,
-        where: (
-            str | Path | Bucket | Callable[[str, Mapping[str, T]], None] | None
-        ) = None,
-    ) -> None:
+    def store(self, items: Mapping[str, T]) -> None:
         """Store items related to the trial.
 
         ```python exec="true" source="material-block" result="python" title="store" hl_lines="5"
         from amltk.optimization import Trial
         from amltk.store import PathBucket
 
-        trial = Trial(name="trial", config={"x": 1}, bucket=PathBucket("results"))
+        trial = Trial(name="trial", config={"x": 1}, bucket=PathBucket("my-trial"))
         trial.store({"config.json": trial.config})
-
-        print(trial.storage)
-        ```
-
-        You could also specify `where=` exactly to store the thing
-
-        ```python exec="true" source="material-block" result="python" title="store-bucket" hl_lines="7"
-        from amltk.optimization import Trial
-
-        trial = Trial(name="trial", config={"x": 1})
-        trial.store({"config.json": trial.config}, where="./results")
-
         print(trial.storage)
         ```
 
@@ -546,48 +526,12 @@ class Trial(RichRenderable, Generic[I]):
                 to the item itself.If using a `str`, `Path` or `PathBucket`,
                 the keys of the items should be a valid filename, including
                 the correct extension. e.g. `#!python {"config.json": trial.config}`
-
-            where: Where to store the items.
-
-                * If `None`, will use the bucket attached to the `Trial` if any,
-                    otherwise it will raise an error.
-
-                * If a `str` or `Path`, will store
-                a bucket will be created at the path, and the items will be
-                stored in a sub-bucket with the name of the trial.
-
-                * If a `Bucket`, will store the items **in a sub-bucket** with the
-                name of the trial.
-
-                * If a `Callable`, will call the callable with the name of the
-                trial and the key-valued pair of items to store.
         """  # noqa: E501
-        method: Bucket
-        match where:
-            case None:
-                method = self.bucket
-                method.sub(self.name).store(items)
-            case str() | Path():
-                method = PathBucket(where, create=True)
-                method.sub(self.name).store(items)
-            case Bucket():
-                method = where
-                method.sub(self.name).store(items)
-            case _:
-                # Leave it up to supplied method
-                where(self.name, items)
-
+        self.bucket.store(items)
         # Add the keys to storage
         self.storage.update(items)
 
-    def delete_from_storage(
-        self,
-        items: Iterable[str],
-        *,
-        where: (
-            str | Path | Bucket | Callable[[str, Iterable[str]], dict[str, bool]] | None
-        ) = None,
-    ) -> dict[str, bool]:
+    def delete_from_storage(self, items: Iterable[str]) -> dict[str, bool]:
         """Delete items related to the trial.
 
         ```python exec="true" source="material-block" result="python" title="delete-storage" hl_lines="6"
@@ -620,39 +564,13 @@ class Trial(RichRenderable, Generic[I]):
 
         Args:
             items: The items to delete, an iterable of keys
-            where: Where the items are stored
-
-                * If `None`, will use the bucket attached to the `Trial` if any,
-                    otherwise it will raise an error.
-
-                * If a `str` or `Path`, will lookup a bucket at the path,
-                and the items will be deleted from a sub-bucket with the name of the trial.
-
[…]
