### 2024-01-28 9893144
refactor(Metric): Apply changes to trial/history/smac
