+++ src/amltk/optimization/history.py
@@ -67,13 +67,13 @@ import pandas as pd
 
 from amltk._functional import compare_accumulate
 from amltk._richutil import RichRenderable
+from amltk.optimization.metric import Metric
 from amltk.optimization.trial import Trial
 from amltk.types import Comparable
 
 if TYPE_CHECKING:
     from rich.console import RenderableType
 
-    from amltk.optimization.metric import Metric
 
 T = TypeVar("T")
 CT = TypeVar("CT", bound=Comparable)
@@ -139,7 +139,7 @@ class History(RichRenderable):
         history.add(reports)
         return history
 
-    def best(self, metric: str | None = None) -> Trial.Report:
+    def best(self, metric: str | Metric | None = None) -> Trial.Report:
         """Returns the best report in the history.
 
         Args:
@@ -150,26 +150,30 @@ class History(RichRenderable):
         Returns:
             The best report.
         """
-        if metric is None:
-            if len(self.metrics) > 1:
-                raise ValueError(
-                    "There are multiple metrics in the history, "
-                    "please specify which metric to sort by.",
-                )
+        match metric:
+            case None:
+                if len(self.metrics) > 1:
+                    raise ValueError(
+                        "There are multiple metrics in the history, "
+                        "please specify which metric to sort by for best.",
+                    )
 
-            _metric_def = next(iter(self.metrics.values()))
-            _metric_name = _metric_def.name
-        else:
-            if metric not in self.metrics:
-                raise ValueError(
-                    f"Metric {metric} not found in history. "
-                    f"Available metrics: {list(self.metrics.keys())}",
-                )
-            _metric_def = self.metrics[metric]
-            _metric_name = metric
+                _metric_def = next(iter(self.metrics.values()))
+                _metric_name = _metric_def.name
+            case str():
+                if metric not in self.metrics:
+                    raise ValueError(
+                        f"Metric {metric} not found in history. "
+                        f"Available metrics: {list(self.metrics.keys())}",
+                    )
+                _metric_def = self.metrics[metric]
+                _metric_name = metric
+            case Metric():
+                _metric_def = metric
+                _metric_name = metric.name
 
         _by = min if _metric_def.minimize else max
-        return _by(self.reports, key=lambda r: r.metrics[_metric_name])
+        return _by(self.reports, key=lambda r: r.values[_metric_name])
 
     def add(self, report: Trial.Report | Iterable[Trial.Report]) -> None:
         """Adds a report or reports to the history.
@@ -179,16 +183,7 @@ class History(RichRenderable):
         """
         match report:
             case Trial.Report():
-                for m in report.metric_values:
-                    if (_m := self.metrics.get(m.name)) is not None:
-                        if m.metric != _m:
-                            raise ValueError(
-                                f"Metric {m.name} has conflicting definitions:"
-                                f"\n{m.metric} != {_m}",
-                            )
-                    else:
-                        self.metrics[m.name] = m.metric
-
+                self.metrics.update(report.metrics)
                 self.reports.append(report)
                 self._lookup[report.name] = len(self.reports) - 1
             case Iterable():
@@ -487,29 +482,28 @@ class History(RichRenderable):
 
         Args:
             key: The key to sort by. If given a str, it will sort by
-                the value of that key in the `.metrics` and also filter
+                the value of that key in the `.values` and also filter
                 out anything that does not contain this key.
             reverse: Whether to sort in some given order. By
                 default (`None`), if given a metric key, the reports with
                 the best metric values will be sorted first. If
                 given a `#!python Callable`, the reports with the
                 smallest values will be sorted first. Using
-                `reverse=True` will always reverse this order, while
-                `reverse=False` will always preserve it.
[…]
