+++ pyproject.toml
@@ -177,6 +177,7 @@ ignore = [
   "PLC1901", # "" can be simplified to be falsey
   "TCH003",  # Move stdlib import into TYPE_CHECKING
   "B010",    # Do not use `setattr`
+  "PD011",   # Use .to_numpy() instead of .values  (triggers on report.values)
   # These tend to be lighweight and confuse pyright
 ]
 
