### 2024-01-28 45da1cf
refactor: Move to redone metric setup
