### 2024-01-29 09d0df2
test(History): `best` shouldn't fail if a trial doesn't contain the metric
