### 2024-01-29 fd9d8cb
fix: Normalized loss calculation
