+++ tests/optimizers/test_metric.py
@@ -17,6 +17,7 @@ class MetricTest:
     expected_loss: float
     expected_distance_from_optimal: float | None
     expected_score: float
+    expected_normalized_loss: float
     expected_str: str
 
 
@@ -28,6 +29,7 @@ def case_metric_score_bounded() -> MetricTest:
         value=0.3,
         expected_loss=-0.3,
         expected_distance_from_optimal=0.7,
+        expected_normalized_loss=0.7,
         expected_score=0.3,
         expected_str="score_bounded [0.0, 1.0] (maximize)",
     )
@@ -41,6 +43,7 @@ def case_metric_score_unbounded() -> MetricTest:
         value=0.3,
         expected_loss=-0.3,
         expected_distance_from_optimal=None,
+        expected_normalized_loss=-0.3,
         expected_score=0.3,
         expected_str="score_unbounded (maximize)",
     )
@@ -54,6 +57,7 @@ def case_metric_loss_unbounded() -> MetricTest:
         value=0.8,
         expected_loss=0.8,
         expected_distance_from_optimal=None,
+        expected_normalized_loss=0.8,
         expected_score=-0.8,
         expected_str="loss_unbounded (minimize)",
     )
@@ -67,6 +71,7 @@ def case_metric_loss_bounded() -> MetricTest:
         value=0.8,
         expected_loss=0.8,
         expected_distance_from_optimal=1.8,
+        expected_normalized_loss=0.9,
         expected_score=-0.8,
         expected_str="loss_bounded [-1.0, 1.0] (minimize)",
     )
@@ -135,10 +140,17 @@ def test_distance_to_optimal_is_always_positive_for_bounded(C: MetricTest) -> No
     assert C.metric.distance_to_optimal(C.value) >= 0
 
 
+@parametrize_with_cases(argnames="C", cases=".")
+def test_normalized_loss(C: MetricTest) -> None:
+    assert C.metric.normalized_loss(C.value) == C.expected_normalized_loss
+
+
 @parametrize_with_cases(argnames="C", cases=".", has_tag=["bounded"])
 def test_normalized_loss_for_bounded(C: MetricTest) -> None:
-    assert 0 <= C.metric.normalized_loss(C.value) >= 1
+    assert 0 <= C.metric.normalized_loss(C.value) <= 1
     assert C.metric.normalized_loss(C.metric.optimal) == 0
+    mid = (C.metric.optimal + C.metric.worst) / 2
+    assert C.metric.normalized_loss(mid) == 0.5
     assert C.metric.normalized_loss(C.metric.worst) == 1
 
 
