### 2024-01-29 0e70e80
refactor(Metric, Trial): Cleanup of metrics and `Trial` (#242)
