+++ src/amltk/optimization/__init__.py
@@ -1,5 +1,5 @@
 from amltk.optimization.history import History
-from amltk.optimization.metric import Metric
+from amltk.optimization.metric import Metric, MetricCollection
 from amltk.optimization.optimizer import Optimizer
 from amltk.optimization.trial import Trial
 
@@ -7,5 +7,6 @@ __all__ = [
     "Optimizer",
     "Trial",
     "Metric",
+    "MetricCollection",
     "History",
 ]
