+++ tests/optimizers/test_optimizers.py
@@ -42,7 +42,10 @@ def target_function(trial: Trial, err: Exception | None = None) -> Trial.Report:
             return trial.fail(err)
 
         return trial.success(
-            **{metric.name: metric.optimal.value for metric in trial.metrics},
+            **{
+                metric_name: metric.optimal
+                for metric_name, metric in trial.metrics.items()
+            },
         )
 
 
@@ -112,7 +115,9 @@ def test_report_success(optimizer: Optimizer) -> None:
     assert report.status == Trial.Status.SUCCESS
     assert valid_time_interval(report.profiles["trial"].time)
     assert report.trial.info is trial.info
-    assert report.metric_values == tuple(metric.optimal for metric in optimizer.metrics)
+    assert report.values == {
+        name: metric.optimal for name, metric in optimizer.metrics.items()
+    }
 
 
 @parametrize_with_cases("optimizer", cases=".", prefix="opt_")
@@ -125,7 +130,7 @@ def test_report_failure(optimizer: Optimizer):
     assert valid_time_interval(report.profiles["trial"].time)
     assert isinstance(report.exception, ValueError)
     assert isinstance(report.traceback, str)
-    assert report.metric_values == tuple(metric.worst for metric in optimizer.metrics)
+    assert report.values == {}
 
 
 @parametrize_with_cases("optimizer", cases=".", prefix="opt_")
