+++ src/amltk/optimization/trial.py
@@ -29,7 +29,13 @@ from __future__ import annotations
 import copy
 import logging
 import traceback as traceback_module
-from collections.abc import Callable, Iterable, Iterator, Mapping, Sequence
+from collections.abc import (
+    Hashable,
+    Iterable,
+    Iterator,
+    Mapping,
+    MutableMapping,
+)
 from contextlib import contextmanager
 from dataclasses import dataclass, field
 from datetime import datetime
@@ -44,9 +50,9 @@ import pandas as pd
 from amltk._functional import dict_get_not_none, mapping_select, prefix_keys
 from amltk._richutil.renderable import RichRenderable
 from amltk._util import parse_timestamp_object
-from amltk.optimization.metric import Metric
+from amltk.optimization.metric import Metric, MetricCollection
 from amltk.profiling import Memory, Profile, Profiler, Timer
-from amltk.store import Bucket, PathBucket
+from amltk.store import PathBucket
 
 if TYPE_CHECKING:
     from rich.console import RenderableType
@@ -107,7 +113,11 @@ class Trial(RichRenderable, Generic[I]):
             return trial.success(cost=cost)
 
         # ... usually obtained from an optimizer
-        trial = Trial(name="some-unique-name", config={"x": 1, "y": 2}, metrics=[cost])
+        trial = Trial.create(
+            name="some-unique-name",
+            config={"x": 1, "y": 2},
+            metrics=[cost]
+        )
 
         report = target_function(trial)
         print(report.df())
@@ -154,7 +164,7 @@ class Trial(RichRenderable, Generic[I]):
             from amltk.optimization import Trial, Metric
 
             # Gotten from some optimizer usually, i.e. via `optimizer.ask()`
-            trial = Trial(
+            trial = Trial.create(
                 name="example_trial",
                 config={"param": 42},
                 metrics=[Metric(name="accuracy", minimize=False)]
@@ -188,7 +198,7 @@ class Trial(RichRenderable, Generic[I]):
         ```python exec="true" source="material-block" result="python" title="profile"
         from amltk.optimization import Trial
 
-        trial = Trial(name="some-unique-name", config={})
+        trial = Trial.create(name="some-unique-name", config={})
 
         # ... somewhere where you've begun your trial.
         with trial.profile("some_interval"):
@@ -225,40 +235,102 @@ class Trial(RichRenderable, Generic[I]):
     config: Mapping[str, Any]
     """The config of the trial provided by the optimizer."""
 
-    bucket: PathBucket = field(
-        default_factory=lambda: PathBucket("unknown-trial-bucket"),
-    )
+    bucket: PathBucket
     """The bucket to store trial related output to."""
 
-    info: I | None = field(default=None, repr=False)
+    info: I | None = field(repr=False)
     """The info of the trial provided by the optimizer."""
 
-    metrics: Sequence[Metric] = field(default_factory=list)
-    """The metrics associated with the trial."""
+    metrics: MetricCollection
+    """The metrics associated with the trial.
+
+    You can access the metrics by name, e.g. `#!python trial.metrics["loss"]`.
+    """
 
     seed: int | None = None
     """The seed to use if suggested by the optimizer."""
 
-    fidelities: dict[str, Any] | None = None
+    fidelities: Mapping[str, Any]
     """The fidelities at which to evaluate the trial, if any."""
 
-    profiler: Profiler = field(
-        repr=False,
-        default_factory=lambda: Profiler(memory_unit="B", time_kind="wall"),
-    )
+    profiler: Profiler = field(repr=False)
     """A profiler for this trial."""
 
-    summary: dict[str, Any] = field(default_factory=dict)
+    summary: MutableMapping[str, Any]
[…]

[…]
