### 2024-01-30 cabd841
feat(sklearn): Provide a standard CVEvaluator
