+++ src/amltk/sklearn/evaluation.py
@@ -1,0 +1,600 @@
+"""TODO upon review."""
+from __future__ import annotations
+
+import logging
+import tempfile
+from collections import defaultdict
+from collections.abc import Callable, Iterable, Iterator, Mapping, Sized
+from functools import partial
+from pathlib import Path
+from typing import TYPE_CHECKING, Any, ClassVar, Literal, TypeAlias, TypeVar
+from typing_extensions import override
+
+import numpy as np
+import pandas as pd
+from sklearn import clone
+from sklearn.base import BaseEstimator
+from sklearn.exceptions import UnsetMetadataPassedError
+from sklearn.metrics import get_scorer
+from sklearn.metrics._scorer import _MultimetricScorer, _Scorer
+from sklearn.model_selection import (
+    KFold,
+    ShuffleSplit,
+    StratifiedKFold,
+    StratifiedShuffleSplit,
+)
+from sklearn.model_selection._validation import _score
+from sklearn.utils.metadata_routing import (
+    MetadataRouter,
+    MethodMapping,
+    process_routing,
+)
+from sklearn.utils.metaestimators import _safe_split
+from sklearn.utils.multiclass import type_of_target
+from sklearn.utils.validation import _check_method_params
+
+import amltk.randomness
+from amltk._functional import subclass_map
+from amltk.optimization.evaluation import EvaluationProtocol
+from amltk.profiling.profiler import Profiler
+from amltk.store import Stored
+from amltk.store.paths.path_bucket import PathBucket
+
+if TYPE_CHECKING:
+    from sklearn.model_selection import (
+        BaseCrossValidator,
+        BaseShuffleSplit,
+    )
+    from sklearn.utils import Bunch
+
+    from amltk.optimization import Trial
+    from amltk.pipeline import Node
+    from amltk.randomness import Seed
+    from amltk.scheduling import Plugin, Scheduler, Task
+
+
+logger = logging.getLogger(__name__)
+
+BaseEstimatorT = TypeVar("BaseEstimatorT", bound=BaseEstimator)
+TaskTypeName: TypeAlias = Literal[
+    "binary",
+    "multiclass",
+    "multilabel-indicator",
+    "continuous",
+    "continuous-multioutput",
+    "multiclass-multioutput",
+]
+
+
+def _route_params(
+    splitter: BaseShuffleSplit | BaseCrossValidator,
+    estimator: BaseEstimator,
+    _scorer: _Scorer | _MultimetricScorer,
+    **params: Any,
+) -> Bunch:
+    # NOTE: This is basically copied out of sklearns 1.4 cross_validate
+
+    # For estimators, a MetadataRouter is created in get_metadata_routing
+    # methods. For these router methods, we create the router to use
+    # `process_routing` on it.
+    router = (
+        MetadataRouter(owner="cross_validate")
+        .add(
+            splitter=splitter,
+            method_mapping=MethodMapping().add(caller="fit", callee="split"),
+        )
+        .add(
+            estimator=estimator,
+            # TODO(SLEP6): also pass metadata to the predict method for
+            # scoring?
+            # ^ Taken from cross_validate source code in sklearn
+            method_mapping=MethodMapping().add(caller="fit", callee="fit"),
+        )
+        .add(
+            scorer=_scorer,
+            method_mapping=MethodMapping().add(caller="fit", callee="score"),
+        )
+    )
+    try:
+        return process_routing(router, "fit", **params)  # type: ignore
+    except UnsetMetadataPassedError as e:
+        # The default exception would mention `fit` since in the above
+        # `process_routing` code, we pass `fit` as the caller. However,
+        # the user is not calling `fit` directly, so we change the message
+        # to make it more suitable for this case.
[…]
