### 2024-01-30 0d3c414
doc: Better class documentation on usage
