### 2024-01-30 876b3fb
doc(sklearn): Raise warnings about conflicting inference
