+++ src/amltk/sklearn/evaluation.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 
 import logging
 import tempfile
+import warnings
 from collections import defaultdict
 from collections.abc import Callable, Iterable, Iterator, Mapping, Sized
 from functools import partial
@@ -115,56 +116,46 @@ def _route_params(
         ) from e
 
 
-def _default_resampler(
+def _default_holdout(
     task_type: TaskTypeName,
+    holdout_size: float,
     *,
-    n_splits: int,
     random_state: Seed | None = None,
-    train_size: float = 0.67,
-) -> BaseShuffleSplit | BaseCrossValidator:
-    if n_splits < 1:
-        raise ValueError("Must have at least one split")
+) -> ShuffleSplit | StratifiedShuffleSplit:
+    if not (0 < holdout_size < 1):
+        raise ValueError(f"`{holdout_size=}` must be in (0, 1)")
 
-    is_holdout = n_splits == 1
-    random_state = amltk.randomness.as_int(random_state)
-
-    # TODO: Custom splitter for edge case where we only have one label
-    # For a given class
-    # if "The least populated class in y has only" in e.args[0]:
-    # Required for both of them here
+    rs = amltk.randomness.as_int(random_state)
     match task_type:
         case "binary" | "multiclass":
-            if is_holdout:
-                return StratifiedShuffleSplit(
-                    n_splits=1,
-                    random_state=random_state,
-                    train_size=train_size,
-                )
-            return StratifiedKFold(
-                n_splits=n_splits,
-                shuffle=True,
-                random_state=random_state,
-            )
-        # NOTE: They don't natively support multilabel-indicator for stratified
+            return StratifiedShuffleSplit(1, random_state=rs, test_size=holdout_size)
         case "multilabel-indicator" | "multiclass-multioutput":
-            if is_holdout:
-                return ShuffleSplit(
-                    n_splits=1,
-                    random_state=random_state,
-                    train_size=train_size,
-                )
-
-            return KFold(n_splits=n_splits, shuffle=True, random_state=random_state)
+            return ShuffleSplit(1, random_state=rs, test_size=holdout_size)
         case "continuous" | "continuous-multioutput":
-            if is_holdout:
-                return ShuffleSplit(
-                    n_splits=1,
-                    random_state=random_state,
-                    train_size=train_size,
-                )
+            return ShuffleSplit(1, random_state=rs, test_size=holdout_size)
+        case _:
+            raise ValueError(f"Don't know how to handle {task_type=}")
 
-            return KFold(n_splits=n_splits, shuffle=True, random_state=random_state)
 
+def _default_cv_resampler(
+    task_type: TaskTypeName,
+    n_splits: int,
+    *,
+    random_state: Seed | None = None,
+) -> StratifiedKFold | KFold:
+    if n_splits < 1:
+        raise ValueError(f"Must have at least one split but got {n_splits=}")
+
+    rs = amltk.randomness.as_int(random_state)
+
+    match task_type:
+        case "binary" | "multiclass":
+            return StratifiedKFold(n_splits=n_splits, shuffle=True, random_state=rs)
+        case "multilabel-indicator" | "multiclass-multioutput":
+            # NOTE: They don't natively support multilabel-indicator for stratified
+            return KFold(n_splits=n_splits, shuffle=True, random_state=rs)
+        case "continuous" | "continuous-multioutput":
+            return KFold(n_splits=n_splits, shuffle=True, random_state=rs)
         case _:
             raise ValueError(f"Don't know how to handle {task_type=} with {n_splits=}")
 
[…]
