### 2024-01-30 7d6ef07
fix: Remove `set_output(transform="pandas")`
