### 2024-01-30 ccfa0c3
fix: Add explicit check for finite floats in scores
