### 2024-01-30 7c8d860
fix: Ensure data is compartmentalized for memory conservation
