### 2024-01-30 3130726
test: Update tests to reflect changes
