+++ tests/sklearn/test_evaluation.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import warnings
 from collections.abc import Mapping
 from dataclasses import dataclass
 from pathlib import Path
@@ -26,7 +27,8 @@ from amltk.pipeline import Component
 from amltk.sklearn.evaluation import (
     CVEvaluation,
     TaskTypeName,
-    _default_resampler,
+    _default_cv_resampler,
+    _default_holdout,
     identify_task_type,
 )
 
@@ -49,6 +51,8 @@ def data_for_task_type(task_type: TaskTypeName) -> tuple[np.ndarray, np.ndarray]
             return make_regression(random_state=42, n_targets=1)  # type: ignore
         case "continuous-multioutput":
             return make_regression(random_state=42, n_targets=2)  # type: ignore
+
+    raise ValueError(f"Unknown task type {task_type}")
 
 
 def _sample_y(task_type: TaskTypeName) -> np.ndarray:
@@ -92,51 +96,57 @@ def test_identify_task_type(
     is_classification: bool,
     expected: str,
 ) -> None:
-    identified = identify_task_type(y, is_classification=is_classification)
+    with warnings.catch_warnings():
+        warnings.filterwarnings("ignore", category=UserWarning)
+        identified = identify_task_type(y, is_classification=is_classification)
+
     assert identified == expected
 
     pd_y = pd.Series(y) if np.ndim(y) == 1 else pd.DataFrame(y)
-    identified = identify_task_type(pd_y, is_classification=is_classification)
+    with warnings.catch_warnings():
+        warnings.filterwarnings("ignore", category=UserWarning)
+        identified = identify_task_type(pd_y, is_classification=is_classification)
+
     assert identified == expected
 
 
 @parametrize(
-    "task_type, n_splits, train_size, expected",
+    "task_type, expected",
     [
-        # CV - Notable, only binary and multiclass can be stratified
-        ("binary", 3, 0.0, StratifiedKFold),
-        ("multiclass", 3, 0.0, StratifiedKFold),
-        ("multilabel-indicator", 3, 0.0, KFold),
-        ("multiclass-multioutput", 3, 0.0, KFold),
-        ("continuous", 3, 0.0, KFold),
-        ("continuous-multioutput", 3, 0.0, KFold),
         # Holdout
-        ("binary", 1, 0.3, StratifiedShuffleSplit),  # With only one fold and test size
-        ("multiclass", 1, 0.3, StratifiedShuffleSplit),
-        ("multilabel-indicator", 1, 0.3, ShuffleSplit),
-        ("multiclass-multioutput", 1, 0.3, ShuffleSplit),
-        ("continuous", 1, 0.3, ShuffleSplit),
-        ("continuous-multioutput", 1, 0.3, ShuffleSplit),
+        ("binary", StratifiedShuffleSplit),
+        ("multiclass", StratifiedShuffleSplit),
+        ("multilabel-indicator", ShuffleSplit),
+        ("multiclass-multioutput", ShuffleSplit),
+        ("continuous", ShuffleSplit),
+        ("continuous-multioutput", ShuffleSplit),
     ],
 )
-def test_default_resampling(
-    task_type: TaskTypeName,
-    n_splits: int,
-    train_size: float,
-    expected: type,
-) -> None:
-    sampler = _default_resampler(
-        task_type,
-        n_splits=n_splits,
-        random_state=42,
-        train_size=train_size,
-    )
+def test_default_holdout(task_type: TaskTypeName, expected: type) -> None:
+    sampler = _default_holdout(task_type, holdout_size=0.387, random_state=42)
     assert isinstance(sampler, expected)
[…]
