+++ tests/sklearn/test_evaluation.py
@@ -3,8 +3,9 @@ from __future__ import annotations
 import warnings
 from collections.abc import Mapping
 from dataclasses import dataclass
+from functools import cache
 from pathlib import Path
-from typing import Any
+from typing import Any, Literal
 
 import numpy as np
 import pandas as pd
@@ -13,8 +14,6 @@ from sklearn.datasets import make_classification, make_regression
 from sklearn.metrics import get_scorer
 from sklearn.metrics._scorer import _Scorer
 from sklearn.model_selection import (
-    BaseCrossValidator,
-    BaseShuffleSplit,
     KFold,
     ShuffleSplit,
     StratifiedKFold,
@@ -23,34 +22,58 @@ from sklearn.model_selection import (
 from sklearn.tree import DecisionTreeClassifier, DecisionTreeRegressor
 
 from amltk.optimization.trial import Metric, Trial
-from amltk.pipeline import Component
+from amltk.pipeline import Component, request
 from amltk.sklearn.evaluation import (
     CVEvaluation,
+    ImplicitMetricConversionWarning,
     TaskTypeName,
+    TaskTypeWarning,
     _default_cv_resampler,
     _default_holdout,
     identify_task_type,
 )
 
 
+# NOTE: We can cache this as it doesn't get changed
+@cache
 def data_for_task_type(task_type: TaskTypeName) -> tuple[np.ndarray, np.ndarray]:
     match task_type:
         case "binary":
-            return make_classification(random_state=42, n_classes=2, n_informative=3)  # type: ignore
+            return make_classification(
+                random_state=42,
+                n_samples=20,
+                n_classes=2,
+                n_informative=3,
+            )  # type: ignore
         case "multiclass":
-            return make_classification(random_state=42, n_classes=4, n_informative=3)  # type: ignore
+            return make_classification(
+                random_state=42,
+                n_samples=20,
+                n_classes=4,
+                n_informative=3,
+            )  # type: ignore
         case "multilabel-indicator":
-            x, y = make_classification(random_state=42, n_classes=2, n_informative=3)
+            x, y = make_classification(
+                random_state=42,
+                n_samples=20,
+                n_classes=2,
+                n_informative=3,
+            )
             y = np.vstack([y, y]).T
             return x, y  # type: ignore
         case "multiclass-multioutput":
-            x, y = make_classification(random_state=42, n_classes=4, n_informative=3)
+            x, y = make_classification(
+                random_state=42,
+                n_samples=20,
+                n_classes=4,
+                n_informative=3,
+            )
             y = np.vstack([y, y, y]).T
             return x, y  # type: ignore
         case "continuous":
-            return make_regression(random_state=42, n_targets=1)  # type: ignore
+            return make_regression(random_state=42, n_samples=20, n_targets=1)  # type: ignore
         case "continuous-multioutput":
-            return make_regression(random_state=42, n_targets=2)  # type: ignore
+            return make_regression(random_state=42, n_samples=20, n_targets=2)  # type: ignore
 
     raise ValueError(f"Unknown task type {task_type}")
 
@@ -60,54 +83,84 @@ def _sample_y(task_type: TaskTypeName) -> np.ndarray:
 
 
 @parametrize(
-    "y, is_classification, expected",
+    "real, task_hint, expected",
     [
-        # with is_classification = None
-        # --- classification
-        (_sample_y("binary"), None, "binary"),
-        (_sample_y("multiclass"), None, "multiclass"),
-        (_sample_y("multilabel-indicator"), None, "multilabel-indicator"),
[…]
