+++ src/amltk/sklearn/evaluation.py
@@ -60,10 +60,43 @@ TaskTypeName: TypeAlias = Literal[
     "binary",
     "multiclass",
     "multilabel-indicator",
+    "multiclass-multioutput",
     "continuous",
     "continuous-multioutput",
-    "multiclass-multioutput",
 ]
+_valid_task_types: tuple[TaskTypeName, ...] = (
+    "binary",
+    "multiclass",
+    "multilabel-indicator",
+    "multiclass-multioutput",
+    "continuous",
+    "continuous-multioutput",
+)
+
+
+class ImplicitMetricConversionWarning(UserWarning):
+    """A warning raised when a metric is implicitly converted to an sklearn scorer.
+
+    This is raised when a metric is provided with a custom function and is
+    implicitly converted to an sklearn scorer. This may fail in some cases
+    and it is recommended to explicitly convert the metric to an sklearn
+    scorer with `make_scorer` and then pass it to the metric with
+    [`Metric(fn=...)`][amltk.optimization.Metric].
+    """
+
+
+class TaskTypeWarning(UserWarning):
+    """A warning raised about the task type."""
+
+
+class NoProvidedIsClassificationWarning(TaskTypeWarning):
+    """A warning raised when the task type is inferred from the target data."""
+
+
+class MismatchedTaskTypeWarning(TaskTypeWarning):
+    """A warning raised when inferred task type with `task_hint` does not
+    match the inferred task type from the target data.
+    """
 
 
 def _route_params(
@@ -162,80 +195,110 @@ def _default_cv_resampler(
 def identify_task_type(  # noqa: PLR0911
     y: np.ndarray | pd.Series | pd.DataFrame,
     *,
-    is_classification: bool | None = None,
+    task_hint: Literal["classification", "regression"] | None = None,
 ) -> TaskTypeName:
     """Identify the task type from the target data."""
-    sklearn_target_type = type_of_target(y)
-    match (sklearn_target_type, is_classification):
-        case (task_type, None):
+    inferred_type: TaskTypeName = type_of_target(y)
+    if task_hint is None:
+        warnings.warn(
+            "`task_hint` was not provided. The task type was inferred from"
+            f" the target data to be '{inferred_type}'."
+            " To silence this warning, please provide `task_hint`.",
+            NoProvidedIsClassificationWarning,
+            stacklevel=2,
+        )
+        return inferred_type
+
+    match task_hint, inferred_type:
+        # First two cases are everything is fine
+        case (
+            "classification",
+            "binary"
+            | "multiclass"
+            | "multilabel-indicator"
+            | "multiclass-multioutput",
+        ):
+            return inferred_type
+        case ("regression", "continuous" | "continuous-multioutput"):
+            return inferred_type
+        # Hinted to be regression but we got a single column classification task
+        case ("regression", "binary" | "multiclass"):
             warnings.warn(
-                "`is_classification` was not provided. The task type was inferred from"
-                f" the target data to be {task_type=}."
-                " To silence this warning, please provide `is_classification`"
-                " as either `True` or `False`.",
-                UserWarning,
-                stacklevel=2,
-            )
-            return sklearn_target_type
-        case ("continuous", True) if len(np.unique(y)) == 2:  # noqa: PLR2004
-            warnings.warn(
-                f"`{is_classification=}` but the target"
-                f" data was inferred to be `{sklearn_target_type}` originally."
-                " We have corrected this to be `binary` as there were only"
-                " 2 unique values. This may cause crashes. To silence this"
-                " warning, you can encode your target data as `binary` by"
-                " making it a categorical or by using a `LabelEncoder`.",
-                UserWarning,
-                stacklevel=2,
-            )
-            return "binary"
-        case ("continuous", True):
[…]
