### 2024-01-30 dac9419
test: Give tmp_path fixture as trial bucket
