### 2024-01-30 9b7e488
fix: Delay loading of data until last moment
