### 2024-01-30 de0ede5
doc: Make sure to remove empty dirs created by trials
