### 2024-01-30 92942dc
doc: Add module docstring for `CVEvaluator`
