### 2024-01-30 24b3dbc
doc: Fix continuation of doc exec session
