### 2024-01-30 e8f6947
fix: Use `mkdtemp` instead of `TemporaryDirectory`
`TemporaryDirectory` will always cleanup when it goes out of scope
and the `delete: bool` parameter is only in python 3.12 upwards
