### 2024-01-30 1b16f8f
doc: Remove n_workers from example
