### 2024-01-30 5ccad97
refactor: Unify `datadir` with `optimizes` `working_dir`
No need to introduce a name
