### 2024-01-30 d6aa8a8
refactor: Unify `datadir` with `optimizes` `working_dir`
No need to introduce a name
