### 2024-01-30 e4073e0
feat: Add an automatic parameter warning type
