### 2024-01-30 cf8b9a3
refactor: Move warnings to seperate file, use automatic warning type
