### 2024-01-30 3bff968
refactor: Use `"auto"` for indicating auto parameters
