### 2024-01-30 2d59f12
doc: Ignore automatic parameter warnings
