### 2024-01-31 #6133 @JBorrow
Hi @GarrettMerz, I found your handle using the JOSS review volunteer application. Would you be interested in reviewing this package for JOSS? Our review guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html should you want to take a look before agreeing to review.
