### 2024-01-31 03698a3
feat: Test metadata routing
