+++ tests/sklearn/test_evaluation.py
@@ -9,11 +9,15 @@ from typing import Any, Literal
 
 import numpy as np
 import pandas as pd
+import pytest
 from pytest_cases import case, parametrize, parametrize_with_cases
+from sklearn import config_context as sklearn_config_context
 from sklearn.datasets import make_classification, make_regression
-from sklearn.metrics import get_scorer
+from sklearn.dummy import DummyClassifier
+from sklearn.metrics import get_scorer, make_scorer
 from sklearn.metrics._scorer import _Scorer
 from sklearn.model_selection import (
+    GroupKFold,
     KFold,
     ShuffleSplit,
     StratifiedKFold,
@@ -317,14 +321,14 @@ def case_regression(
 @parametrize("store_models", [True, False])
 @parametrize("train_score", [True, False])
 @parametrize_with_cases("item", cases=".", prefix="case_")
-@parametrize("cv_value, strategy", [(2, "cv"), (0.3, "holdout")])
+@parametrize("cv_value, splitter", [(2, "cv"), (0.3, "holdout")])
 def test_evaluator(
     as_pd: bool,
     store_models: bool,
     train_score: bool,
     item: _EvalKwargs,
     cv_value: int | float,
-    strategy: str,
+    splitter: str,
 ) -> None:
     x = pd.DataFrame(item.X) if as_pd else item.X
     y = (
@@ -333,10 +337,10 @@ def test_evaluator(
         else (pd.DataFrame(item.y) if np.ndim(item.y) > 1 else pd.Series(item.y))
     )
     trial = item.trial
-    if strategy == "cv":
-        cv_kwargs = {"n_splits": cv_value, "strategy": "cv"}
+    if splitter == "cv":
+        cv_kwargs = {"n_splits": cv_value, "splitter": "cv"}
     else:
-        cv_kwargs = {"holdout_size": cv_value, "strategy": "holdout"}
+        cv_kwargs = {"holdout_size": cv_value, "splitter": "holdout"}
 
     evaluator = CVEvaluation(
         X=x,
@@ -406,11 +410,11 @@ def test_evaluator(
         "continuous-multioutput",
     ],
 )
-@parametrize("cv_value, strategy", [(2, "cv"), (0.3, "holdout")])
+@parametrize("cv_value, splitter", [(2, "cv"), (0.3, "holdout")])
 def test_consistent_results_across_seeds(
     tmp_path: Path,
     cv_value: int | float,
-    strategy: Literal["cv", "holdout"],
+    splitter: Literal["cv", "holdout"],
     task_type: TaskTypeName,
 ) -> None:
     x, y = data_for_task_type(task_type)
@@ -440,10 +444,10 @@ def test_consistent_results_across_seeds(
                 fn=lambda y_true, y_pred: (y_pred == y_true).mean().mean(),
             )
 
-    if strategy == "cv":
-        cv_kwargs = {"n_splits": cv_value, "strategy": "cv"}
+    if splitter == "cv":
+        cv_kwargs = {"n_splits": cv_value, "splitter": "cv"}
     else:
-        cv_kwargs = {"holdout_size": cv_value, "strategy": "holdout"}
+        cv_kwargs = {"holdout_size": cv_value, "splitter": "holdout"}
 
     evaluator_1 = CVEvaluation(
         X=x,
@@ -495,3 +499,152 @@ def test_consistent_results_across_seeds(
     df_1 = report_1.df(profiles=False).drop(columns=["reported_at"])
     df_2 = report_2.df(profiles=False).drop(columns=["reported_at"])
     pd.testing.assert_frame_equal(df_1, df_2)
+
+
+def test_scoring_params_get_forwarded(tmp_path: Path) -> None:
+    with sklearn_config_context(enable_metadata_routing=True):
[…]
