### 2024-02-01 ad9fb4e
test: Check evaluator works with clustering methods
