### 2024-02-01 7410272
feat(sklearn): Provide a standard CVEvaluator (#244)
