+++ tests/sklearn/test_evaluation.py
@@ -1,0 +1,682 @@
+from __future__ import annotations
+
+import warnings
+from collections.abc import Mapping
+from dataclasses import dataclass
+from functools import cache
+from pathlib import Path
+from typing import Any, Literal
+
+import numpy as np
+import pandas as pd
+import pytest
+import sklearn.datasets
+from pytest_cases import case, parametrize, parametrize_with_cases
+from sklearn import config_context as sklearn_config_context
+from sklearn.cluster import KMeans
+from sklearn.datasets import make_classification, make_regression
+from sklearn.dummy import DummyClassifier
+from sklearn.metrics import get_scorer, make_scorer
+from sklearn.metrics._scorer import _Scorer
+from sklearn.model_selection import (
+    GroupKFold,
+    KFold,
+    ShuffleSplit,
+    StratifiedKFold,
+    StratifiedShuffleSplit,
+)
+from sklearn.tree import DecisionTreeClassifier, DecisionTreeRegressor
+
+from amltk.exceptions import TaskTypeWarning
+from amltk.optimization.trial import Metric, Trial
+from amltk.pipeline import Component, request
+from amltk.sklearn.evaluation import (
+    CVEvaluation,
+    ImplicitMetricConversionWarning,
+    TaskTypeName,
+    _default_cv_resampler,
+    _default_holdout,
+    identify_task_type,
+)
+
+
+# NOTE: We can cache this as it doesn't get changed
+@cache
+def data_for_task_type(task_type: TaskTypeName) -> tuple[np.ndarray, np.ndarray]:
+    match task_type:
+        case "binary":
+            return make_classification(
+                random_state=42,
+                n_samples=20,
+                n_classes=2,
+                n_informative=3,
+            )  # type: ignore
+        case "multiclass":
+            return make_classification(
+                random_state=42,
+                n_samples=20,
+                n_classes=4,
+                n_informative=3,
+            )  # type: ignore
+        case "multilabel-indicator":
+            x, y = make_classification(
+                random_state=42,
+                n_samples=20,
+                n_classes=2,
+                n_informative=3,
+            )
+            y = np.vstack([y, y]).T
+            return x, y  # type: ignore
+        case "multiclass-multioutput":
+            x, y = make_classification(
+                random_state=42,
+                n_samples=20,
+                n_classes=4,
+                n_informative=3,
+            )
+            y = np.vstack([y, y, y]).T
+            return x, y  # type: ignore
+        case "continuous":
+            return make_regression(random_state=42, n_samples=20, n_targets=1)  # type: ignore
+        case "continuous-multioutput":
+            return make_regression(random_state=42, n_samples=20, n_targets=2)  # type: ignore
+
+    raise ValueError(f"Unknown task type {task_type}")
+
+
+def _sample_y(task_type: TaskTypeName) -> np.ndarray:
+    return data_for_task_type(task_type)[1]
+
+
+@parametrize(
+    "real, task_hint, expected",
+    [
+        ("binary", "auto", "binary"),
+        ("binary", "classification", "binary"),
+        ("binary", "regression", "continuous"),
+        #
+        ("multiclass", "auto", "multiclass"),
+        ("multiclass", "classification", "multiclass"),
+        ("multiclass", "regression", "continuous"),
+        #
+        ("multilabel-indicator", "auto", "multilabel-indicator"),
+        ("multilabel-indicator", "classification", "multilabel-indicator"),
+        ("multilabel-indicator", "regression", "continuous-multioutput"),
+        #
+        ("multiclass-multioutput", "auto", "multiclass-multioutput"),
[…]
