### 2024-02-02 07ee9a3
fix(Trial): Give trials a created_at stamp
When plotting, it's important to have some notion
of a start. In this case, it's usually as soon as the first
Trial is created. Otherwise, it would be up for the user
to get a timestamp such as `datetime.now()`
