### 2024-02-02 518c2b4
test: Fix test which expected created_at to be the same
