### 2024-02-02 baa0f82
fix(CVEvaluator): When on_error="raise", inform of which trial failed
Previously it would give the error but it was not really clear which
trial failed. No it's explicitly in the error.
