+++ src/amltk/sklearn/evaluation.py
@@ -49,6 +49,7 @@ from amltk.exceptions import (
     AutomaticTaskTypeInferredWarning,
     ImplicitMetricConversionWarning,
     MismatchedTaskTypeWarning,
+    TrialError,
 )
 from amltk.optimization.evaluation import EvaluationProtocol
 from amltk.profiling.profiler import Profiler
@@ -564,7 +565,7 @@ def cross_validate_task(  # noqa: D103, PLR0913, C901, PLR0912
     except Exception as e:  # noqa: BLE001
         trial.dump_exception(e)
         if on_error == "raise":
-            raise e
+            raise TrialError(f"Trial failed: {trial}") from e
         return trial.fail(e)
     else:
         mean_val_scores = {k: np.mean(v) for k, v in all_val_scores.items()}
