### 2024-02-02 104e087
test: Fix raises assertion to use new error type
