### 2024-02-02 1f6cb69
fix(CVEvaluator): When on_error="raise", inform of which trial failed (#247)
