### 2024-02-02 1271071
fix(Node): Ensure that parent name does not conflict with children
This was causing an issue when configuring, where essentially the
parent would steal the configuration of its child.
