### 2024-02-02 1a07dc2
fix(CVEvaluator): `clone` the estimator before use
whoops
