### 2024-02-02 2553f6a
feat(sklearn): CVEvaluator params for `configure` and `build`
