### 2024-02-02 9e6b59a
feat(sklearn): CVEvaluator now fully parametrized
