+++ src/amltk/sklearn/evaluation.py
@@ -14,7 +14,7 @@ import logging
 import tempfile
 import warnings
 from collections import defaultdict
-from collections.abc import Callable, Iterable, Iterator, Mapping, MutableMapping, Sized
+from collections.abc import Iterable, Iterator, Mapping, MutableMapping, Sized
 from datetime import datetime
 from functools import partial
 from pathlib import Path
@@ -468,7 +468,7 @@ def _iter_cross_validate(
         )
 
 
-def cross_validate_task(  # noqa: D103, PLR0913, C901, PLR0912
+def cross_validate_task(  # noqa: D103, C901, PLR0912, PLR0915
     trial: Trial,
     pipeline: Node,
     *,
@@ -479,7 +479,6 @@ def cross_validate_task(  # noqa: D103, PLR0913, C901, PLR0912
     train_score: bool = False,
     store_models: bool = True,
     params: MutableMapping[str, Stored[Any] | Any] | None = None,
-    builder: Literal["sklearn"] | Callable[[Node], BaseEstimator] = "sklearn",
     on_error: Literal["fail", "raise"] = "fail",
 ) -> Trial.Report:
     params = {} if params is None else params
@@ -491,21 +490,29 @@ def cross_validate_task(  # noqa: D103, PLR0913, C901, PLR0912
             f"Expected `params['configure']` to be a dict but got {configure_params=}",
         )
 
-    build_params = params.pop("build", {})
+    if "random_state" in configure_params:
+        raise ValueError(
+            "You should not provide `'random_state'` in `params['configure']`"
+            " as the seed is set by the optimizer.",
+        )
+    random_state = amltk.randomness.as_randomstate(trial.seed)
+    configure_params["random_state"] = random_state
+
+    build_params = params.pop("build", {"builder": "sklearn"})  # type: ignore
     if not isinstance(build_params, MutableMapping):
         raise ValueError(
             f"Expected `params['build']` to be a dict but got {build_params=}",
         )
 
-    builder = params.pop("builder", "sklearn")  # type: ignore
+    transform_context = params.pop("transform_context", None)  # type: ignore
 
-    random_state = amltk.randomness.as_randomstate(trial.seed)
-
-    # TODO: Could possibly include `transform_context` here to `configure()`
     estimator = pipeline.configure(
         trial.config,
-        params={**configure_params, "random_state": random_state},
-    ).build(builder, **build_params)
+        transform_context=transform_context,
+        params=configure_params,
+    ).build(
+        **build_params,
+    )
 
     scorers: dict[str, _Scorer] = {}
     for metric_name, metric in trial.metrics.items():
@@ -654,6 +661,52 @@ class CVEvaluation(EvaluationProtocol):
     print(history.df())
     evaluator.bucket.rmdir()  # Cleanup
     ```
+
+    If you need to pass dataset specific items to the pipeline, you can do so
+    use the [`request`][amltk.pipeline.request] in the config of your pipeline.
+
+    ```python exec="true" source="material-block" result="python"
+    from amltk.sklearn import CVEvaluation
+    from amltk.pipeline import Component, request
+    from amltk.optimization import Metric
+
+    from sklearn.ensemble import RandomForestClassifier
+    from sklearn.metrics import get_scorer
+    from sklearn.datasets import load_iris
+    from pathlib import Path
+
+    working_dir = Path("./some-path")
+    X, y = load_iris(return_X_y=True)
+
+    pipeline = Component(
+        RandomForestClassifier,
+        config={
+            "random_state": request("random_state"),
+            # Allow it to be configured with n_jobs
+            "n_jobs": request("n_jobs", default=None)
+        },
+        space={"n_estimators": (10, 100), "criterion": ["gini", "entropy"]},
+    )
+
+    evaluator = CVEvaluation(
[…]
