### 2024-02-02 40d33c2
test: Remove unnecesary test
