### 2024-02-02 0588391
doc: Fixup docstring
