### 2024-02-03 82b6988
chore: Update ruff and pre-commit
