+++ pyproject.toml
@@ -107,8 +107,14 @@ changelog_start_rev = "1.0.0"
 [tool.ruff]
 target-version = "py310"
 line-length = 88
-show-source = true
+output-format = "full"
 src = ["src", "tests", "examples"]
+
+[tool.ruff.lint]
+# Extend what ruff is allowed to fix, even it it may break
+# This is okay given we use it all the time and it ensures
+# better practices. Would be dangerous if using for first
+# time on established project.
 extend-safe-fixes = ["ALL"]
 
 # Allow unused variables when underscore-prefixed.
@@ -205,7 +211,7 @@ exclude = [
 ]
 
 # Exclude a variety of commonly ignored directories.
-[tool.ruff.per-file-ignores]
+[tool.ruff.lint.per-file-ignores]
 "tests/*.py" = [
   "S101",
   "D103",
@@ -225,7 +231,7 @@ exclude = [
 "docs/*" = ["INP001"]
 
 
-[tool.ruff.isort]
+[tool.ruff.lint.isort]
 known-first-party = ["amltk"]
 known-third-party = ["sklearn"]
 no-lines-before = ["future"]
@@ -234,10 +240,10 @@ combine-as-imports = true
 extra-standard-library = ["typing_extensions"]
 force-wrap-aliases = true
 
-[tool.ruff.pydocstyle]
+[tool.ruff.lint.pydocstyle]
 convention = "google"
 
-[tool.ruff.pylint]
+[tool.ruff.lint.pylint]
 max-args = 10 # Changed from default of 5
 
 [tool.mypy]
