### 2024-02-03 9927dd2
feat(sklearn): CVEvaluator allows `configure` and `build` params (#250)
