+++ src/amltk/sklearn/evaluation.py
@@ -14,7 +14,7 @@ import logging
 import tempfile
 import warnings
 from collections import defaultdict
-from collections.abc import Callable, Iterable, Iterator, Mapping, Sized
+from collections.abc import Iterable, Iterator, Mapping, MutableMapping, Sized
 from datetime import datetime
 from functools import partial
 from pathlib import Path
@@ -468,7 +468,7 @@ def _iter_cross_validate(
         )
 
 
-def cross_validate_task(  # noqa: D103, PLR0913, C901, PLR0912
+def cross_validate_task(  # noqa: D103, C901, PLR0912, PLR0915
     trial: Trial,
     pipeline: Node,
     *,
@@ -478,22 +478,41 @@ def cross_validate_task(  # noqa: D103, PLR0913, C901, PLR0912
     additional_scorers: Mapping[str, _Scorer] | None,
     train_score: bool = False,
     store_models: bool = True,
-    params: Mapping[str, Stored[Any] | Any] | None = None,
-    builder: Literal["sklearn"] | Callable[[Node], BaseEstimator] = "sklearn",
-    build_params: Mapping[str, Any] | None = None,
+    params: MutableMapping[str, Stored[Any] | Any] | None = None,
     on_error: Literal["fail", "raise"] = "fail",
 ) -> Trial.Report:
     params = {} if params is None else params
     # Make sure to load all the stored values
 
-    build_params = {} if build_params is None else build_params
-    random_state = amltk.randomness.as_randomstate(trial.seed)
+    configure_params = params.pop("configure", {})
+    if not isinstance(configure_params, MutableMapping):
+        raise ValueError(
+            f"Expected `params['configure']` to be a dict but got {configure_params=}",
+        )
 
-    # TODO: Could possibly include `transform_context` here to `configure()`
+    if "random_state" in configure_params:
+        raise ValueError(
+            "You should not provide `'random_state'` in `params['configure']`"
+            " as the seed is set by the optimizer.",
+        )
+    random_state = amltk.randomness.as_randomstate(trial.seed)
+    configure_params["random_state"] = random_state
+
+    build_params = params.pop("build", {"builder": "sklearn"})  # type: ignore
+    if not isinstance(build_params, MutableMapping):
+        raise ValueError(
+            f"Expected `params['build']` to be a dict but got {build_params=}",
+        )
+
+    transform_context = params.pop("transform_context", None)  # type: ignore
+
     estimator = pipeline.configure(
         trial.config,
-        params={"random_state": random_state},
-    ).build(builder, **build_params)
+        transform_context=transform_context,
+        params=configure_params,
+    ).build(
+        **build_params,
+    )
 
     scorers: dict[str, _Scorer] = {}
     for metric_name, metric in trial.metrics.items():
@@ -642,6 +661,54 @@ class CVEvaluation(EvaluationProtocol):
     print(history.df())
     evaluator.bucket.rmdir()  # Cleanup
     ```
+
+    If you need to pass specific configuration items to your pipeline during
+    configuration, you can do so using a [`request()`][amltk.pipeline.request]
+    in the config of your pipeline.
+
+    In the below example, we allow the pipeline to be configured with `"n_jobs"`
+    and pass it in to the `CVEvalautor` using the `params` argument.
+
+    ```python exec="true" source="material-block" result="python"
+    from amltk.sklearn import CVEvaluation
+    from amltk.pipeline import Component, request
+    from amltk.optimization import Metric
+
+    from sklearn.ensemble import RandomForestClassifier
+    from sklearn.metrics import get_scorer
+    from sklearn.datasets import load_iris
+    from pathlib import Path
+
+    working_dir = Path("./some-path")
[…]
