### 2024-02-03 e858f59
chore: Update ruff and pre-commit (#251)
