### 2024-02-04 00476c8
refactor(pipeline): `optimize` now requires one of `timeout` or `max_trials` set
