### 2024-02-04 e677212
doc: Fix docstring with specified parameter
