### 2024-02-08 43627e7
fix(metalearning): Portfolio Check for Dataframe as Input (#253)
