### 2024-02-08 #6133 @JBorrow
Hi @woznicak, you were suggested as a potential reviewer for this paper and repository by one of the authors. Would you be interested in reviewing this package for the Journal of Open Source Software? We are a diamond open access journal focused on sharing and improving software. Our review guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html should you want to take a look before agreeing to review.
