+++ src/amltk/sklearn/evaluation.py
@@ -13,16 +13,29 @@ from __future__ import annotations
 import logging
 import tempfile
 import warnings
+from asyncio import Future
 from collections import defaultdict
-from collections.abc import Iterable, Iterator, Mapping, MutableMapping, Sized
+from collections.abc import Callable, Iterator, Mapping, MutableMapping, Sized
+from contextlib import nullcontext
 from datetime import datetime
 from functools import partial
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, ClassVar, Literal, TypeAlias, TypeVar
+from typing import (
+    TYPE_CHECKING,
+    Any,
+    ClassVar,
+    Literal,
+    NamedTuple,
+    ParamSpec,
+    Protocol,
+    TypeAlias,
+    TypeVar,
+)
 from typing_extensions import override
 
 import numpy as np
 import pandas as pd
+from more_itertools import all_equal
 from sklearn.base import BaseEstimator, clone
 from sklearn.exceptions import UnsetMetadataPassedError
 from sklearn.metrics import get_scorer
@@ -47,12 +60,16 @@ import amltk.randomness
 from amltk._functional import subclass_map
 from amltk.exceptions import (
     AutomaticTaskTypeInferredWarning,
+    CVEarlyStoppedError,
     ImplicitMetricConversionWarning,
     MismatchedTaskTypeWarning,
     TrialError,
 )
 from amltk.optimization.evaluation import EvaluationProtocol
 from amltk.profiling.profiler import Profiler
+from amltk.scheduling import Plugin, Task
+from amltk.scheduling.events import Emitter, Event
+from amltk.scheduling.plugins.comm import Comm
 from amltk.store import Stored
 from amltk.store.paths.path_bucket import PathBucket
 
@@ -66,7 +83,17 @@ if TYPE_CHECKING:
     from amltk.optimization import Trial
     from amltk.pipeline import Node
     from amltk.randomness import Seed
-    from amltk.scheduling import Plugin, Scheduler, Task
+
+P = ParamSpec("P")
+R = TypeVar("R")
+
+
+class _CVEarlyStopper(Protocol):
+    def update(self, report: Trial.Report) -> None:
+        ...
+
+    def should_stop(self, info: CVEvaluation.FoldInfo) -> bool | Exception:
+        ...
 
 
 logger = logging.getLogger(__name__)
@@ -302,7 +329,7 @@ def _fit(
     profiler: Profiler,
     fit_params: Mapping[str, Any],
     scorer_params: Mapping[str, Any],
-    scorers: Mapping[str, _Scorer] | None = None,
+    scorers: _MultimetricScorer | None = None,
 ) -> tuple[BaseEstimatorT, Mapping[str, float] | None]:
     _fit_params = _check_method_params(X, params=fit_params, indices=i_train)
     _scorer_params_train = _check_method_params(X, scorer_params, indices=i_train)
@@ -348,10 +375,10 @@ def _score_val_fold(
     i_train: np.ndarray,
     i_val: np.ndarray,
     scorer_params: Mapping[str, Any],
-    scorers: Mapping[str, _Scorer],
+    scorers: _MultimetricScorer,
     profiler: Profiler,
 ) -> Mapping[str, float]:
-    _scorer_params_test = _check_method_params(X, scorer_params, indices=i_val)
+    _scorer_params_test = _check_method_params(X, params=scorer_params, indices=i_val)
     X_t, y_t = _safe_split(estimator, X, y, indices=i_val, train_indices=i_train)
 
     with profiler("score"):
@@ -389,7 +416,7 @@ def _evaluate_fold(
     i_train: np.ndarray,
     i_val: np.ndarray,
[…]
