+++ src/amltk/scheduling/task.py
@@ -16,23 +16,23 @@ with callbacks and act accordingly.
     for more on how to customize these callbacks. You can also take a look
     at the API of [`on()`][amltk.scheduling.task.Task.on] for more information.
 
-    === "`@on_result`"
+    === "`@on-result`"
 
         ::: amltk.scheduling.task.Task.on_result
 
-    === "`@on_exception`"
+    === "`@on-exception`"
 
         ::: amltk.scheduling.task.Task.on_exception
 
-    === "`@on_done`"
+    === "`@on-done`"
 
         ::: amltk.scheduling.task.Task.on_done
 
-    === "`@on_submitted`"
+    === "`@on-submitted`"
 
         ::: amltk.scheduling.task.Task.on_submitted
 
-    === "`@on_cancelled`"
+    === "`@on-cancelled`"
 
         ::: amltk.scheduling.task.Task.on_cancelled
 
@@ -75,15 +75,14 @@ from __future__ import annotations
 import logging
 from asyncio import Future
 from collections.abc import Callable, Iterable
-from typing import TYPE_CHECKING, Any, Concatenate, Generic, TypeVar, overload
+from typing import TYPE_CHECKING, Any, Concatenate, Generic, TypeVar
 from typing_extensions import ParamSpec, Self, override
 
-from more_itertools import first_true
+from dask.utils import funcname
 
 from amltk._functional import callstring
 from amltk._richutil.renderable import RichRenderable
-from amltk.exceptions import EventNotKnownError, SchedulerNotRunningError
-from amltk.randomness import randuid
+from amltk.exceptions import SchedulerNotRunningError
 from amltk.scheduling.events import Emitter, Event, Subscriber
 from amltk.scheduling.plugins.plugin import Plugin
 
@@ -103,7 +102,7 @@ R2 = TypeVar("R2")
 CallableT = TypeVar("CallableT", bound=Callable)
 
 
-class Task(RichRenderable, Generic[P, R]):
+class Task(Emitter, RichRenderable, Generic[P, R]):
     """The task class."""
 
     unique_ref: str
@@ -121,7 +120,7 @@ class Task(RichRenderable, Generic[P, R]):
     emitter: Emitter
     """The emitter for events of this task."""
 
-    on_submitted: Subscriber[Concatenate[Future[R], P]]
+    on_submitted: Subscriber[Concatenate[Future[R], P], Any]
     """An event that is emitted when a future is submitted to the
     scheduler. It will pass the future as the first argument with args and
     kwargs following.
@@ -133,7 +132,7 @@ class Task(RichRenderable, Generic[P, R]):
         print(f"Future {future} was submitted with {args=} and {kwargs=}")
     ```
     """
-    on_done: Subscriber[Future[R]]
+    on_done: Subscriber[[Future[R]], Any]
     """Called when a task is done running with a result or exception.
     ```python
     @task.on_done
@@ -141,7 +140,7 @@ class Task(RichRenderable, Generic[P, R]):
         print(f"Future {future} is done")
     ```
     """
-    on_cancelled: Subscriber[Future[R]]
+    on_cancelled: Subscriber[[Future[R]], Any]
     """Called when a task is cancelled.
     ```python
     @task.on_cancelled
@@ -149,7 +148,7 @@ class Task(RichRenderable, Generic[P, R]):
         print(f"Future {future} was cancelled")
     ```
     """
-    on_result: Subscriber[Future[R], R]
+    on_result: Subscriber[[Future[R], R], Any]
     """Called when a task has successfully returned a value.
     Comes with Future
     ```python
@@ -158,7 +157,7 @@ class Task(RichRenderable, Generic[P, R]):
         print(f"Future {future} returned {result}")
     ```
     """
[…]
