### 2024-02-09 16a2bae
fix: Off by one error with `max_trials`
