### 2024-02-09 08b9a6f
fix: Work with sklearn 1.4 metadata routing disabled
