+++ src/amltk/sklearn/evaluation.py
@@ -47,6 +47,8 @@ from sklearn.model_selection import (
     StratifiedShuffleSplit,
 )
 from sklearn.model_selection._validation import _score
+from sklearn.utils import Bunch
+from sklearn.utils._metadata_requests import _routing_enabled
 from sklearn.utils.metadata_routing import (
     MetadataRouter,
     MethodMapping,
@@ -78,7 +80,6 @@ if TYPE_CHECKING:
         BaseCrossValidator,
         BaseShuffleSplit,
     )
-    from sklearn.utils import Bunch
 
     from amltk.optimization import Trial
     from amltk.pipeline import Node
@@ -123,47 +124,54 @@ def _route_params(
     _scorer: _Scorer | _MultimetricScorer,
     **params: Any,
 ) -> Bunch:
-    # NOTE: This is basically copied out of sklearns 1.4 cross_validate
+    if _routing_enabled():
+        # NOTE: This is basically copied out of sklearns 1.4 cross_validate
 
-    # For estimators, a MetadataRouter is created in get_metadata_routing
-    # methods. For these router methods, we create the router to use
-    # `process_routing` on it.
-    router = (
-        MetadataRouter(owner="cross_validate")
-        .add(
-            splitter=splitter,
-            method_mapping=MethodMapping().add(caller="fit", callee="split"),
+        # For estimators, a MetadataRouter is created in get_metadata_routing
+        # methods. For these router methods, we create the router to use
+        # `process_routing` on it.
+        router = (
+            MetadataRouter(owner="cross_validate")
+            .add(
+                splitter=splitter,
+                method_mapping=MethodMapping().add(caller="fit", callee="split"),
+            )
+            .add(
+                estimator=estimator,
+                # TODO(SLEP6): also pass metadata to the predict method for
+                # scoring?
+                # ^ Taken from cross_validate source code in sklearn
+                method_mapping=MethodMapping().add(caller="fit", callee="fit"),
+            )
+            .add(
+                scorer=_scorer,
+                method_mapping=MethodMapping().add(caller="fit", callee="score"),
+            )
         )
-        .add(
-            estimator=estimator,
-            # TODO(SLEP6): also pass metadata to the predict method for
-            # scoring?
-            # ^ Taken from cross_validate source code in sklearn
-            method_mapping=MethodMapping().add(caller="fit", callee="fit"),
-        )
-        .add(
-            scorer=_scorer,
-            method_mapping=MethodMapping().add(caller="fit", callee="score"),
-        )
-    )
-    try:
-        return process_routing(router, "fit", **params)  # type: ignore
-    except UnsetMetadataPassedError as e:
-        # The default exception would mention `fit` since in the above
-        # `process_routing` code, we pass `fit` as the caller. However,
-        # the user is not calling `fit` directly, so we change the message
-        # to make it more suitable for this case.
-        raise UnsetMetadataPassedError(
-            message=(
-                f"{sorted(e.unrequested_params.keys())} are passed to cross"
-                " validation but are not explicitly requested or unrequested. See"
-                " the Metadata Routing User guide"
-                " <https://scikit-learn.org/stable/metadata_routing.html> for more"
-                " information."
-            ),
-            unrequested_params=e.unrequested_params,
-            routed_params=e.routed_params,
-        ) from e
+        try:
+            return process_routing(router, "fit", **params)  # type: ignore
+        except UnsetMetadataPassedError as e:
+            # The default exception would mention `fit` since in the above
+            # `process_routing` code, we pass `fit` as the caller. However,
+            # the user is not calling `fit` directly, so we change the message
[…]
