### 2024-02-09 97c2419
doc: Remove stale comment
