### 2024-02-09 2e57243
fix: Match with isinstance for np.bool_
