### 2024-02-09 39cee5a
fix: Don't fail is `np.bool_` object is falsy
