+++ src/amltk/sklearn/evaluation.py
@@ -651,8 +651,9 @@ def cross_validate_task(  # noqa: D103, C901, PLR0915, PLR0913
                             raise CVEarlyStoppedError("Early stopped!")
                         case False:
                             pass
-                        case np.bool_() if bool(response) is True:
-                            raise CVEarlyStoppedError("Early stopped!")
+                        case np.bool_():
+                            if bool(response) is True:
+                                raise CVEarlyStoppedError("Early stopped!")
                         case Exception():
                             raise response
                         case _:
