### 2024-02-09 fcd80dc
refactor: Remove concept of `Evaluator`
