### 2024-02-09 114e93f
doc: Add docstring to plugin to elaborate more
