### 2024-02-09 cb12a18
test: Remove `xfails`
