+++ src/amltk/optimization/trial.py
@@ -343,7 +343,22 @@ class Trial(RichRenderable, Generic[I]):
 
     @property
     def profiles(self) -> Mapping[str, Profile.Interval]:
-        """The profiles of the trial."""
+        """The profiles of the trial.
+
+        These are indexed by the name of the profile indicated by:
+
+        ```python
+        with trial.profile("key_to_index"):
+            # ...
+
+        profile = trial.profiles["key_to_index"]
+        ```
+
+        The values are a [`Profile.Interval`][amltk.profiling.profile.Interval],
+        which contain a [`Memory.Interval`][amltk.profiling.memory.Interval] and
+        a [`Timer.Interval`][amltk.profiling.timer.Interval]. Please see the
+        respective documentation for more.
+        """
         return self.profiler.profiles
 
     def dump_exception(
