+++ src/amltk/sklearn/evaluation.py
@@ -88,17 +88,10 @@ P = ParamSpec("P")
 R = TypeVar("R")
 
 
-class _CVEarlyStopper(Protocol):
-    def update(self, report: Trial.Report) -> None:
-        ...
-
-    def should_stop(self, info: CVEvaluation.SplitInfo) -> bool | Exception:
-        ...
-
-
 logger = logging.getLogger(__name__)
 
 BaseEstimatorT = TypeVar("BaseEstimatorT", bound=BaseEstimator)
+
 TaskTypeName: TypeAlias = Literal[
     "binary",
     "multiclass",
@@ -107,6 +100,8 @@ TaskTypeName: TypeAlias = Literal[
     "continuous",
     "continuous-multioutput",
 ]
+"""A type alias for the task type name as defined by sklearn."""
+
 _valid_task_types: tuple[TaskTypeName, ...] = (
     "binary",
     "multiclass",
@@ -367,14 +362,14 @@ def _fit(
             # God I feel like I'm doing javascript
             if np.isfinite(v) != True:  # noqa: E712
                 raise ValueError(
-                    f"Scorer {k} returned {v} for train fold. The scorer should"
+                    f"Scorer {k} returned {v} for train split. The scorer should"
                     " should return a finite float",
                 )
 
     return estimator, train_scores
 
 
-def _score_val_fold(
+def _score_val_split(
     estimator: BaseEstimator,
     X: pd.DataFrame | np.ndarray,  # noqa: N803
     y: pd.Series | pd.DataFrame | np.ndarray,
@@ -408,14 +403,14 @@ def _score_val_fold(
             # God I feel like I'm doing javascript
             if np.isfinite(v) != True:  # noqa: E712
                 raise ValueError(
-                    f"Scorer {k} returned {v} for validation fold. The scorer"
+                    f"Scorer {k} returned {v} for validation split. The scorer"
                     " should return a finite float",
                 )
 
         return scores
 
 
-def _evaluate_fold(
+def _evaluate_split(
     estimator: BaseEstimatorT,
     X: pd.DataFrame | np.ndarray,  # noqa: N803
     y: pd.Series | pd.DataFrame | np.ndarray,
@@ -441,7 +436,7 @@ def _evaluate_fold(
         scorers=scorers if train_score else None,
     )
 
-    val_scores = _score_val_fold(
+    val_scores = _score_val_split(
         estimator=fitted_estimator,
         X=X,
         y=y,
@@ -495,7 +490,7 @@ def _iter_cross_validate(
 
     for i_train, i_test in indicies:
         # Sadly this function needs the full X and y due to its internal checks
-        yield _evaluate_fold(
+        yield _evaluate_split(
             estimator=estimator,
             X=X_loaded,
             y=y_loaded,
@@ -612,21 +607,21 @@ def cross_validate_task(  # noqa: D103, C901, PLR0915, PLR0913
             for i, (_trained_est, _val_scores, _train_scores) in trial.profiler.each(
                 enumerate(cv_iter),
                 name="cv",
-                itr_name="fold",
+                itr_name="split",
             ):
                 # Update the report
                 if store_models:
                     trial.store({f"model_{i}.pkl": _trained_est})
 
-                fold_scores = {f"fold_{i}:{k}": v for k, v in _val_scores.items()}
+                split_scores = {f"split_{i}:{k}": v for k, v in _val_scores.items()}
 
-                trial.summary.update(fold_scores)
+                trial.summary.update(split_scores)
                 for k, v in _val_scores.items():
                     all_val_scores[k].append(v)
 
                 if _train_scores is not None:
                     train_scores = {
-                        f"fold_{i}:train_{k}": v for k, v in _train_scores.items()
[…]
