### 2024-02-12 7af22d9
fix(Comm): Check for cancelled task before looking for exception
