### 2024-02-12 7e42aa8
test: Fix assertion to use `"split_"` instead of `"fold_"`
