### 2024-02-12 82df97f
feat: CVEarlyStopping (#254)
