+++ src/amltk/scheduling/events.py
@@ -243,19 +243,22 @@ import logging
 import math
 import time
 from collections import Counter, defaultdict
-from collections.abc import Callable, Iterable, Iterator, Mapping
+from collections.abc import Callable, Iterable
 from dataclasses import dataclass, field
 from functools import partial
-from typing import TYPE_CHECKING, Any, Generic, TypeVar, overload
+from typing import TYPE_CHECKING, Any, Generic, TypeAlias, TypeVar, overload
 from typing_extensions import ParamSpec, override
-from uuid import uuid4
 
-from amltk._functional import callstring, funcname
+from more_itertools import first_true
+
+from amltk._functional import funcname
 from amltk._richutil.renderers.function import Function
+from amltk.exceptions import EventNotKnownError
+from amltk.randomness import randuid
 
 if TYPE_CHECKING:
+    from rich.console import RenderableType
     from rich.text import Text
-    from rich.tree import Tree
 
 T = TypeVar("T")
 P = ParamSpec("P")
@@ -273,12 +276,12 @@ class RegisteredTimeCallOrderStrategy:
         cls,
         events: Iterable[
             tuple[
-                Iterable[Handler[P]],
+                Iterable[Handler[P, R]],
                 tuple[Any, ...] | None,
                 dict[str, Any] | None,
             ]
         ],
-    ) -> None:
+    ) -> list[tuple[Handler[P, R], R | None]]:
         """Call all events in the scheduler."""
         all_handlers = []
         for handlers, args, kwargs in events:
@@ -287,12 +290,14 @@ class RegisteredTimeCallOrderStrategy:
             ]
 
         sorted_handlers = sorted(all_handlers, key=lambda item: item[0].registered_at)
-        for handler, args, kwargs in sorted_handlers:
-            handler(*args, **kwargs)
+        return [
+            (handler, handler(*args, **kwargs))
+            for handler, args, kwargs in sorted_handlers
+        ]
 
 
 @dataclass(frozen=True)
-class Event(Generic[P]):
+class Event(Generic[P, R]):
     """An event that can be emitted.
 
     Attributes:
@@ -321,7 +326,7 @@ class Event(Generic[P]):
 
 
 @dataclass
-class Subscriber(Generic[P]):
+class Subscriber(Generic[P, R]):
     """An object that can be used to easily subscribe to a certain event.
 
     ```python
@@ -353,53 +358,24 @@ class Subscriber(Generic[P]):
     """
 
     emitter: Emitter
-    event: Event[P]
-    when: Callable[[], bool] | None = None
-    max_calls: int | None = None
-    repeat: int = 1
-    every: int = 1
+    event: Event[P, R]
 
     @property
     def event_counts(self) -> int:
         """The number of times this event has been emitted."""
         return self.emitter.event_counts[self.event]
 
-    @overload
-    def __call__(
+    def register(
         self,
-        callback: None = None,
-        *,
-        when: Callable[[], bool] | None = ...,
-        max_calls: int | None = ...,
-        repeat: int = ...,
-        every: int = ...,
-    ) -> partial[Callable[P, Any]]:
-        ...
-
-    @overload
-    def __call__(
-        self,
-        callback: Callable[P, Any],
-        *,
-        when: Callable[[], bool] | None = ...,
-        max_calls: int | None = ...,
-        repeat: int = ...,
-        every: int = ...,
-        hidden: bool = ...,
-    ) -> Callable[P, Any]:
-        ...
-
-    def __call__(
-        self,
-        callback: Callable[P, Any] | None = None,
+        callback: Callable[P, R],
         *,
         when: Callable[[], bool] | None = None,
         max_calls: int | None = None,
         repeat: int = 1,
         every: int = 1,
[…]
