### 2024-02-14 a7316b5
fix(scheduling): Safety against process lifetime race condition
