### 2024-02-14 099dfca
fix(scheduling): Buffer between `SIGTERM` and `SIGKILL` to 5 seconds
