### 2024-02-14 b579188
fix(scheduling): Safe termination of processes, avoiding lifetime race condition (#256)
