### 2024-02-14 fd8c5a3
feat(sklearn): `CVEvaluator` now can compute test scores
