### 2024-02-14 b1bea39
test: `CVEvaluator` accepts test_scorer params
