### 2024-02-14 e393a6c
doc: Error msg typo fix
