### 2024-02-14 3b41ebf
feat(sklearn): `X_test`, `y_test` to CVEvaluator (#258)
