+++ src/amltk/sklearn/evaluation.py
@@ -345,26 +345,27 @@ def _fit(
 
     train_scores = None
     if scorers is not None:
-        train_scores = _score(
-            estimator=estimator,
-            X_test=X_train,
-            y_test=y_train,
-            scorer=scorers,
-            score_params=_scorer_params_train,
-            error_score="raise",
-        )
-        assert isinstance(train_scores, dict)
-        for k, v in train_scores.items():
-            # Can return list or np.bool_
-            # We do not want a list to pass (i.e. if [x] shouldn't pass if check)
-            # Also, we can't use `np.bool_` is `True` as `np.bool_(True) is not True`.
-            # Hence we have to use equality checking
-            # God I feel like I'm doing javascript
-            if np.isfinite(v) != True:  # noqa: E712
-                raise ValueError(
-                    f"Scorer {k} returned {v} for train split. The scorer should"
-                    " should return a finite float",
-                )
+        with profiler("train_score"):
+            train_scores = _score(
+                estimator=estimator,
+                X_test=X_train,
+                y_test=y_train,
+                scorer=scorers,
+                score_params=_scorer_params_train,
+                error_score="raise",
+            )
+            assert isinstance(train_scores, dict)
+            for k, v in train_scores.items():
+                # Can return list or np.bool_
+                # We do not want a list to pass (i.e. if [x] shouldn't pass if check)
+                # We can't use `np.bool_` is `True` as `np.bool_(True) is not True`.
+                # Hence we have to use equality checking
+                # God I feel like I'm doing javascript
+                if np.isfinite(v) != True:  # noqa: E712
+                    raise ValueError(
+                        f"Scorer {k} returned {v} for train split. The scorer should"
+                        " should return a finite float",
+                    )
 
     return estimator, train_scores
 
@@ -410,19 +411,27 @@ def _score_val_split(
         return scores
 
 
-def _evaluate_split(
+def _evaluate_split(  # noqa: PLR0913
     estimator: BaseEstimatorT,
     X: pd.DataFrame | np.ndarray,  # noqa: N803
     y: pd.Series | pd.DataFrame | np.ndarray,
     *,
+    X_test: np.ndarray | pd.DataFrame | None = None,  # noqa: N803
+    y_test: np.ndarray | pd.Series | pd.DataFrame | None = None,
     i_train: np.ndarray,
     i_val: np.ndarray,
     profiler: Profiler,
     scorers: _MultimetricScorer,
     fit_params: Mapping[str, Any],
     scorer_params: Mapping[str, Any],
+    test_scorer_params: Mapping[str, Any],
     train_score: bool,
-) -> tuple[BaseEstimatorT, Mapping[str, float], Mapping[str, float] | None]:
+) -> tuple[
+    BaseEstimatorT,
+    Mapping[str, float],
+    Mapping[str, float] | None,
+    Mapping[str, float] | None,
+]:
     # These return new dictionaries
 
     fitted_estimator, train_scores = _fit(
@@ -447,7 +456,34 @@ def _evaluate_split(
         profiler=profiler,
     )
 
-    return fitted_estimator, val_scores, train_scores
+    test_scores = None
+    if X_test is not None and y_test is not None:
+        with profiler("test_score"):
+            test_scores = _score(
+                estimator=fitted_estimator,
+                X_test=X_test,
+                y_test=y_test,
+                scorer=scorers,
+                score_params=test_scorer_params,
+                error_score="raise",
+            )
+
+            # NOTE: Despite `error_score="raise"`, this will not raise
+            # for `inf` or `nan` values.
+            assert isinstance(test_scores, dict)
+            for k, v in test_scores.items():
+                # Can return list or np.bool_
[…]
