+++ src/amltk/sklearn/evaluation.py
@@ -14,11 +14,12 @@ import logging
 import tempfile
 import warnings
 from asyncio import Future
-from collections import defaultdict
 from collections.abc import Callable, Iterator, Mapping, MutableMapping, Sized
 from contextlib import nullcontext
+from dataclasses import dataclass
 from datetime import datetime
 from functools import partial
+from numbers import Number
 from pathlib import Path
 from typing import (
     TYPE_CHECKING,
@@ -110,6 +111,47 @@ _valid_task_types: tuple[TaskTypeName, ...] = (
     "continuous",
     "continuous-multioutput",
 )
+
+XLike: TypeAlias = pd.DataFrame | np.ndarray
+"""A type alias for X input data type as defined by sklearn."""
+
+YLike: TypeAlias = pd.Series | pd.DataFrame | np.ndarray
+"""A type alias for y input data type as defined by sklearn."""
+
+PostSplitSignature: TypeAlias = Callable[
+    ["CVEvaluation.SplitEval"],
+    "CVEvaluation.SplitEval",
+]
+"""A type alias for the post split callback signature.
+
+Please see [`CVEvaluation.SplitEval`][amltk.sklearn.evaluation.CVEvaluation.SplitEval]
+for more information on the information available to act upon.
+
+```python
+def my_post_split(eval: CVEvalauation.SplitEval) -> CVEvaluation.SplitEval:
+    ...
+```
+"""
+
+
+def _check_valid_scores(
+    scores: Mapping[str, float] | Number,
+    split: str,
+) -> Mapping[str, float]:
+    assert isinstance(scores, Mapping)
+    for k, v in scores.items():
+        # Can return list or np.bool_
+        # We do not want a list to pass (i.e. if [x] shouldn't pass if check)
+        # We can't use `np.bool_` is `True` as `np.bool_(True) is not True`.
+        # Hence we have to use equality checking
+        # God I feel like I'm doing javascript
+        if np.isfinite(v) != True:  # noqa: E712
+            raise ValueError(
+                f"Scorer {k} returned {v} for {split} split. The scorer should"
+                " should return a finite float",
+            )
+
+    return scores
 
 
 def _route_params(
@@ -213,7 +255,7 @@ def _default_cv_resampler(
 
 
 def identify_task_type(  # noqa: PLR0911
-    y: np.ndarray | pd.Series | pd.DataFrame,
+    y: YLike,
     *,
     task_hint: Literal["classification", "regression", "auto"] = "auto",
 ) -> TaskTypeName:
@@ -324,19 +366,19 @@ def identify_task_type(  # noqa: PLR0911
 
 def _fit(
     estimator: BaseEstimatorT,
-    X: pd.DataFrame | np.ndarray,  # noqa: N803
-    y: pd.Series | pd.DataFrame | np.ndarray,
-    *,
+    X: XLike,  # noqa: N803
+    y: YLike,
     i_train: np.ndarray,
+    *,
     profiler: Profiler,
     fit_params: Mapping[str, Any],
+    train_score: bool,
+    scorers: dict[str, _Scorer] | _MultimetricScorer,
     scorer_params: Mapping[str, Any],
-    scorers: _MultimetricScorer | None = None,
-) -> tuple[BaseEstimatorT, Mapping[str, float] | None]:
+) -> tuple[BaseEstimatorT, dict[str, Any], Mapping[str, float] | None, dict[str, Any]]:
     _fit_params = _check_method_params(X, params=fit_params, indices=i_train)
-    _scorer_params_train = _check_method_params(X, scorer_params, indices=i_train)
-
     X_train, y_train = _safe_split(estimator, X, y, indices=i_train)
+
     with profiler("fit"):
         if y_train is None:
[…]
