### 2024-02-16 7b13c67
test: Ensure callback hooks work
