+++ src/amltk/sklearn/evaluation.py
@@ -68,6 +68,7 @@ from amltk.exceptions import (
     MismatchedTaskTypeWarning,
     TrialError,
 )
+from amltk.optimization import Trial
 from amltk.profiling.profiler import Profiler
 from amltk.scheduling import Plugin, Task
 from amltk.scheduling.events import Emitter, Event
@@ -81,7 +82,6 @@ if TYPE_CHECKING:
         BaseShuffleSplit,
     )
 
-    from amltk.optimization import Trial
     from amltk.pipeline import Node
     from amltk.randomness import Seed
 
@@ -119,8 +119,8 @@ YLike: TypeAlias = pd.Series | pd.DataFrame | np.ndarray
 """A type alias for y input data type as defined by sklearn."""
 
 PostSplitSignature: TypeAlias = Callable[
-    ["CVEvaluation.SplitEval"],
-    "CVEvaluation.SplitEval",
+    [Trial, int, "CVEvaluation.PostSplitInfo"],
+    "CVEvaluation.PostSplitInfo",
 ]
 """A type alias for the post split callback signature.
 
@@ -128,7 +128,11 @@ Please see [`CVEvaluation.SplitEval`][amltk.sklearn.evaluation.CVEvaluation.Spli
 for more information on the information available to act upon.
 
 ```python
-def my_post_split(eval: CVEvalauation.SplitEval) -> CVEvaluation.SplitEval:
+def my_post_split(
+    trial: Trial,
+    split_number: int,
+    eval: CVEvalauation.PostSplitInfo
+) -> CVEvaluation.PostSplitInfo:
     ...
 ```
 """
@@ -452,7 +456,7 @@ def _evaluate_split(  # noqa: PLR0913
     scorer_params: Mapping[str, Any],
     test_scorer_params: Mapping[str, Any],
     train_score: bool,
-) -> CVEvaluation.SplitEval:
+) -> CVEvaluation.PostSplitInfo:
     # We shove all logic that requires indexing into X for train into `_fit`.
     # This is because it's easy to create an accidental copy, i.e. with _safe_split.
     # We want that memory to only exist inside that `_fit` function and to not persists
@@ -493,7 +497,11 @@ def _evaluate_split(  # noqa: PLR0913
             )
             test_scores = _check_valid_scores(test_scores, split="test")
 
-    return CVEvaluation.SplitEval(
+    return CVEvaluation.PostSplitInfo(
+        X=X,
+        y=y,
+        X_test=X_test,
+        y_test=y_test,
         i_train=i_train,
         i_val=i_val,
         model=fitted_estimator,
@@ -516,14 +524,13 @@ def _iter_cross_validate(  # noqa: PLR0913
     *,
     X_test: XLike | None = None,  # noqa: N803
     y_test: YLike | None = None,
-    post_split: PostSplitSignature | None = None,
     fit_params: Mapping[str, Any] | None = None,
     scorer_params: Mapping[str, Any] | None = None,
     splitter_params: Mapping[str, Any] | None = None,
     test_scorer_params: Mapping[str, Any] | None = None,
     profiler: Profiler | None = None,
     train_score: bool = False,
-) -> Iterator[CVEvaluation.SplitEval]:
+) -> Iterator[CVEvaluation.PostSplitInfo]:
     if (X_test is not None and y_test is None) or (
         y_test is not None and X_test is None
     ):
@@ -540,7 +547,7 @@ def _iter_cross_validate(  # noqa: PLR0913
 
     for i_train, i_val in splitter.split(X, y, **splitter_params):
         # Sadly this function needs the full X and y due to its internal checks
-        split_eval = _evaluate_split(
+        yield _evaluate_split(
             estimator=estimator,
             X=X,
             y=y,
[…]
