### 2024-02-17 a0d2378
refactor(Evaluation): Group scores for easier mean/std calculation
