+++ src/amltk/sklearn/evaluation.py
@@ -14,6 +14,7 @@ import logging
 import tempfile
 import warnings
 from asyncio import Future
+from collections import defaultdict
 from collections.abc import Callable, Iterator, Mapping, MutableMapping, Sized
 from contextlib import nullcontext
 from dataclasses import dataclass
@@ -729,7 +730,11 @@ def cross_validate_task(  # noqa: D103, C901, PLR0915, PLR0913
         test_scorer_params=test_scorer_params,
     )
 
-    split_infos: list[CVEvaluation.SplitScores] = []
+    split_scores = CVEvaluation.SplitScores(
+        val=defaultdict(list),
+        train=defaultdict(list) if train_score else None,
+        test=defaultdict(list) if X_test is not None else None,
+    )
     models: list[BaseEstimator] | None = (
         None if not post_processing_requires_models else []
     )
@@ -756,6 +761,8 @@ def cross_validate_task(  # noqa: D103, C901, PLR0915, PLR0913
                 trial.summary.update(
                     {f"split_{i}:val_{k}": v for k, v in split_eval.val_scores.items()},
                 )
+                for k, v in split_eval.val_scores.items():
+                    split_scores.val[k].append(v)
 
                 if split_eval.train_scores is not None:
                     trial.summary.update(
@@ -764,6 +771,8 @@ def cross_validate_task(  # noqa: D103, C901, PLR0915, PLR0913
                             for k, v in split_eval.train_scores.items()
                         },
                     )
+                    for k, v in split_eval.train_scores.items():
+                        split_scores.train[k].append(v)  # type: ignore
 
                 if split_eval.test_scores is not None:
                     trial.summary.update(
@@ -772,14 +781,8 @@ def cross_validate_task(  # noqa: D103, C901, PLR0915, PLR0913
                             for k, v in split_eval.test_scores.items()
                         },
                     )
-
-                split_infos.append(
-                    CVEvaluation.SplitScores(
-                        val_scores=split_eval.val_scores,
-                        train_scores=split_eval.train_scores,
-                        test_scores=split_eval.test_scores,
-                    ),
-                )
+                    for k, v in split_eval.test_scores.items():
+                        split_scores.test[k].append(v)  # type: ignore
 
                 if post_processing_requires_models:
                     assert models is not None
@@ -795,7 +798,7 @@ def cross_validate_task(  # noqa: D103, C901, PLR0915, PLR0913
                 # should we continue or stop?
                 if comm is not None and i < n_splits:
                     match response := comm.request(
-                        (trial, list(split_infos)),
+                        (trial, split_scores),
                         timeout=10,
                     ):
                         case True:
@@ -828,7 +831,7 @@ def cross_validate_task(  # noqa: D103, C901, PLR0915, PLR0913
                     y_test=_y_test,
                     splitter=splitter,
                     max_splits=n_splits,
-                    split_info=split_infos,
+                    scores=split_scores,
                     scorers=scorers,
                     models=models,
                     splitter_params=splitter_params,
@@ -840,37 +843,21 @@ def cross_validate_task(  # noqa: D103, C901, PLR0915, PLR0913
 
             return report
         else:
+            for mname, fold_scores in split_scores.val.items():
+                trial.summary[f"val_mean_{mname}"] = float(np.mean(fold_scores))
+                trial.summary[f"val_std_{mname}"] = float(np.std(fold_scores))
 
-            def _process_scores(
-                _scores: dict[str, list[float]],
-                _name: Literal["train", "val", "test"],
-            ) -> dict[str, float]:
-                _means = {k: float(np.mean(scores)) for k, scores in _scores.items()}
[…]
