### 2024-02-17 dead41d
doc: Type fix
