+++ tests/sklearn/test_evaluation.py
@@ -14,6 +14,7 @@ import sklearn.datasets
 import sklearn.pipeline
 from pytest_cases import case, parametrize, parametrize_with_cases
 from sklearn import config_context as sklearn_config_context
+from sklearn.base import check_is_fitted
 from sklearn.cluster import KMeans
 from sklearn.datasets import make_classification, make_regression
 from sklearn.dummy import DummyClassifier
@@ -30,7 +31,7 @@ from sklearn.tree import DecisionTreeClassifier, DecisionTreeRegressor
 
 from amltk.exceptions import TaskTypeWarning, TrialError
 from amltk.optimization.trial import Metric, Trial
-from amltk.pipeline import Component, request
+from amltk.pipeline import Component, Node, request
 from amltk.pipeline.builders.sklearn import build as sklearn_pipeline_builder
 from amltk.sklearn.evaluation import (
     CVEvaluation,
@@ -396,9 +397,9 @@ def test_evaluator(  # noqa: PLR0912
     ]
     for metric_name in expected_summary_scorers:
         for i in range(n_splits):
-            assert f"split_{i}:{metric_name}" in report.summary
-        assert f"mean_{metric_name}" in report.summary
-        assert f"std_{metric_name}" in report.summary
+            assert f"split_{i}:val_{metric_name}" in report.summary
+        assert f"val_mean_{metric_name}" in report.summary
+        assert f"val_std_{metric_name}" in report.summary
 
     if train_score:
         assert "cv:train_score" in report.profiles
@@ -675,7 +676,7 @@ def test_estimator_params_get_forward(tmp_path: Path) -> None:
 
 
 def test_evaluator_with_clustering(tmp_path: Path) -> None:
-    x, y = sklearn.datasets.make_blobs(
+    x, y = sklearn.datasets.make_blobs(  # type: ignore
         n_samples=20,
         centers=2,
         n_features=2,
@@ -813,7 +814,11 @@ def test_early_stopping_plugin(tmp_path: Path) -> None:
         def update(self, report: Trial.Report) -> None:
             pass  # Normally you would update w.r.t. a finished trial
 
-        def should_stop(self, info: CVEvaluation.SplitInfo) -> bool:  # noqa: ARG002
+        def should_stop(
+            self,
+            trial: Trial,  # noqa: ARG002
+            scores: CVEvaluation.SplitScores,  # noqa: ARG002
+        ) -> bool:
             # Just say yes, should stop
             return True
 
@@ -834,8 +839,8 @@ def test_early_stopping_plugin(tmp_path: Path) -> None:
     assert "Early stop" in str(report.exception)
 
     # Only the first fold should have been run and put in summary
-    assert "split_0:accuracy" in report.summary
-    assert "split_1:accuracy" not in report.summary
+    assert "split_0:val_accuracy" in report.summary
+    assert "split_1:val_accuracy" not in report.summary
 
 
 def test_that_test_scorer_params_can_be_forwarded(tmp_path: Path) -> None:
@@ -917,3 +922,85 @@ def test_that_test_scorer_params_can_be_forwarded(tmp_path: Path) -> None:
 
         assert report.values["custom_metric"] == 1
         assert report.summary["test_mean_custom_metric"] == 2
+
+
+def record_split_number(
+    trial: Trial,
+    split_number: int,
+    info: CVEvaluation.PostSplitInfo,
+) -> CVEvaluation.PostSplitInfo:
+    # Should get the test data if it was passed in as it is in the function below
+    assert info.X_test is not None
+    assert info.y_test is not None
+    check_is_fitted(info.model)
+
[…]
