+++ src/amltk/sklearn/evaluation.py
@@ -17,8 +17,10 @@ from asyncio import Future
 from collections import defaultdict
 from collections.abc import Callable, Iterator, Mapping, MutableMapping, Sized
 from contextlib import nullcontext
+from dataclasses import dataclass
 from datetime import datetime
 from functools import partial
+from numbers import Number
 from pathlib import Path
 from typing import (
     TYPE_CHECKING,
@@ -67,6 +69,7 @@ from amltk.exceptions import (
     MismatchedTaskTypeWarning,
     TrialError,
 )
+from amltk.optimization import Trial
 from amltk.profiling.profiler import Profiler
 from amltk.scheduling import Plugin, Task
 from amltk.scheduling.events import Emitter, Event
@@ -80,7 +83,6 @@ if TYPE_CHECKING:
         BaseShuffleSplit,
     )
 
-    from amltk.optimization import Trial
     from amltk.pipeline import Node
     from amltk.randomness import Seed
 
@@ -110,6 +112,51 @@ _valid_task_types: tuple[TaskTypeName, ...] = (
     "continuous",
     "continuous-multioutput",
 )
+
+XLike: TypeAlias = pd.DataFrame | np.ndarray
+"""A type alias for X input data type as defined by sklearn."""
+
+YLike: TypeAlias = pd.Series | pd.DataFrame | np.ndarray
+"""A type alias for y input data type as defined by sklearn."""
+
+PostSplitSignature: TypeAlias = Callable[
+    [Trial, int, "CVEvaluation.PostSplitInfo"],
+    "CVEvaluation.PostSplitInfo",
+]
+"""A type alias for the post split callback signature.
+
+Please see [`PostSplitInfo`][amltk.sklearn.evaluation.CVEvaluation.PostSplitInfo]
+for more information on the information available to this callback.
+
+```python
+def my_post_split(
+    trial: Trial,
+    split_number: int,
+    eval: CVEvalauation.PostSplitInfo
+) -> CVEvaluation.PostSplitInfo:
+    ...
+```
+"""
+
+
+def _check_valid_scores(
+    scores: Mapping[str, float] | Number,
+    split: str,
+) -> Mapping[str, float]:
+    assert isinstance(scores, Mapping)
+    for k, v in scores.items():
+        # Can return list or np.bool_
+        # We do not want a list to pass (i.e. if [x] shouldn't pass if check)
+        # We can't use `np.bool_` is `True` as `np.bool_(True) is not True`.
+        # Hence we have to use equality checking
+        # God I feel like I'm doing javascript
+        if np.isfinite(v) != True:  # noqa: E712
+            raise ValueError(
+                f"Scorer {k} returned {v} for {split} split. The scorer should"
+                " should return a finite float",
+            )
+
+    return scores
 
 
 def _route_params(
@@ -213,7 +260,7 @@ def _default_cv_resampler(
 
 
 def identify_task_type(  # noqa: PLR0911
-    y: np.ndarray | pd.Series | pd.DataFrame,
+    y: YLike,
     *,
     task_hint: Literal["classification", "regression", "auto"] = "auto",
 ) -> TaskTypeName:
@@ -324,19 +371,19 @@ def identify_task_type(  # noqa: PLR0911
 
 def _fit(
     estimator: BaseEstimatorT,
-    X: pd.DataFrame | np.ndarray,  # noqa: N803
-    y: pd.Series | pd.DataFrame | np.ndarray,
-    *,
+    X: XLike,  # noqa: N803
+    y: YLike,
     i_train: np.ndarray,
+    *,
     profiler: Profiler,
     fit_params: Mapping[str, Any],
+    train_score: bool,
[…]
