### 2024-02-17 #6367 @editorialbot
[REVIEW]: AMLTK: A Modular AutoML Toolkit in Python
**Submitting author:** @eddiebergman (<a href="http://orcid.org/0009-0003-4390-7614">Edward Bergman</a>)
**Repository:** https://github.com/automl/amltk
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.12.1
