### 2024-02-22 703fddb
fix(pipeline): Ensure optimizer is updated with report
