### 2024-02-22 74993c9
fix(trial): Don't record metric values for deserialized NaN's or None (#263)
