### 2024-02-23 f291cbf
fix(trial): Bucket name typo fix
