### 2024-02-29 9a6010b
fix(bib): Include doi for references
