+++ paper.bib
@@ -3,6 +3,8 @@
   author    = {N. Erickson and J. Mueller and A. Shirkov and H. Zhang and P. Larroy and M. Li and A. Smola},
   year      = 2020,
   journal   = {arXiv:2003.06505 [stat.ML]},
+  doi       = {10.48550/arXiv.2003.06505},
+}
 }
 
 @inproceedings{feurer-nips15a,
@@ -36,6 +38,7 @@
   title     = {{AMLB}: an AutoML Benchmark},
   journal   = {arXiv:2207.12560 {[stat.ML]}},
   year      = {2023},
+  doi       = {10.48550/arxiv.2207.12560},
 }
 
 @article{zimmer-tpami21a,
@@ -46,6 +49,7 @@
   volume    = {43},
   pages     = {3079--3090},
   issue     = {9},
+  doi       = {10.1109/tpami.2021.3067763},
 }
 
 @book{hutter-book19a,
@@ -55,6 +59,7 @@
   publisher = {Springer},
   note      = {Available for free at \url{http://automl.org/book}},
   editor    = {F. Hutter and L. Kotthoff and J. Vanschoren},
+  doi       = {10.1007/978-3-030-05318-5},
 }
 
 @inproceedings{gijsbers-kdd21a,
@@ -64,6 +69,7 @@
   year      = {2021},
   publisher = {Springer},
   pages     = {560--564},
+  doi       = {10.1007/978-3-030-67670-4_39},
 }
 
 @article{gijsbers-joss19a,
@@ -73,6 +79,7 @@
   journal   = {Journal of Open Source Software},
   volume    = {4},
   issue     = {33}
+  doi       = {10.21105/joss.01132},
 }
 
 @article{lindauer-jmlr22a,
@@ -90,6 +97,7 @@
   author    = {R. Sass and E. Bergman and A. Biedenkapp and F. Hutter and M. Lindauer},
   year      = {2022},
   booktitle = {{ICML} Adaptive Experimental Design and Active Learning in the Real World (ReALML Workshop 2022)},
+  doi       = {10.48550/arxiv.2206.03493},
 }
 
 @article{lindauer-arxiv19a,
@@ -97,6 +105,8 @@
   author    = {M. Lindauer and K. Eggensperger and M. Feurer and A. Biedenkapp and J. Marben and P. Müller and F. Hutter},
   journal   = {arXiv:1908.06756 {[cs.LG]}},
   year      = {2019},
+  doi       = {10.48550/arxiv.1908.06756},
+}
 }
 
 @article{mohr-ml23a,
@@ -107,6 +117,7 @@
   pages     = {1131–1170},
   volume    = {112},
   number    = {4}
+  doi       = {10.1007/s10994-022-06200-0},
 }
 
 @inproceedings{akiba-kdd19a,
@@ -115,6 +126,7 @@
   year      = {2019},
   booktitle = {Proceedings of the 25th {ACM} {SIGKDD} International Conference on Knowledge Discovery {\&} Data Mining ({KDD}'19)},
   pages     = {2623--2631},
+  doi       = {10.1145/3292500.3330701},
 }
 
 @inproceedings{thornton-kdd13a,
@@ -123,6 +135,7 @@
   year      = {2013},
   booktitle = {Proceedings of the 19th {ACM} {SIGKDD} international conference on Knowledge discovery {\&} Data Mining ({KDD}'13)},
   pages     = {847--855},
+  doi       = {10.1007/978-3-030-05318-5_4},
 }
 
 @inproceedings{olson-gecco16a,
@@ -131,6 +144,7 @@
   booktitle = {Proceedings of the Genetic and Evolutionary Computation Conference 2016},
   year      = {2016},
[…]
