### 2024-02-29 0789516
fix(bib): Trailing commas
