+++ paper.bib
@@ -78,7 +78,7 @@
   title     = {{GAMA}: Genetic Automated Machine Learning Assistant},
   journal   = {Journal of Open Source Software},
   volume    = {4},
-  issue     = {33}
+  issue     = {33},
   doi       = {10.21105/joss.01132},
 }
 
@@ -89,7 +89,7 @@
   journal   = {Journal of Machine Learning Research},
   volume    = {23},
   number    = {54},
-  pages     = {1-9}
+  pages     = {1-9},
 }
 
 @inproceedings{sass-realml22a,
@@ -116,7 +116,7 @@
   journal   = {Machine Learning},
   pages     = {1131–1170},
   volume    = {112},
-  number    = {4}
+  number    = {4},
   doi       = {10.1007/s10994-022-06200-0},
 }
 
@@ -151,7 +151,7 @@
   author    = {F. Imrie and B. Cebere and E. McKinney and M. van der Schaar},
   title     = {AutoPrognosis 2.0: Democratizing Diagnostic and Prognostic Modeling in Healthcare with Automated Machine Learning},
   year      = {2022},
-  journal   = {arXiv:2210.12090 {[cs.LG]}}
+  journal   = {arXiv:2210.12090 {[cs.LG]}},
   doi       = {10.1371/journal.pdig.0000276},
 }
 
@@ -161,5 +161,5 @@
   journal   = {Proceedings of Machine Learning and Systems},
   volume    = {3},
   pages     = {434--447},
-  year      = {2021}
+  year      = {2021},
 }
