### 2024-02-29 #6367 @eddiebergman
Apologies, I missed that this was required in the `paper.bib`. I've included a doi for all entries where I could find one. Is there any command to have the bot re-check this?
