### 2024-03-01 af190b1
feat(Pytorch): Add an example of building an MLP from a pipeline
