+++ src/pytorch_example.py
@@ -1,313 +1,0 @@
-# Check the `main()` function to get started and follow it through.
-# Note that performance is irrelevant for now.
-# Most of my pytorch stuff as an example is just taken from here.
-# https://github.com/pytorch/examples/blob/main/mnist/main.py
-from __future__ import annotations
-
-from collections import OrderedDict
-from typing import TYPE_CHECKING
-
-import torch
-import torch.nn.functional as F  # noqa: N812
-from torch import nn, optim
-from torch.optim.lr_scheduler import StepLR
-from torchvision import datasets, transforms
-
-from amltk import Component, Metric, Sequential
-
-# Change this to optuna if you prefer
-# from amltk.optimization.optimizers.optuna import OptunaParser
-from amltk.optimization.optimizers.smac import SMACOptimizer
-
-if TYPE_CHECKING:
-    from amltk import Node, Trial
-
-# This is a nice import :)
-from rich import print
-
-
-# NOTE: This is the reference model, slowly try to build up to this
-# but make it parametrizable.
-# The goal would be that users don't define this class (maybe?)
-# but they can define it using the pipeline structure.
-# We can handle that later, for now, the pipeline definition below should be enough.
-class Net(nn.Module):
-    def __init__(self):
-        super().__init__()
-        self.conv1 = nn.Conv2d(1, 32, 3, 1)
-        self.conv2 = nn.Conv2d(32, 64, 3, 1)
-        self.dropout1 = nn.Dropout(0.25)
-        self.dropout2 = nn.Dropout(0.5)
-        self.fc1 = nn.Linear(9216, 128)
-        self.fc2 = nn.Linear(128, 10)
-
-    def forward(self, x):
-        x = self.conv1(x)
-        x = F.relu(x)
-        x = self.conv2(x)
-        x = F.relu(x)
-        x = F.max_pool2d(x, 2)
-        x = self.dropout1(x)
-        x = torch.flatten(x, 1)
-        x = self.fc1(x)
-        x = F.relu(x)
-        x = self.dropout2(x)
-        x = self.fc2(x)
-        return F.log_softmax(x, dim=1)
-
-
-def train(args, model, device, train_loader, optimizer, epoch):
-    model.train()
-    for batch_idx, (data, target) in enumerate(train_loader):
-        data, target = data.to(device), target.to(device)
-        optimizer.zero_grad()
-        output = model(data)
-        loss = F.nll_loss(output, target)
-        loss.backward()
-        optimizer.step()
-        if batch_idx % args.log_interval == 0:
-            print('Train Epoch: {} [{}/{} ({:.0f}%)]\tLoss: {:.6f}'.format(
-                epoch, batch_idx * len(data), len(train_loader.dataset),
-                       100. * batch_idx / len(train_loader), loss.item()))
-            if args.dry_run:
-                break
-
-
-# Just taken from the pytorch example
-def test(
-        model: nn.Module,
-        device: torch.device,
-        test_loader: torch.utils.data.DataLoader,
-) -> tuple[float, float]:
-    model.eval()
-    test_loss = 0
-    correct = 0
-    with torch.no_grad():
-        for data, target in test_loader:
-            data, target = data.to(device), target.to(device)
-            output = model(data)
-            test_loss += F.nll_loss(output, target, reduction="sum").item()
-            pred = output.argmax(dim=1, keepdim=True)
-            correct += pred.eq(target.view_as(pred)).sum().item()
-
-    test_loss /= len(test_loader.dataset)
-    accuracy = 100.0 * correct / len(test_loader.dataset)
-    return float(test_loss), float(accuracy)
-
-
-# NOTE: The idea for this would be to integrate a general enough builder
[…]
