+++ src/amltk/pytorch/pytorch_example.py
@@ -1,0 +1,327 @@
+"""Script for building and evaluating an example of PyTorch MLP model on MNIST dataset.
+The script defines functions for constructing a neural network model from a pipeline, training the model,
+and evaluating its performance.
+
+References:
+- PyTorch MNIST example: https://github.com/pytorch/examples/blob/main/mnist/main.py
+"""
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+import torch
+import torch.nn.functional as F  # noqa: N812
+from torch import nn, optim
+from torch.optim.lr_scheduler import StepLR
+from torchvision import datasets, transforms
+
+from amltk import Component, Metric, Sequential
+
+# Change this to optuna if you prefer
+# from amltk.optimization.optimizers.optuna import OptunaParser
+from amltk.optimization.optimizers.smac import SMACOptimizer
+
+if TYPE_CHECKING:
+    from amltk import Node, Trial
+
+from rich import print
+
+
+class Net(nn.Module):
+    """Reference neural network model."""
+
+    def __init__(self):
+        super().__init__()
+        self.conv1 = nn.Conv2d(1, 32, 3, 1)
+        self.conv2 = nn.Conv2d(32, 64, 3, 1)
+        self.dropout1 = nn.Dropout(0.25)
+        self.dropout2 = nn.Dropout(0.5)
+        self.fc1 = nn.Linear(9216, 128)
+        self.fc2 = nn.Linear(128, 10)
+
+    def forward(self, x):
+        x = self.conv1(x)
+        x = F.relu(x)
+        x = self.conv2(x)
+        x = F.relu(x)
+        x = F.max_pool2d(x, 2)
+        x = self.dropout1(x)
+        x = torch.flatten(x, 1)
+        x = self.fc1(x)
+        x = F.relu(x)
+        x = self.dropout2(x)
+        x = self.fc2(x)
+        return F.log_softmax(x, dim=1)
+
+
+def test(
+    model: nn.Module,
+    device: torch.device,
+    test_loader: torch.utils.data.DataLoader,
+) -> tuple[float, float]:
+    model.eval()
+    test_loss = 0
+    correct = 0
+    with torch.no_grad():
+        for data, target in test_loader:
+            data, target = data.to(device), target.to(device)
+            output = model(data)
+            test_loss += F.nll_loss(output, target, reduction="sum").item()
+            pred = output.argmax(dim=1, keepdim=True)
+            correct += pred.eq(target.view_as(pred)).sum().item()
+
+    test_loss /= len(test_loader.dataset)
+    accuracy = 100.0 * correct / len(test_loader.dataset)
+    return float(test_loss), float(accuracy)
+
+
+def build_model_from_pipeline(pipeline: Sequential) -> nn.Module:
+    """Builds a model from the provided pipeline.
+
+    This function iterates through the pipeline nodes, constructing the model
+    layers dynamically based on the node types and configurations. It also
+    utilizes the `MatchDimensions` objects to handle dimension matching
+    between layers with search spaces.
+
+    Args:
+        pipeline: The pipeline containing the model architecture.
+
+    Returns:
+        The constructed PyTorch model.
+    """
+    print("Building model from pipeline")
+    print(pipeline)
+
+    model_layers = []
+
+    # Add a Flatten layer to ensure input data is flattened
+    model_layers.append(nn.Flatten(start_dim=1))
+
+    for node in pipeline.iter():
+        if isinstance(node.item, type) and issubclass(node.item, nn.Module):
+            layer_config = node.config or {}
+
+            # Evaluate MatchDimensions objects
+            for key, value in layer_config.items():
+                if isinstance(value, MatchDimensions):
[…]
