### 2024-03-04 #6367 @gomezzz
(starting to go through the checklist now)
@eddiebergman The paper has a fairly long author list and I believe JOSS has somewhat more restrictive [authorship criteria](https://joss.readthedocs.io/en/latest/submitting.html#authorship). Could you clarify the contributions of the authors who did not contribute significantly to the code (deduced from here https://github.com/automl/amltk/graphs/contributors)?
