### 2024-03-07 #6367 @eddiebergman
Hi @gomezzz,

The contributions are as follows:
* Myself (Eddie/Edward), Lennart and Aron contributed directly to the code:
    * Aron had put a lot of work in making sure this library could be used to reconstruct the pipelines of AutoSklearn and contributing issues and design
    * Lennart has been invaluable with reviewing every (at least reviewable) PR and discussin features and API design. Unfortunately PR review is not listed in the contributions graph.
* Matthias/Katharina are the original authors of Auto-Sklearn and have been vital for consulting on the technical details of implementation when it comes to interfacing w/ various aspects of Automated Machine Learning, including but not limited to Optimization, Evaluation and Meta-Learning.
* Amir has been one of the earliest forms of quality assurance we have gotten, i.e. he is a user of the tool for his own research projects and has been giving feedback on what does and doesn't work with suggestions.
* Sarah maintains [DeepCave](https://github.com/automl/DeepCAVE) and [SMAC](https://github.com/automl/SMAC3) and we are working on having these tools interop more seamlessly. She has also volunteered to contribute visualizations, once ready, and maintenance to the tool in the long term.
* Frank and Marius have been invaluable in terms of scope, design and desiderata of the project in the larger landscape of AutoML as well as giving strategic feedback on prioritization.

Please let me know if I can further clarify any of these points :)
