### 2024-03-07 #6367 @gomezzz
Thanks @eddiebergman for the details. I am not sure how this interacts with the authorship guidelines of JOSS and if all these are sufficient for authorship. I have no strong opinions on the matter, I'll check the box and maybe @JBorrow can decide on this aspect if there remains any uncertainty.
