### 2024-03-16 9614fae
feat(Pytorch): Add an example of building an MLP from a pipeline
