+++ src/amltk/pytorch/pytorch_example.py
@@ -16,11 +16,12 @@ from torch import nn, optim
 from torch.optim.lr_scheduler import StepLR
 from torchvision import datasets, transforms
 
-from amltk import Component, Metric, Sequential, Choice, Fixed
+from amltk import Component, Metric, Sequential, Choice
 
 # Change this to optuna if you prefer
 # from amltk.optimization.optimizers.optuna import OptunaParser
 from amltk.optimization.optimizers.smac import SMACOptimizer
+from amltk.pytorch.builders import build_model_from_pipeline, MatchDimensions
 
 if TYPE_CHECKING:
     from amltk import Node, Trial
@@ -28,37 +29,10 @@ if TYPE_CHECKING:
 from rich import print
 
 
-class Net(nn.Module):
-    """Reference neural network model."""
-
-    def __init__(self):
-        super().__init__()
-        self.conv1 = nn.Conv2d(1, 32, 3, 1)
-        self.conv2 = nn.Conv2d(32, 64, 3, 1)
-        self.dropout1 = nn.Dropout(0.25)
-        self.dropout2 = nn.Dropout(0.5)
-        self.fc1 = nn.Linear(9216, 128)
-        self.fc2 = nn.Linear(128, 10)
-
-    def forward(self, x):
-        x = self.conv1(x)
-        x = F.relu(x)
-        x = self.conv2(x)
-        x = F.relu(x)
-        x = F.max_pool2d(x, 2)
-        x = self.dropout1(x)
-        x = torch.flatten(x, 1)
-        x = self.fc1(x)
-        x = F.relu(x)
-        x = self.dropout2(x)
-        x = self.fc2(x)
-        return F.log_softmax(x, dim=1)
-
-
 def test(
-    model: nn.Module,
-    device: torch.device,
-    test_loader: torch.utils.data.DataLoader,
+        model: nn.Module,
+        device: torch.device,
+        test_loader: torch.utils.data.DataLoader,
 ) -> tuple[float, float]:
     model.eval()
     test_loss = 0
@@ -76,62 +50,15 @@ def test(
     return float(test_loss), float(accuracy)
 
 
-def build_model_from_pipeline(pipeline: Sequential) -> nn.Module:
-    """Builds a model from the provided pipeline.
-
-    This function iterates through the pipeline nodes, constructing the model
-    layers dynamically based on the node types and configurations. It also
-    utilizes the `MatchDimensions` objects to handle dimension matching
-    between layers with search spaces.
-
-    Args:
-        pipeline: The pipeline containing the model architecture.
-
-    Returns:
-        The constructed PyTorch model.
-    """
-    print("Building model from pipeline")
-    print(pipeline)
-
-    model_layers = []
-
-    for node in pipeline.iter():
-
-        # Check if node is a Flatten layer, ReLU or similar.
-        if isinstance(node, Fixed):
-            model_layers.append( node.item)
-
-        # Check if node is a Component with config parameter
-        elif isinstance(node.item, type) and issubclass(node.item, nn.Module) and node.config:
-            layer_config = node.config or {}
-
-            # Evaluate MatchDimensions objects
-            for key, value in layer_config.items():
-                if isinstance(value, MatchDimensions):
-                    layer_config[key] = value.evaluate(pipeline)
-
-            layer = node.build_item(**layer_config)
-            model_layers.append(layer)
-
-    model = nn.Sequential(*model_layers)
-
-    print("-" * 80)
-    print("Model built")
-    print(model)
-    print("-" * 80)
-
-    return model
-
-
 def eval_configuration(
-    trial: Trial,
-    pipeline: Node,
-    device: str = "cpu",  # Change if you have a GPU
[…]
