### 2024-03-18 automl/amltk#267 @gomezzz
[JOSS Review] Doc Improvements
Looking at the docs, I think there are some small improvements I would suggest to help users find information:

- [ ] The docs page at https://automl.github.io/amltk/latest/ seems to miss some of the essential information provided in the README (e.g. installation instructions, minimal working example). I would suggest to mirror the readme into, e.g., the `Home` part of the docs or at least point to the readme for those instructions.
- [ ] Please mention in the README and/or contribution guideline that GitHub issues should be used for support inquiries, bug reports etc. (This is the case, right?)
- [ ] (optional) For me, the distinction between `guides` and `examples` in the docs is not immediately clear, maybe there are better names? Similarly the reference also looks a lot like the API. I don't know if you use some automated to make sure they stay synchronized? Otherwise, it seems that this may quickly lead to one of the two becoming out-of-date.
- [ ] From an ML standpoint, I think, it would be great if you could provide an overview of frameworks you can interface with. I think this is partially done here https://github.com/automl/amltk/?tab=readme-ov-file#installation but I am not sure if it is exhaustive? Maybe reformulate to clarify if this the full list of supported frameworks?

(opened as part of JOSS Review https://github.com/openjournals/joss-reviews/issues/6367 )
