### 2024-03-20 automl/amltk#267 @eddiebergman
Hi @gomezzz,

Thanks for the feedback! The docs definitely need some refresh in love but to answer some of the points:
* Yup I think some quick install makes sense on the docs homepage.
* Definitely should mention github as the source of issues/feedback, will do.
* I agree there's too much and I was looking at [Textual](https://textual.textualize.io/) as inspiration:
    * API: API doc autogenerated
    * Reference: A hand-written quick summary, i.e. like a cheatsheet
    * Guide: A walkthrough to actually learn about how things work
    * Examples (Admittedly lacking): Copy-and-pastable code with some minor commentary
