### 2024-03-20 automl/amltk#268 @gomezzz
 [JOSS Review] Small error in installation instructions
In the [installation instructions](https://github.com/automl/amltk?tab=readme-ov-file#installation) it is stated for optional dependencies to run

```bash
pip install amltk[notebook]
...
```

I believe this has to be

```bash
pip install "amltk[notebook]"
...
```

to function - at least on Linux. :)

(opened as part of JOSS Review https://github.com/openjournals/joss-reviews/issues/6367 )
