### 2024-03-20 automl/amltk#269 @gomezzz
[JOSS Review]  Pipeline example not working
The [minimal example on pipelines](https://github.com/automl/amltk?tab=readme-ov-file#pipelines) is not working for me.

- [x] There seem to be stray `>>` in the example.
- [x] I believe it requires `sklearn` and `dask`  which you might want to mention in the description.
- [x] There is a typo in the import of RandomForestClassifier
- [x] After fixing all of the above I receive
```bash
    search_space = pipeline.search_space(parser=...)
                   ^^^^^^^^^^^^^^^^^^^^^
AttributeError: 'tuple' object has no attribute 'search_space'
```
- [x] `config` seems to be undefined

(opened as part of JOSS Review https://github.com/openjournals/joss-reviews/issues/6367 )
