### 2024-03-20 e0d15f4
doc: Fixup example in readme
