### 2024-03-25 #6367 @JBorrow
@gomezzz, I have discussed this with a few people and we are happy with author scope. Thank you for raising this.
