### 2024-03-25 #6367 @eddiebergman
Hi @JBorrow,

Apologies, I was away _partially-working_ last week. I will be able to address all comments fully from Wednesday onwards.

---

@hirzel thanks for reviewing! Many apologies for the citation being missing. This was there in a previously longer submission and I guess it got cut out from there, **we will add Lale back in**. Regarding the `>>` operator and associated operators, you're right that heavy inspiration was drawn from Lale for this. Regarding the three raised bullet points:

* **Performance** - We did consider doing some empirical evaluations but refrained from doing so as:
    * AMLTK itself is not an AutoML-System, rather the objective is to provide some framework for building one. We have already begun doing so by rebuilding some of our Lab's AutoML tools in AMLTK. The empirical analysis of comparing different strategies employed by existing AutoML systems would be too large of a scope for this submission which I believe is largely about justifying the need for such software.
* **Architecture** - Absolutely! We had a previous version from early on in the development cycle. However as all things go early on, the accuracy between the actual implementation and the developed version begin to differ. I've included an old version which provides an abstract overview for reference **which should not be taken fully at face value anymore**. It's largely correct however there's some modifications I would like to do.
![overview-amltk](https://github.com/openjournals/joss-reviews/assets/15933394/3c52f5c4-5a1e-4a25-937f-44063e2afd6d)
I also believe a code overview would be helpful as was raised by @gomezzz in this [issue](https://github.com/automl/amltk/issues/267#issuecomment-2009203722). I will try to address both of these issues together.

* **Novelty** - You're correct that there's nothing inherently novel that does not exist disjoint from what's out there. However there are very few libraries (none?) that could allow someone to build an AutoML system using scikit-learn/PyTorch for tabular data but then also allow someone to use the same framework to build an AutoML system for image data, for example. This comes at a tradeoff of being less immediately useful (hence framework, not tool) but aims to address the fact building an AutoML system requires a lot of systems work which is constantly rebuilt for every tool out there. I believe this to be _a_ novelty but not in the traditional research paper form. Anyhelp on communicating this to improve the paper would be greatly appreciated!
