### 2024-03-26 #6367 @hirzel
Thanks for your detailed reply. That sounds like a good plan. Regarding your question:

> Anyhelp on communicating this to improve the paper would be greatly appreciated

I think you can pretty much add what you wrote above into the paper and/or readme:

> there are very few libraries (none?) that could allow someone to build an AutoML system using scikit-learn/PyTorch for tabular data but then also allow someone to use the same framework to build an AutoML system for image data, for example
