### 2024-03-26 automl/amltk#269 @gomezzz
Hi @eddiebergman ,

trying to run the example from #270 I now receive

```
  .../envs/amltk/lib/python3.11/site-packages/amltk/pipeline/parsers/configspace.py", line 146, in <module>
    from ConfigSpace import Categorical, ConfigurationSpace, Constant
ModuleNotFoundError: No module named 'ConfigSpace'
```
