### 2024-03-26 automl/amltk#273 @gomezzz
[JOSS Review] Cannot install package to run Scheduler with SLURM
I am trying to follow the example here https://automl.github.io/amltk/latest/examples/dask-jobqueue/

But I am unable to run `pip install openml "amltk[smac, sklearn, dask-jobqueue]"` successfully in a new conda environment

with

```
# Name                    Version                   Build  Channel
_libgcc_mutex             0.1                 conda_forge    conda-forge
_openmp_mutex             4.5                       2_gnu    conda-forge
bzip2                     1.0.8                hd590300_5    conda-forge
ca-certificates           2024.2.2             hbcca054_0    conda-forge
ld_impl_linux-64          2.40                 h41732ed_0    conda-forge
libexpat                  2.6.2                h59595ed_0    conda-forge
libffi                    3.4.2                h7f98852_5    conda-forge
libgcc-ng                 13.2.0               h807b86a_5    conda-forge
libgomp                   13.2.0               h807b86a_5    conda-forge
libnsl                    2.0.1                hd590300_0    conda-forge
libsqlite                 3.45.2               h2797004_0    conda-forge
libuuid                   2.38.1               h0b41bf4_0    conda-forge
libxcrypt                 4.4.36               hd590300_1    conda-forge
libzlib                   1.2.13               hd590300_5    conda-forge
ncurses                   6.4.20240210         h59595ed_0    conda-forge
openssl                   3.2.1                hd590300_1    conda-forge
pip                       24.0               pyhd8ed1ab_0    conda-forge
python                    3.11.8          hab00c5b_0_cpython    conda-forge
readline                  8.2                  h8228510_1    conda-forge
setuptools                69.2.0             pyhd8ed1ab_0    conda-forge
tk                        8.6.13          noxft_h4845f30_101    conda-forge
tzdata                    2024a                h0c530f3_0    conda-forge
wheel                     0.43.0             pyhd8ed1ab_0    conda-forge
xz                        5.2.6                h166bdaf_0    conda-forge
```

running the pip install I receive

```
Building wheels for collected packages: pyrfr
  Building wheel for pyrfr (setup.py) ... error
  error: subprocess-exited-with-error

  × python setup.py bdist_wheel did not run successfully.
  │ exit code: 1
  ╰─> [2 lines of output]
      [<setuptools.extension.Extension('pyrfr._regression') at 0x7b20d16df910>, <setuptools.extension.Extension('pyrfr._util') at 0x7b20d10c7850>]
      error: command 'swig' failed: No such file or directory
      [end of output]

  ERROR: Failed building wheel for pyrfr
  Running setup.py clean for pyrfr
Failed to build pyrfr
ERROR: Could not build wheels for pyrfr, which is required to install pyproject.toml-based projects

```

which seems to be a problem in one of the dependencies, however I am unsure how to proceed? If you have maybe a conda environment.yml or a docker or such, providing it may help users setting up their env?

(opened as part of JOSS Review https://github.com/openjournals/joss-reviews/issues/6367 )
