### 2024-03-26 automl/amltk#273 @gomezzz
Same seems to apply to the HPO example https://automl.github.io/amltk/latest/examples/hpo/
