### 2024-03-26 automl/amltk#274 @gomezzz
[JOSS REVIEW] Running 'Performing HPO with Post-Hoc Ensembling' example fails
After the workaround with swig mentioned in #273 I am trying the HPO examples now. The first one works well but for [the second](https://automl.github.io/amltk/latest/examples/hpo_with_ensembling/) I receive

```
Traceback (most recent call last):
  File ".../amltk/ex5.py", line 305, in <module>
    scheduler.run(timeout=30, wait=True)(9)
TypeError: 'ExitState' object is not callable
```

I tried increasing the timeout to 30 to see if that would fixed but without success. Is there anything else I might have to change?

(opened as part of JOSS Review https://github.com/openjournals/joss-reviews/issues/6367 )
