### 2024-03-27 eb794d0
feat(Pytorch): update iter method to handle children of the Choice class
