+++ src/amltk/pipeline/node.py
@@ -468,15 +468,34 @@ class Node(RichRenderable, Generic[Item, Space]):
 
         return prefix_keys(d, f"{self.name}:")
 
-    def iter(self) -> Iterator[Node]:
-        """Iterate the the nodes, including this node.
+    def iter(self, skip_unchosen: bool = False) -> Iterator[Node]:
+        """Recursively iterate through the nodes starting from this node.
+
+        This method traverses the nodes in a depth-first manner, including
+        the current node and its children nodes.
+
+        Args:
+            skip_unchosen (bool): Flag to skip unchosen nodes in Choice nodes.
 
         Yields:
-            The nodes connected to this node
+            Iterator[Node]: Nodes connected to this node.
         """
+        # Import Choice node to avoid circular imports
+        from amltk.pipeline.components import Choice
+
+        # Yield the current node
         yield self
+
+        # Iterate through the child nodes
         for node in self.nodes:
-            yield from node.iter()
+            if skip_unchosen and isinstance(node, Choice):
+                # If the node is a Choice and skipping unchosen nodes is enabled
+                chosen_node = node.chosen()
+                if chosen_node:
+                    yield from chosen_node.iter(skip_unchosen)
+            else:
+                # Recursively iterate through the child nodes
+                yield from node.iter(skip_unchosen)
 
     def mutate(self, **kwargs: Any) -> Self:
         """Mutate the node with the given keyword arguments.
