### 2024-03-27 379570d
feat(Pytorch): Add Choice class to the pipeline
