+++ src/amltk/pytorch/builders.py
@@ -1,69 +1,13 @@
 from __future__ import annotations
 
-from torch import nn, optim
-
-from amltk import Component, Metric, Sequential, Choice, Fixed
-
 from rich import print
+from torch import nn
 
+from amltk import Choice, Fixed, Sequential
 
-
-def build_model_from_pipeline(pipeline: Sequential) -> nn.Module:
-    """Builds a model from the provided pipeline.
-
-    This function iterates through the pipeline nodes, constructing the model
-    layers dynamically based on the node types and configurations. It also
-    utilizes the `MatchDimensions` objects to handle dimension matching
-    between layers with search spaces.
-
-    Args:
-        pipeline: The pipeline containing the model architecture.
-
-    Returns:
-        The constructed PyTorch model.
-    """
-    print("Building model from pipeline")
-    # print(pipeline)
-
-    model_layers = []
-
-    for node in pipeline.iter():
-        print("Node:",  node.name, "isinstance Choice:", isinstance(node, Choice))
-        print("-" * 80)
-
-        if isinstance(node, Choice):
-            chosen_node = node.chosen()
-            print("Choice node")
-            print(chosen_node)
-            print("*" * 80)
-
-        # Check if node is a Flatten layer, ReLU or similar.
-        if isinstance(node, Fixed):
-            model_layers.append(node.item)
-
-        # Check if node is a Component with config parameter
-        elif isinstance(node.item, type) and issubclass(node.item, nn.Module) and node.config:
-            layer_config = node.config or {}
-
-            # Evaluate MatchDimensions objects
-            for key, value in layer_config.items():
-                if isinstance(value, MatchDimensions):
-                    layer_config[key] = value.evaluate(pipeline)
-
-            layer = node.build_item(**layer_config)
-            model_layers.append(layer)
-
-    model = nn.Sequential(*model_layers)
-
-    print("-" * 80)
-    print("Model built")
-    print(model)
-    print("-" * 80)
-
-    return model
 
 class MatchDimensions:
-    """Class to handle matching dimensions between layers in a pipeline.
+    """Handles matching dimensions between layers in a pipeline.
 
     This class helps ensure compatibility between layers with search spaces
     during HPO optimization. It takes the layer name and parameter name
@@ -96,3 +40,103 @@ class MatchDimensions:
                 f"Parameter {self.param} not found in config of layer {self.layer_name}",
             )
         return value
+
+
+class MatchChosenDimensions:
+    """Handles matching dimensions for chosen nodes (children of Choice class) in a pipeline.
+
+    This class helps ensure compatibility between layers with search spaces
+    during HPO optimization. It takes the choice name and the corresponding
+    dimensions for that choice and stores them for later reference.
+
+    When called, it retrieves the corresponding dimension based on the name
+    of the chosen node.
+    """
+
+    chosen_nodes = {}  # Class variable to store chosen node names
+
+    def __init__(self, choice_name: str, choices: dict):
+        self.choice_name = choice_name
+        self.choices = choices
+
+    @staticmethod
+    def collect_chosen_nodes(pipeline: Sequential) -> None:
+        """Collects the names of chosen nodes in the pipeline.
+
+        Args:
+            pipeline: The pipeline containing the model architecture.
+        """
+        for node in pipeline.iter():
+            if isinstance(node, Choice):
+                chosen_node = node.chosen()
+                if chosen_node:
+                    MatchChosenDimensions.chosen_nodes[node.name] = chosen_node.name
+
+    def evaluate(self) -> int:
+        """Retrieves the corresponding dimension for the chosen node.
+
+        Returns:
+            The value of the matching dimension for a chosen node.
+        """
[…]
