### 2024-03-27 automl/amltk#273 @gomezzz
Hi @eddiebergman , glad to hear, no rush!

Yeah, just so you know, I am also not particularly worried about the quality of the module, just want to help make it accessible to an audience, who might be new to it like myself. :)
