### 2024-03-28 automl/amltk#273 @eddiebergman
Hi @gomezzz,

**Regarding the issues**

Assuming this is on Windows, I think the bulk of these issues stem from the fact I do not test/work on Windows. I had a student use Windows and he showed me that many of the dask tests failed. I will see what I can do about it. You should still be able to use a `Scheduler.with_processes()`.

The `pyrfr` problem is due to technical debt of `SMAC`, one of the optimizers, which relies on a custom C++ random forest which is only compiled for Linux up to Python `3.10`. This is why `swig` was required, so `pip` can build it locally. You should be able to use `OptunaOptimizer` instead of SMAC.

**Documenting these issues**

I'm not entirely sure how to do so, this library allows you to plug and play different tools into it, all of them optional so there's no required dependancies. Of course for the examples I need to choose one of them, but as you pointed out, there are issues with it. I'm not entirely sure what to do about this, any recommendations would be greatly appreciated.
