### 2024-03-28 automl/amltk#274 @eddiebergman
I'm not sure why that `(9)` is uncommeneted at the end of the line as in the example, it's commented out. Essentially `scheduler.run()` returns an `ExitState` which has information about why the `Scheduler` stopped running. The `(9)` is then trying to call that `ExitState` object.

```python
if __name__ == "__main__":
    scheduler.run(timeout=5, wait=True)  # (9)!
```

Did you have a formatter or something which removed the comment `#`? That would be good to know about! Perhaps you're using something like `PyCharm` and it has odd behaviours?
