### 2024-03-28 automl/amltk#269 @eddiebergman
Aye, that would need to be installed, as referenced in another issue, it's an entirely optional dependancy (but required with SMAC).

One option would be `pip install ConfigSpace` or `pip install amltk[smac]` which installs all the optional dependacies for the smac optimizer.

The other option would be to use `pip install amltk[optuna]` and use the Optuna optimizer instead.

---

I guess the main issue that needs to be addressed here is including a comment or something that mentions what dependancies are needed for which example.
