### 2024-03-28 automl/amltk#268 @eddiebergman
Good spot, I've ran into the issue a few times in other repos. Funnily enough, I think it is also dependant on which shell is being used, i.e `bash` vs `zsh`
