### 2024-03-28 7fc6f6f
feat(Pytorch): Add custom exceptions classes
