+++ src/amltk/__init__.py
@@ -40,6 +40,11 @@ from amltk.store import (
     TxtLoader,
     YAMLLoader,
 )
+from amltk.pytorch import (
+    MatchDimensions,
+    MatchChosenDimensions,
+    build_model_from_pipeline,
+)
 
 __all__ = [
     "Bucket",
@@ -77,4 +82,7 @@ __all__ = [
     "Trial",
     "TxtLoader",
     "YAMLLoader",
+    "MatchDimensions",
+    "MatchChosenDimensions",
+    "build_model_from_pipeline",
 ]
