+++ src/amltk/pytorch/__init__.py
@@ -1,0 +1,7 @@
+from amltk.pytorch.builders import MatchDimensions, MatchChosenDimensions, build_model_from_pipeline
+
+__all__ = [
+    "MatchDimensions",
+    "MatchChosenDimensions",
+    "build_model_from_pipeline",
+]
