+++ src/amltk/pytorch/__init__.py
@@ -1,4 +1,8 @@
-from amltk.pytorch.builders import MatchDimensions, MatchChosenDimensions, build_model_from_pipeline
+from amltk.pytorch.builders import (
+    MatchChosenDimensions,
+    MatchDimensions,
+    build_model_from_pipeline,
+)
 
 __all__ = [
     "MatchDimensions",
