+++ tests/pytorch/test_build_model_from_pipeline.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 import torch
 
-from amltk import Component, Fixed, Sequential, build_model_from_pipeline
-
+from amltk import Component, Fixed, Sequential
+from amltk.pytorch import build_model_from_pipeline
 
 def test_build_model_from_pipeline():
     # Define a simple pipeline for a multi-layer perceptron
