+++ tests/pytorch/test_match_chosen_dimensions.py
@@ -9,8 +9,8 @@ from amltk import (
     Fixed,
     Sequential,
 )
-from amltk.pytorch import MatchChosenDimensions, build_model_from_pipeline
 from amltk.exceptions import MatchChosenDimensionsError
+from amltk.pytorch import MatchChosenDimensions, build_model_from_pipeline
 from tests.pytorch.common import create_optimizer
 
 
