+++ pyproject.toml
@@ -208,6 +208,8 @@ exclude = [
 [tool.ruff.per-file-ignores]
 "tests/*.py" = [
   "S101",
+  "D101",
+  "D102",
   "D103",
   "ANN001",
   "ANN201",
