### 2024-04-04 743c9b8
feat(Pytorch): Minor updates to make ruff happy
