### 2024-04-05 automl/amltk#273 @gomezzz
Hi @eddiebergman ,

> Assuming this is on Windows, I think the bulk of these issues stem from the fact I do not test/work on Windows. I had a student use Windows and he showed me that many of the dask tests failed. I will see what I can do about it. You should still be able to use a Scheduler.with_processes().

I am on Linux :)

> Documenting these issues

> I'm not entirely sure how to do so, this library allows you to plug and play different tools into it, all of them optional so there's no required dependancies. Of course for the examples I need to choose one of them, but as you pointed out, there are issues with it. I'm not entirely sure what to do about this, any recommendations would be greatly appreciated.

I think a small note in the examples along the lines of "If you encounter problems installing with `pip install openml "amltk[smac]"` we recommend instead ...." or something like this would help already. Alternatively, you could provide a conda environment file or something with fixed versions? I think the main problem here is that normally you expect you install a framework and then you are ready to play with it but it seems you always need more dependencies here. So, alternatively, you could also consider providing a default config with one optimizer that works and then you only need to install more dependencies if you want to use another?
