### 2024-04-05 9dfda85
feat(Pytorch): Minor updates to make ruff happy
