+++ src/amltk/pytorch/pytorch_example.py
@@ -54,12 +54,12 @@ def test(
     test_loss = 0
     correct = 0
     with torch.no_grad():
-        for data, target in test_loader:
-            data, target = data.to(device), target.to(device)
-            output = model(data)
-            test_loss += f.nll_loss(output, target, reduction="sum").item()
+        for test_data, test_target in test_loader:
+            test_data, test_target = test_data.to(device), test_target.to(device)
+            output = model(test_data)
+            test_loss += f.nll_loss(output, test_target, reduction="sum").item()
             pred = output.argmax(dim=1, keepdim=True)
-            correct += pred.eq(target.view_as(pred)).sum().item()
+            correct += pred.eq(test_target.view_as(pred)).sum().item()
 
     test_loss /= len(test_loader.dataset)
     accuracy = 100.0 * correct / len(test_loader.dataset)
