### 2024-04-05 e3a10d7
feat(Pytorch): Minor updates to make ruff happy
