+++ src/amltk/pytorch/pytorch_example.py
@@ -51,8 +51,8 @@ def test(
         tuple[float, float]: Test loss and accuracy.
     """
     model.eval()
-    test_loss = 0
-    correct = 0
+    test_loss = 0.0
+    correct = 0.0
     with torch.no_grad():
         for test_data, test_target in test_loader:
             test_data, test_target = test_data.to(device), test_target.to(device)
