### 2024-04-05 0dfd923
feat(PyTorch): Add functionality to construct a PyTorch Model from a pipeline (#276)
