+++ src/amltk/pytorch/builders.py
@@ -1,0 +1,169 @@
+"""This module contains functionality to construct a pytorch model from a pipeline.
+
+It also includes classes for handling dimension matching between layers.
+"""
+
+from __future__ import annotations
+
+from typing import Any
+
+from torch import nn
+
+from amltk import Choice, Fixed, Node
+from amltk.exceptions import MatchChosenDimensionsError, MatchDimensionsError
+
+
+class MatchDimensions:
+    """Handles matching dimensions between layers in a pipeline.
+
+    This class helps ensure compatibility between layers with search spaces
+    during HPO optimization. It takes the layer name and parameter name
+    and stores them for later reference.
+
+    Not intended to be used inside a Choice node.
+
+    Attributes:
+        layer_name (str): The name of the layer.
+        param (str | None): The name of the parameter.
+    """
+
+    def __init__(self, layer_name: str, param: str | None) -> None:
+        """Initializes the MatchDimensions object.
+
+        Args:
+            layer_name (str): The name of the layer.
+            param (str | None): The name of the parameter.
+        """
+        self.layer_name = layer_name
+        self.param = param
+
+    def evaluate(self, pipeline: Node) -> int:
+        """Retrieves the corresponding configuration value from the pipeline.
+
+        Args:
+            pipeline: The pipeline to search for the matching configuration.
+
+        Returns:
+            The value of the matching configuration parameter.
+        """
+        layer = pipeline[self.layer_name]
+        layer_config = layer.config
+        if layer_config is None:
+            raise MatchDimensionsError(self.layer_name, None)
+        value = layer_config.get(self.param)
+        if value is None:
+            raise MatchDimensionsError(self.layer_name, self.param)
+        return value
+
+
+class MatchChosenDimensions:
+    """Handles matching dimensions for chosen nodes in a pipeline.
+
+    This class helps ensure compatibility between layers with search spaces
+    during HPO optimization. It takes the choice name and the corresponding
+    dimensions for that choice and stores them for later reference.
+
+    Attributes:
+        choice_name (str): The name of the choice.
+        choices (dict): A dictionary containing dimensions for choices.
+    """
+
+    def __init__(self, choice_name: str, choices: dict) -> None:
+        """Initializes the MatchChosenDimensions object.
+
+        Args:
+            choice_name (str): The name of the choice.
+            choices (dict): A dictionary containing dimensions for choices.
+        """
+        self.choice_name = choice_name
+        self.choices = choices
+
+    def evaluate(self, chosen_nodes: dict[str, Any]) -> int:
+        """Retrieves the corresponding dimension for the chosen node.
+
+        If the chosen node is not found in the choices dictionary, an error is raised.
+        If the dimensions provided are not valid, an error is not raised.
+        It is up to the user to ensure that the dimensions are valid.
+
+        Args:
+            chosen_nodes: The chosen nodes.
+
+        Returns:
+            The value of the matching dimension for a chosen node.
+        """
+        chosen_node_name = chosen_nodes.get(self.choice_name)
+
+        try:
+            return self.choices[chosen_node_name]
+        except KeyError as e:
+            raise MatchChosenDimensionsError(self.choice_name, chosen_node_name) from e
+
+    @staticmethod
+    def collect_chosen_nodes_names(pipeline: Node) -> dict[str, str]:
+        """Collects the names of chosen nodes in the pipeline.
+
+        Each pipeline has a unique set of chosen nodes, which we collect separately
+        to handle dimension matching between layers with search spaces.
+
+        Args:
+            pipeline: The pipeline containing the model architecture.
+
+        Returns:
+            The names of the chosen nodes in the pipeline.
+        """
[…]
