### 2024-04-05 #6367 @gomezzz
@JBorrow @eddiebergman

I opened a bunch of minor issues mostly related to installation instructions and docs (these are behind the currently unchecked boxes for guidelines and installation instructions) but the required changes should be trivial, I think.

Aside from that, from my end, I have finished trying out amltk and looking at the code and paper. I am happy to say, it's a very comprehensive and mature project that deserves publication in JOSS. :)
