### 2024-04-12 fb43b88
refactor: Move pytorch example to examples folder
