### 2024-04-12 920d8a6
refactor: Convert to match things to dataclasses
