### 2024-04-12 5f40a69
fix: Handle when key doesn't exist in chosen nodes
