### 2024-04-12 19fde2a
refactor: Use `match` for more explicit case handling
