### 2024-04-12 d21ceb2
doc: Fix bad tooltip numbering
