### 2024-04-24 ee1e0e3
fix(Trial): Issue with recursive rendering
