### 2024-04-24 30610c3
feat(CVEvaluator): stratified splitting with minorty classes
