### 2024-04-24 4470f33
doc(CVEvaluator): Example of `CVEvalautor` with `RandomSearch` (#283)
