+++ examples/sklearn-hpo-cv.py
@@ -1,0 +1,277 @@
+"""Random Search with CVEvaluation.
+
+This example demonstrates the [`CVEvaluation`][amltk.sklearn.CVEvaluation] class,
+which builds a custom cross-validation task that can be used to evaluate
+[`pipelines`](../guides/pipelines.md) with cross-validation, using
+[`RandomSearch`][amltk.optimization.optimizers.random_search.RandomSearch].
+"""
+from collections.abc import Mapping
+from pathlib import Path
+from typing import Any
+
+import numpy as np
+import openml
+import pandas as pd
+from ConfigSpace import Categorical, Integer
+from sklearn.ensemble import RandomForestClassifier
+from sklearn.impute import SimpleImputer
+from sklearn.metrics import get_scorer
+from sklearn.preprocessing import OneHotEncoder, OrdinalEncoder
+
+from amltk.optimization.optimizers.random_search import RandomSearch
+from amltk.optimization.trial import Metric, Trial
+from amltk.pipeline import Choice, Component, Node, Sequential, Split, request
+from amltk.sklearn import CVEvaluation
+
+
+def get_fold(
+    openml_task_id: int,
+    fold: int,
+) -> tuple[
+    pd.DataFrame,
+    pd.DataFrame,
+    pd.DataFrame | pd.Series,
+    pd.DataFrame | pd.Series,
+]:
+    """Get the data for a specific fold of an OpenML task.
+
+    Args:
+        openml_task_id: The OpenML task id.
+        fold: The fold number.
+        n_splits: The number of splits that will be applied. This is used
+            to resample training data such that enough at least instance for each class is present for
+            every stratified split.
+        seed: The random seed to use for reproducibility of resampling if necessary.
+    """
+    task = openml.tasks.get_task(
+        openml_task_id,
+        download_splits=True,
+        download_data=True,
+        download_qualities=True,
+        download_features_meta_data=True,
+    )
+    train_idx, test_idx = task.get_train_test_split_indices(fold=fold)
+    X, y = task.get_X_and_y(dataset_format="dataframe")  # type: ignore
+    X_train, y_train = X.iloc[train_idx], y.iloc[train_idx]
+    X_test, y_test = X.iloc[test_idx], y.iloc[test_idx]
+    return X_train, X_test, y_train, y_test
+
+
+preprocessing = Split(
+    {
+        "numerical": Component(SimpleImputer, space={"strategy": ["mean", "median"]}),
+        "categorical": [
+            Component(
+                OrdinalEncoder,
+                config={
+                    "categories": "auto",
+                    "handle_unknown": "use_encoded_value",
+                    "unknown_value": -1,
+                    "encoded_missing_value": -2,
+                },
+            ),
+            Choice(
+                "passthrough",
+                Component(
+                    OneHotEncoder,
+                    space={"max_categories": (2, 20)},
+                    config={
+                        "categories": "auto",
+                        "drop": None,
+                        "sparse_output": False,
+                        "handle_unknown": "infrequent_if_exist",
+                    },
+                ),
+                name="one_hot",
+            ),
+        ],
+    },
+    name="preprocessing",
+)
+
+
+def rf_config_transform(config: Mapping[str, Any], _: Any) -> dict[str, Any]:
+    new_config = dict(config)
+    if new_config["class_weight"] == "None":
+        new_config["class_weight"] = None
+    return new_config
+
+
+# NOTE: This space should not be used for evaluating how good this RF is
+# vs other algorithms
+rf_classifier = Component(
+    item=RandomForestClassifier,
+    config_transform=rf_config_transform,
+    space={
+        "criterion": ["gini", "entropy"],
+        "max_features": Categorical(
[…]
