### 2024-05-11 #6367 @woznicak
I apologize very much for the delays due to fortuitous reasons. I am very sorry to have delayed the whole review process.

I think a very big work has been done and in my opinion in paper can be accepted to JOSS. I see great novelty and impact on the AutoML field.

The paper and software is very interesting and important for the development of AutoML. It addresses key issues related to the comparability of entire frameworks or individual operations used in ML pipelines. The AutoML Fall School script is very useful for seeing the full capabilities of AMLTK.

The only aspect that, it is worth considering is missing examples that would show how we can reproduce or map the operation of already existing frameworks such as Autogluon, or at least their key elements such as Multilayer stacking. The examples that have been shown focus on showing the basic pipelines that can be built in scikitlearn and rather show the ease of implementing optimizations in SMAC or Optuna.
