### 2024-05-14 8695d29
fix(pipeline): `request` now fails without default
Previously when serialized to a new process, object ids would no longer
match, causing the `has_default` check to succeed, even if not set.
