### 2024-05-22 17dbaa6
ci: Remove unnecessary Python setup step
The Python setup isn't actually used in the remainder of the jobs.
