+++ .github/workflows/release.yml
@@ -32,12 +32,8 @@ jobs:
       - uses: actions/checkout@v4
         with:
           fetch-tags: 1  # Essential to later commitizen
-          fetch-depth: 0  # Reccommended by the action
+          fetch-depth: 0  # Recommended by the action
           token: ${{ secrets.PUSH_ACCESS }}
-      - uses: actions/setup-python@v5
-        with:
-          python-version: "3.10"
-          cache: pip
       - run: git tag  # Debug statement
       - name: Create bump and changelog
         uses: commitizen-tools/commitizen-action@master
