### 2024-05-27 87922aa
doc: Add citations
