+++ paper.bib
@@ -163,3 +163,21 @@
   pages     = {434--447},
   year      = {2021},
 }
+
+@inproceedings{baudart2021,
+    title = "Pipeline Combinators for Gradual {AutoML}",
+    author = "G. Baudart and M. Hirzel and K. Kate and P. Ram and A. Shinnar and J. Tsay",
+    booktitle = "Advances in Neural Information Processing Systems (NeurIPS)",
+    year = 2021,
+    month = dec,
+    pages = "19705--19718",
+    url = "https://proceedings.neurips.cc/paper/2021/file/a3b36cb25e2e0b93b5f334ffb4e4064e-Paper.pdf"
+}
+@article{ledell2020,
+    title = {{H2O} {A}uto{ML}: Scalable Automatic Machine Learning},
+    author = {E. LeDell and S. Poirier},
+    year = {2020},
+    month = {July},
+    journal = {7th ICML Workshop on Automated Machine Learning (AutoML)},
+    url = {https://www.automl.org/wp-content/uploads/2020/07/AutoML_2020_paper_61.pdf},
+}
