+++ paper.md
@@ -82,7 +82,7 @@ Typically, AutoML systems evaluate ML models sequentially to return a well-perfo
 Existing AutoML systems are typically highly engineered monolithic software developed for specific use cases to perform well and robustly under various conditions.
 
 With the growing amount of data and design decisions for ML, there is also a growing need to improve our understanding of the design decisions of AutoML systems.
-Current state-of-the-art systems vary in implemented paradigms (stacking [@erickson-arxiv20a] vs CASH [@thornton-kdd13a], optimizing a pre-defined pipeline structure [@thornton-kdd13a] vs evolving open-ended pipelines [@olson-gecco16a]) and also use different methods
+Current state-of-the-art systems vary in implemented paradigms (stacking [@ledell2020; @erickson-arxiv20a] vs CASH [@thornton-kdd13a], optimizing a pre-defined pipeline structure [@thornton-kdd13a; @baudart2021] vs evolving open-ended pipelines [@olson-gecco16a]) and also use different methods
 within one paradigm (i.e. Bayesian optimization [@thornton-kdd13a; @feurer-nips15a] or Genetic Programming [@olson-gecco16a; @gijsbers-joss19a] as the optimization algorithm,
 different search spaces for the same machine learning algorithm cf. [@olson-gecco16a; @gijsbers-joss19a; @thornton-kdd13a; @feurer-nips15a], different post-hoc ensemble methods or even no post-hoc ensembling at all cf. [@feurer-nips15a; @autoprognosis; @wang2021flaml]),
 raising many research questions and opportunities to study improved algorithms and novel applications.
