### 2024-05-28 automl/amltk#272 @eddiebergman
Hi @gomezzz, sorry to get back to this so late, which example was it that you were trying to run? Was it this one?

https://automl.github.io/amltk/latest/examples/dask-jobqueue/
