+++ src/amltk/optimization/trial.py
@@ -1,29 +1,4 @@
-"""A [`Trial`][amltk.optimization.Trial] is
-typically the output of
-[`Optimizer.ask()`][amltk.optimization.Optimizer.ask], indicating
-what the optimizer would like to evaluate next. We provide a host
-of convenience methods attached to the `Trial` to make it easy to
-save results, store artifacts, and more.
-
-Paired with the `Trial` is the [`Trial.Report`][amltk.optimization.Trial.Report],
-class, providing an easy way to report back to the optimizer's
-[`tell()`][amltk.optimization.Optimizer.tell] with
-a simple [`trial.success(cost=...)`][amltk.optimization.Trial.success] or
-[`trial.fail(cost=...)`][amltk.optimization.Trial.fail] call..
-
-### Trial
-
-::: amltk.optimization.trial.Trial
-    options:
-        members: False
-
-### Report
-
-::: amltk.optimization.trial.Trial.Report
-    options:
-        members: False
-
-"""
+"""The Trial and Report class."""
 from __future__ import annotations
 
 import copy
@@ -78,159 +53,7 @@ logger = logging.getLogger(__name__)
 
 @dataclass(kw_only=True)
 class Trial(RichRenderable, Generic[I]):
-    """A [`Trial`][amltk.optimization.Trial] encapsulates some configuration
-    that needs to be evaluated. Typically, this is what is generated by an
-    [`Optimizer.ask()`][amltk.optimization.Optimizer.ask] call.
-
-    ??? tip "Usage"
-
-        If all went smooth, your trial was successful and you can use
-        [`trial.success()`][amltk.optimization.Trial.success] to generate
-        a success [`Report`][amltk.optimization.Trial.Report], typically
-        passing what your chosen optimizer expects, e.g., `"loss"` or `"cost"`.
-
-        If your trial failed, you can instead use the
-        [`trial.fail()`][amltk.optimization.Trial.fail] to generate a
-        failure [`Report`][amltk.optimization.Trial.Report]. If use pass
-        in an exception to `fail()`, it will be attached to the report along
-        with any traceback it can deduce.
-        Each [`Optimizer`][amltk.optimization.Optimizer] will take
-        care of what to do from here.
-
-        ```python exec="true" source="material-block" html="true"
-        from amltk.optimization import Trial, Metric
-        from amltk.store import PathBucket
-
-        cost = Metric("cost", minimize=True)
-
-        def target_function(trial: Trial) -> Trial.Report:
-            x = trial.config["x"]
-            y = trial.config["y"]
-
-            with trial.profile("expensive-calculation"):
-                cost = x**2 - y
-
-            return trial.success(cost=cost)
-
-        # ... usually obtained from an optimizer
-        trial = Trial.create(
-            name="some-unique-name",
-            config={"x": 1, "y": 2},
-            metrics=[cost]
-        )
-
-        report = target_function(trial)
-        print(report.df())
-        trial.bucket.rmdir()  # markdown-exec: hide
-        ```
-
-
-    What you can return with [`trial.success()`][amltk.optimization.Trial.success]
-    or [`trial.fail()`][amltk.optimization.Trial.fail] depends on the
-    [`metrics`][amltk.optimization.Trial.metrics] of the trial. Typically,
-    an optimizer will provide the trial with the list of metrics.
-
-    ??? tip "Metrics"
-
-        ::: amltk.optimization.metric.Metric
-            options:
-                members: False
-
-    Some important properties are that they have a unique
-    [`.name`][amltk.optimization.Trial.name] given the optimization run,
-    a candidate [`.config`][amltk.optimization.Trial.config] to evaluate,
[…]
