### 2024-06-06 automl/amltk#273 @gomezzz
Yes, makes sense. Also ideally one that runs without additional deps that may cause problems (see install issues at the start of this issue)
