+++ src/amltk/scheduling/scheduler.py
@@ -1,253 +1,4 @@
-"""The [`Scheduler`][amltk.scheduling.Scheduler] uses
-an [`Executor`][concurrent.futures.Executor], a builtin python native with
-a `#!python submit(f, *args, **kwargs)` function to submit compute to
-be compute else where, whether it be locally or remotely.
-
-The `Scheduler` is primarily used to dispatch compute to an `Executor` and
-emit `@events`, which can trigger user callbacks.
-
-Typically you should not use the `Scheduler` directly for dispatching and
-responding to computed functions, but rather use a [`Task`][amltk.scheduling.Task]
-
-??? note "Running in a Jupyter Notebook/Colab"
-
-    If you are using a Jupyter Notebook, you likley need to use the following
-    at the top of your notebook:
-
-    ```python
-    import nest_asyncio  # Only necessary in Notebooks
-    nest_asyncio.apply()
-
-    scheduler.run(...)
-    ```
-
-    This is due to the fact a notebook runs in an async context. If you do not
-    wish to use the above snippet, you can instead use:
-
-    ```python
-    await scheduler.async_run(...)
-    ```
-
-??? tip "Basic Usage"
-
-    In this example, we create a scheduler that uses local processes as
-    workers. We then create a task that will run a function `fn` and submit it
-    to the scheduler. Lastly, a callback is registered to `@future-result` to print the
-    result when the compute is done.
-
-    ```python exec="true" source="material-block" html="true"
-    from amltk.scheduling import Scheduler
-
-    def fn(x: int) -> int:
-        return x + 1
-    from amltk._doc import make_picklable; make_picklable(fn)  # markdown-exec: hide
-
-    scheduler = Scheduler.with_processes(1)
-
-    @scheduler.on_start
-    def launch_the_compute():
-        scheduler.submit(fn, 1)
-
-    @scheduler.on_future_result
-    def callback(future, result):
-        print(f"Result: {result}")
-
-    scheduler.run()
-    from amltk._doc import doc_print; doc_print(print, scheduler)  # markdown-exec: hide
-    ```
-
-    The last line in the previous example called
-    [`scheduler.run()`][amltk.scheduling.Scheduler.run] is what starts the scheduler
-    running, in which it will first emit the `@start` event. This triggered the
-    callback `launch_the_compute()` which submitted the function `fn` with the
-    arguments `#!python 1`.
-
-    The scheduler then ran the compute and waited for it to complete, emitting the
-    `@future-result` event when it was done successfully. This triggered the callback
-    `callback()` which printed the result.
-
-    At this point, there is no more compute happening and no more events to respond to
-    so the scheduler will halt.
-
-??? example "`@events`"
-
-    === "Scheduler Status Events"
-
-        When the scheduler enters some important state, it will emit an event
-        to let you know.
-
-        === "`@start`"
-
-            ::: amltk.scheduling.Scheduler.on_start
-
-        === "`@finishing`"
-
-            ::: amltk.scheduling.Scheduler.on_finishing
-
-        === "`@finished`"
-
-            ::: amltk.scheduling.Scheduler.on_finished
-
-        === "`@stop`"
-
-            ::: amltk.scheduling.Scheduler.on_stop
-
-        === "`@timeout`"
-
-            ::: amltk.scheduling.Scheduler.on_timeout
-
-        === "`@empty`"
-
-            ::: amltk.scheduling.Scheduler.on_empty
-
-    === "Submitted Compute Events"
-
-        When any compute goes through the `Scheduler`, it will emit an event
-        to let you know. You should however prefer to use a
-        [`Task`][amltk.scheduling.Task] as it will emit specific events
-        for the task at hand, and not all compute.
-
-        === "`@future-submitted`"
-
-            ::: amltk.scheduling.Scheduler.on_future_submitted
-
[…]
