### 2024-06-17 #6367 @eddiebergman
Hi @JBorrow, apologies, some other deadlines took precedence. I've made some paper changes in accordance to what was said here but mainly working on @gomezzz's helpful comments. I'm reducing some complexity and improving documentation for first time users.
