### 2024-08-09 #6367 @JBorrow
@eddiebergman the final steps here are as follows:

1. Tidying up missing DOIs for the above where at all possible
2. Making a new release and associated deposit at Zenodo/Figshare/etc. to get a frozen DOI for the 'accepted' software.

Please let me know if you have any questions!
