### 2024-08-12 e12d64b
doc: Fixup DOIs
