+++ paper.bib
@@ -12,15 +12,7 @@
   author    = {M. Feurer and A. Klein and K. Eggensperger and J. Springenberg and M. Blum and F. Hutter},
   year      = 2015,
   pages     = {2962--2970},
-  crossref  = {nips15},
-}
-
-@proceedings{nips15,
-  title        = {Proceedings of the 29th International Conference on Advances in Neural Information Processing Systems ({N}eur{IPS}'15)},
-  year         = 2015,
-  booktitle    = {Proceedings of the 29th International Conference on Advances in Neural Information Processing Systems ({N}eur{IPS}'15)},
-  publisher    = {Curran Associates},
-  editor       = {C. Cortes and N. Lawrence and D. Lee and M. Sugiyama and R. Garnett},
+  journal   = {Advances in Neural Information Processing Systems 28 (NIPS 2015)},
 }
 
 @article{feurer-jmlr22a,
@@ -31,6 +23,7 @@
   volume    = 23,
   number    = 261,
   pages     = {1--61},
+  doi       = {10.48550/arXiv.2007.04074}
 }
 
 @article{gijsbers-arxiv23a,
