### 2024-08-12 #6367 @eddiebergman
Hi @JBorrow,

Clarification for the `'accepted'` software part, is that the latest version including some fixe based on reviews or is it the updated paper branch with the fixed DOIs?
