### 2024-08-13 ad38d09
Deployed 39a8edc to 1.12.1 with MkDocs 1.6.0 and mike 2.1.3
