### 2024-08-13 #6367 @eddiebergman
Hi @JBorrow, I've published on zenodo here: https://zenodo.org/records/13309538

As for the DOI's missing, I was only able to correct 2 of them, the others did not have a DOI issued to them
