### 2024-08-14 #6367 @crvernon
:wave: @eddiebergman - I just need you to address the following before I can accept this for publication:

- [ ] The paper title and the Zenodo archive title must match exactly.  Please edit the Zenodo archive's metadata to reflect this.

In the paper:

- [ ] LINE 130 and 133:  "bayesian" should have a capital "B".  Formatting can be maintained in the bib file by using curly brackets around the characters you wish to maintain the formatting of.

Thanks
