### 2024-08-14 13a8d08
doc: Fixup capatilization of bib {B}ayesian
