+++ paper.bib
@@ -76,7 +76,7 @@
 }
 
 @article{lindauer-jmlr22a,
-  title     = {{SMAC3}: A Versatile Bayesian Optimization Package for {H}yperparameter {O}ptimization},
+  title     = {{SMAC3}: A Versatile {B}ayesian Optimization Package for {H}yperparameter {O}ptimization},
   author    = {M. Lindauer and K. Eggensperger and M. Feurer and A. Biedenkapp and D. Deng and C. Benjamins and T. Ruhkopf and R. Sass and F. Hutter},
   year      = {2022},
   journal   = {Journal of Machine Learning Research},
@@ -94,7 +94,7 @@
 }
 
 @article{lindauer-arxiv19a,
-  title     = {BOAH: A Tool Suite for Multi-Fidelity Bayesian Optimization & Analysis of Hyperparameters},
+  title     = {BOAH: A Tool Suite for Multi-Fidelity {B}ayesian Optimization & Analysis of Hyperparameters},
   author    = {M. Lindauer and K. Eggensperger and M. Feurer and A. Biedenkapp and J. Marben and P. Müller and F. Hutter},
   journal   = {arXiv:1908.06756 {[cs.LG]}},
   year      = {2019},
