### 2024-08-14 68eb1bb
doc: Add CITATION.cff
