### 2024-08-14 #6367 @crvernon
🥳 Congratulations on your new publication @eddiebergman! Many thanks to @JBorrow for editing and @gomezzz, @woznicak, and @hirzel for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
